Load college basketball games for a given season

Pull in all college basketball games tracked by collegebasketballdata.com for a given season

Usage

load_games(seasons = most_recent_season())

Arguments

seasons

an integer of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of all games in the specified season(s).

Examples

# \donttest{
load_games()
#>           id source_id season_label season season_type tournament
#>        <int>    <char>       <char>  <int>      <char>     <char>
#>    1:     24 401727499     20242025   2025     regular       <NA>
#>    2:     26 401725796     20242025   2025     regular       <NA>
#>    3:     25 401726400     20242025   2025     regular       <NA>
#>    4:     27 401715872     20242025   2025     regular       <NA>
#>    5:     28 401726395     20242025   2025     regular       <NA>
#>   ---                                                            
#> 6296: 209747 401758633     20242025   2025  postseason       <NA>
#> 6297: 209731 401746081     20242025   2025  postseason       NCAA
#> 6298: 209732 401746080     20242025   2025  postseason       NCAA
#> 6299: 209749 401758636     20242025   2025  postseason       <NA>
#> 6300: 209751 401746082     20242025   2025  postseason       NCAA
#>                     start_date start_time_tbd neutral_site conference_game
#>                         <char>         <lgcl>       <lgcl>          <lgcl>
#>    1: 2024-11-04T11:00:00.000Z          FALSE        FALSE           FALSE
#>    2: 2024-11-04T12:00:00.000Z          FALSE        FALSE           FALSE
#>    3: 2024-11-04T12:00:00.000Z          FALSE        FALSE           FALSE
#>    4: 2024-11-04T12:00:00.000Z          FALSE         TRUE           FALSE
#>    5: 2024-11-04T12:30:00.000Z          FALSE        FALSE           FALSE
#>   ---                                                                     
#> 6296: 2025-04-05T20:00:00.000Z          FALSE         TRUE           FALSE
#> 6297: 2025-04-05T22:09:00.000Z          FALSE         TRUE           FALSE
#> 6298: 2025-04-06T01:28:00.000Z          FALSE         TRUE           FALSE
#> 6299: 2025-04-06T21:30:00.000Z          FALSE         TRUE           FALSE
#> 6300: 2025-04-08T00:50:00.000Z          FALSE         TRUE           FALSE
#>       game_type status                                            game_notes
#>          <char> <char>                                                <char>
#>    1:       STD  final                                                  <NA>
#>    2:       STD  final                                                  <NA>
#>    3:       STD  final                                                  <NA>
#>    4:    TRNMNT  final                                  Field of 68 Showcase
#>    5:       STD  final                                                  <NA>
#>   ---                                                                       
#> 6296:    TRNMNT  final                 College Basketball Crown - Semifinals
#> 6297:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 6298:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 6299:    TRNMNT  final            College Basketball Crown Championship Game
#> 6300:    TRNMNT  final Men's Basketball Championship - National Championship
#>       attendance home_team_id          home_team home_conference_id
#>            <int>        <int>             <char>              <int>
#>    1:       3716          115    IU Indianapolis                 14
#>    2:       1001          324       UT Arlington                 30
#>    3:       7501          171   Middle Tennessee                 13
#>    4:          0          273  Southern Illinois                 19
#>    5:       2055          128 Jacksonville State                 13
#>   ---                                                              
#> 6296:       2972          338          Villanova                  7
#> 6297:      68252           16             Auburn                 24
#> 6298:      68252           72               Duke                  2
#> 6299:       3314          312                UCF                  6
#> 6300:      66602          113            Houston                  6
#>       home_conference home_seed home_points home_period_points_1
#>                <char>     <int>       <int>                <int>
#>    1:         Horizon        NA         100                   48
#>    2:             WAC        NA          90                   48
#>    3:            CUSA        NA          97                   51
#>    4:             MVC        NA          80                   42
#>    5:            CUSA        NA         112                   55
#>   ---                                                           
#> 6296:        Big East        NA          98                   44
#> 6297:             SEC        NA          73                   46
#> 6298:             ACC        NA          67                   34
#> 6299:          Big 12        NA          66                   35
#> 6300:          Big 12        NA          63                   31
#>       home_period_points_2 home_period_points_3 home_period_points_4
#>                      <int>                <int>                <int>
#>    1:                   52                   NA                   NA
#>    2:                   42                   NA                   NA
#>    3:                   46                   NA                   NA
#>    4:                   38                   NA                   NA
#>    5:                   57                   NA                   NA
#>   ---                                                               
#> 6296:                   44                   10                   NA
#> 6297:                   27                   NA                   NA
#> 6298:                   33                   NA                   NA
#> 6299:                   31                   NA                   NA
#> 6300:                   32                   NA                   NA
#>       home_period_points_5 home_period_points_6 home_winner away_team_id
#>                      <int>                <int>      <lgcl>        <int>
#>    1:                   NA                   NA        TRUE          492
#>    2:                   NA                   NA        TRUE          494
#>    3:                   NA                   NA        TRUE          493
#>    4:                   NA                   NA       FALSE           46
#>    5:                   NA                   NA        TRUE          495
#>   ---                                                                   
#> 6296:                   NA                   NA       FALSE          312
#> 6297:                   NA                   NA       FALSE           87
#> 6298:                   NA                   NA       FALSE          113
#> 6299:                   NA                   NA       FALSE          188
#> 6300:                   NA                   NA       FALSE           87
#>         away_team away_conference_id away_conference away_seed away_points
#>            <char>              <int>          <char>     <int>       <int>
#>    1: IU Columbus                 NA            <NA>        NA          44
#>    2:  UNT Dallas                 NA            <NA>        NA          55
#>    3:  Oglethorpe                 NA            <NA>        NA          51
#>    4:  Charleston                 12             CAA        NA          90
#>    5:    LaGrange                 NA            <NA>        NA          48
#>   ---                                                                     
#> 6296:         UCF                  6          Big 12        NA         104
#> 6297:     Florida                 24             SEC        NA          79
#> 6298:     Houston                  6          Big 12        NA          70
#> 6299:    Nebraska                 10         Big Ten        NA          77
#> 6300:     Florida                 24             SEC        NA          65
#>       away_period_points_1 away_period_points_2 away_period_points_3
#>                      <int>                <int>                <int>
#>    1:                   21                   23                   NA
#>    2:                   25                   30                   NA
#>    3:                   22                   29                   NA
#>    4:                   45                   45                   NA
#>    5:                   20                   28                   NA
#>   ---                                                               
#> 6296:                   48                   40                   16
#> 6297:                   38                   41                   NA
#> 6298:                   28                   42                   NA
#> 6299:                   35                   42                   NA
#> 6300:                   28                   37                   NA
#>       away_period_points_4 away_period_points_5 away_period_points_6
#>                      <int>                <int>                <int>
#>    1:                   NA                   NA                   NA
#>    2:                   NA                   NA                   NA
#>    3:                   NA                   NA                   NA
#>    4:                   NA                   NA                   NA
#>    5:                   NA                   NA                   NA
#>   ---                                                               
#> 6296:                   NA                   NA                   NA
#> 6297:                   NA                   NA                   NA
#> 6298:                   NA                   NA                   NA
#> 6299:                   NA                   NA                   NA
#> 6300:                   NA                   NA                   NA
#>       away_winner excitement venue_id                 venue         city  state
#>            <lgcl>      <num>    <int>                <char>       <char> <char>
#>    1:       FALSE        0.6       19      Corteva Coliseum Indianapolis     IN
#>    2:       FALSE        1.5       21   College Park Center    Arlington     TX
#>    3:       FALSE        0.6       20         Murphy Center Murfreesboro     TN
#>    4:        TRUE        7.2       22      Sanford Pentagon  Sioux Falls     SD
#>    5:       FALSE        1.0       23 Pete Mathews Coliseum Jacksonville     AL
#>   ---                                                                          
#> 6296:        TRUE       12.3       13        T-Mobile Arena    Las Vegas     NV
#> 6297:        TRUE        8.6      629             Alamodome  San Antonio     TX
#> 6298:        TRUE        6.2      629             Alamodome  San Antonio     TX
#> 6299:        TRUE        7.0       13        T-Mobile Arena    Las Vegas     NV
#> 6300:        TRUE        7.7      629             Alamodome  San Antonio     TX
#>       home_period_points_7 away_period_points_7
#>                      <int>                <int>
#>    1:                   NA                   NA
#>    2:                   NA                   NA
#>    3:                   NA                   NA
#>    4:                   NA                   NA
#>    5:                   NA                   NA
#>   ---                                          
#> 6296:                   NA                   NA
#> 6297:                   NA                   NA
#> 6298:                   NA                   NA
#> 6299:                   NA                   NA
#> 6300:                   NA                   NA

load_games(2024)
#>          id source_id season_label season season_type tournament
#>       <int>    <char>       <char>  <int>      <char>     <char>
#>    1:  5901 401594209     20232024   2024     regular       <NA>
#>    2:  5902 401610539     20232024   2024     regular       <NA>
#>    3:  5903 401591565     20232024   2024     regular       <NA>
#>    4:  5904 401604410     20232024   2024     regular       <NA>
#>    5:  5905 401598014     20232024   2024     regular       <NA>
#>   ---                                                           
#> 6245: 12124 401641122     20232024   2024  postseason        NIT
#> 6246: 12125 401641124     20232024   2024  postseason        NIT
#> 6247: 12126 401638643     20232024   2024  postseason       NCAA
#> 6248: 12127 401638644     20232024   2024  postseason       NCAA
#> 6249: 12128 401638645     20232024   2024  postseason       NCAA
#>                     start_date start_time_tbd neutral_site conference_game
#>                         <char>         <lgcl>       <lgcl>          <lgcl>
#>    1: 2023-11-06T11:00:00.000Z          FALSE        FALSE           FALSE
#>    2: 2023-11-06T11:30:00.000Z          FALSE        FALSE           FALSE
#>    3: 2023-11-06T12:00:00.000Z          FALSE        FALSE           FALSE
#>    4: 2023-11-06T12:30:00.000Z          FALSE        FALSE           FALSE
#>    5: 2023-11-06T13:00:00.000Z          FALSE        FALSE           FALSE
#>   ---                                                                     
#> 6245: 2024-04-02T21:35:00.000Z          FALSE         TRUE           FALSE
#> 6246: 2024-04-04T19:00:00.000Z          FALSE         TRUE           FALSE
#> 6247: 2024-04-06T18:09:00.000Z          FALSE         TRUE           FALSE
#> 6248: 2024-04-06T20:51:00.000Z          FALSE         TRUE           FALSE
#> 6249: 2024-04-08T21:20:00.000Z          FALSE         TRUE           FALSE
#>       game_type status                                            game_notes
#>          <char> <char>                                                <char>
#>    1:       STD  final                                                  <NA>
#>    2:       STD  final                                                  <NA>
#>    3:       STD  final                                                  <NA>
#>    4:       STD  final                                                  <NA>
#>    5:       STD  final                                                  <NA>
#>   ---                                                                       
#> 6245:    TRNMNT  final                                       NIT - Semifinal
#> 6246:    TRNMNT  final                                    NIT - Championship
#> 6247:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 6248:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 6249:    TRNMNT  final Men's Basketball Championship - National Championship
#>       attendance home_team_id         home_team home_conference_id
#>            <int>        <int>            <char>              <int>
#>    1:       4867          115   IU Indianapolis                 14
#>    2:       2147          110           Hofstra                 12
#>    3:       3321          302              Troy                 29
#>    4:       3250          282 Stephen F. Austin                 27
#>    5:        350          232  Prairie View A&M                 25
#>   ---                                                             
#> 6245:          0          263        Seton Hall                  7
#> 6246:       9100          263        Seton Hall                  7
#> 6247:          0          236            Purdue                 10
#> 6248:      74720          314             UConn                  7
#> 6249:      74423          314             UConn                  7
#>       home_conference home_seed home_points home_period_points_1
#>                <char>     <int>       <int>                <int>
#>    1:         Horizon        NA          70                   33
#>    2:             CAA        NA         101                   44
#>    3:        Sun Belt        NA          92                   38
#>    4:       Southland        NA          96                   54
#>    5:            SWAC        NA          89                   43
#>   ---                                                           
#> 6245:        Big East        NA          84                   42
#> 6246:        Big East        NA          79                   39
#> 6247:         Big Ten         1          63                   35
#> 6248:        Big East         1          86                   44
#> 6249:        Big East         1          75                   36
#>       home_period_points_2 home_period_points_3 home_period_points_4
#>                      <int>                <int>                <int>
#>    1:                   37                   NA                   NA
#>    2:                   57                   NA                   NA
#>    3:                   54                   NA                   NA
#>    4:                   42                   NA                   NA
#>    5:                   46                   NA                   NA
#>   ---                                                               
#> 6245:                   42                   NA                   NA
#> 6246:                   40                   NA                   NA
#> 6247:                   28                   NA                   NA
#> 6248:                   42                   NA                   NA
#> 6249:                   39                   NA                   NA
#>       home_period_points_5 home_winner away_team_id                  away_team
#>                      <int>      <lgcl>        <int>                     <char>
#>    1:                   NA        TRUE          831                   Spalding
#>    2:                   NA        TRUE          677 Saint Joseph's Long Island
#>    3:                   NA        TRUE          797            Fort Lauderdale
#>    4:                   NA        TRUE          832             North American
#>    5:                   NA        TRUE          614           Kansas Christian
#>   ---                                                                         
#> 6245:                   NA        TRUE           98                    Georgia
#> 6246:                   NA        TRUE          122              Indiana State
#> 6247:                   NA        TRUE          185                   NC State
#> 6248:                   NA        TRUE            5                    Alabama
#> 6249:                   NA        TRUE          236                     Purdue
#>       away_conference_id away_conference away_seed away_points
#>                    <int>          <char>     <int>       <int>
#>    1:                 NA            <NA>        NA          63
#>    2:                 NA            <NA>        NA          48
#>    3:                 NA            <NA>        NA          47
#>    4:                 NA            <NA>        NA          68
#>    5:                 NA            <NA>        NA          66
#>   ---                                                         
#> 6245:                 24             SEC        NA          67
#> 6246:                 19             MVC        NA          77
#> 6247:                  2             ACC        11          50
#> 6248:                 24             SEC         4          72
#> 6249:                 10         Big Ten         1          60
#>       away_period_points_1 away_period_points_2 away_period_points_3
#>                      <int>                <int>                <int>
#>    1:                   25                   38                   NA
#>    2:                   19                   29                   NA
#>    3:                   18                   29                   NA
#>    4:                   22                   46                   NA
#>    5:                   33                   33                   NA
#>   ---                                                               
#> 6245:                   25                   42                   NA
#> 6246:                   39                   38                   NA
#> 6247:                   29                   21                   NA
#> 6248:                   40                   32                   NA
#> 6249:                   30                   30                   NA
#>       away_period_points_4 away_period_points_5 away_winner excitement venue_id
#>                      <int>                <int>      <lgcl>      <num>    <int>
#>    1:                   NA                   NA       FALSE        2.7       19
#>    2:                   NA                   NA       FALSE        1.3       27
#>    3:                   NA                   NA       FALSE        1.2       65
#>    4:                   NA                   NA       FALSE        0.9      154
#>    5:                   NA                   NA       FALSE        2.2       56
#>   ---                                                                          
#> 6245:                   NA                   NA       FALSE        1.3       40
#> 6246:                   NA                   NA       FALSE        7.2       40
#> 6247:                   NA                   NA       FALSE        2.3      485
#> 6248:                   NA                   NA       FALSE        4.5      485
#> 6249:                   NA                   NA       FALSE        3.0      485
#>                                             venue         city  state
#>                                            <char>       <char> <char>
#>    1:                            Corteva Coliseum Indianapolis     IN
#>    2: David S. Mack Sports and Exhibition Complex    Hempstead     NY
#>    3:                                Trojan Arena         Troy     AL
#>    4:                 William R. Johnson Coliseum  Nacogdoches     TX
#>    5:                      William Nicks Building Prairie View     TX
#>   ---                                                                
#> 6245:                           Hinkle Fieldhouse Indianapolis     IN
#> 6246:                           Hinkle Fieldhouse Indianapolis     IN
#> 6247:                          State Farm Stadium     Glendale     AZ
#> 6248:                          State Farm Stadium     Glendale     AZ
#> 6249:                          State Farm Stadium     Glendale     AZ
#>       home_period_points_6 away_period_points_6 home_period_points_7
#>                      <int>                <int>                <int>
#>    1:                   NA                   NA                   NA
#>    2:                   NA                   NA                   NA
#>    3:                   NA                   NA                   NA
#>    4:                   NA                   NA                   NA
#>    5:                   NA                   NA                   NA
#>   ---                                                               
#> 6245:                   NA                   NA                   NA
#> 6246:                   NA                   NA                   NA
#> 6247:                   NA                   NA                   NA
#> 6248:                   NA                   NA                   NA
#> 6249:                   NA                   NA                   NA
#>       away_period_points_7
#>                      <int>
#>    1:                   NA
#>    2:                   NA
#>    3:                   NA
#>    4:                   NA
#>    5:                   NA
#>   ---                     
#> 6245:                   NA
#> 6246:                   NA
#> 6247:                   NA
#> 6248:                   NA
#> 6249:                   NA

load_games(TRUE)
#>             id source_id season_label season season_type tournament
#>          <int>    <char>       <char>  <int>      <char>     <char>
#>      1: 127127 223180183     20022003   2003     regular       <NA>
#>      2: 127126 223180201     20022003   2003     regular       <NA>
#>      3: 127129 223190222     20022003   2003     regular       <NA>
#>      4: 127130 223192459     20022003   2003     regular       <NA>
#>      5: 127131 223190275     20022003   2003     regular       <NA>
#>     ---                                                            
#> 132536: 209747 401758633     20242025   2025  postseason       <NA>
#> 132537: 209731 401746081     20242025   2025  postseason       NCAA
#> 132538: 209732 401746080     20242025   2025  postseason       NCAA
#> 132539: 209749 401758636     20242025   2025  postseason       <NA>
#> 132540: 209751 401746082     20242025   2025  postseason       NCAA
#>                       start_date start_time_tbd neutral_site conference_game
#>                           <char>         <lgcl>       <lgcl>          <lgcl>
#>      1: 2002-11-14T19:00:00.000Z          FALSE        FALSE           FALSE
#>      2: 2002-11-14T21:30:00.000Z          FALSE        FALSE           FALSE
#>      3: 2002-11-15T18:30:00.000Z          FALSE        FALSE           FALSE
#>      4: 2002-11-15T19:00:00.000Z          FALSE        FALSE           FALSE
#>      5: 2002-11-15T21:00:00.000Z          FALSE        FALSE           FALSE
#>     ---                                                                     
#> 132536: 2025-04-05T20:00:00.000Z          FALSE         TRUE           FALSE
#> 132537: 2025-04-05T22:09:00.000Z          FALSE         TRUE           FALSE
#> 132538: 2025-04-06T01:28:00.000Z          FALSE         TRUE           FALSE
#> 132539: 2025-04-06T21:30:00.000Z          FALSE         TRUE           FALSE
#> 132540: 2025-04-08T00:50:00.000Z          FALSE         TRUE           FALSE
#>         game_type status                                            game_notes
#>            <char> <char>                                                <char>
#>      1:       STD  final                                                  <NA>
#>      2:       STD  final                                                  <NA>
#>      3:       STD  final                                                  <NA>
#>      4:       STD  final                                                  <NA>
#>      5:       STD  final                                                  <NA>
#>     ---                                                                       
#> 132536:    TRNMNT  final                 College Basketball Crown - Semifinals
#> 132537:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 132538:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 132539:    TRNMNT  final            College Basketball Crown Championship Game
#> 132540:    TRNMNT  final Men's Basketball Championship - National Championship
#>         attendance home_team_id         home_team home_conference_id
#>              <int>        <int>            <char>              <int>
#>      1:          0          286          Syracuse                  7
#>      2:       8826          217          Oklahoma                  6
#>      3:          0          338         Villanova                  7
#>      4:          0          208 Northern Illinois                 17
#>      5:      16374          355         Wisconsin                 10
#>     ---                                                             
#> 132536:       2972          338         Villanova                  7
#> 132537:      68252           16            Auburn                 24
#> 132538:      68252           72              Duke                  2
#> 132539:       3314          312               UCF                  6
#> 132540:      66602          113           Houston                  6
#>         home_conference home_seed home_points home_period_points_1
#>                  <char>     <int>       <int>                <int>
#>      1:        Big East        NA          63                   34
#>      2:          Big 12        NA          62                   25
#>      3:        Big East        NA          61                   23
#>      4:             MAC        NA          56                   31
#>      5:         Big Ten        NA          81                   34
#>     ---                                                           
#> 132536:        Big East        NA          98                   44
#> 132537:             SEC        NA          73                   46
#> 132538:             ACC        NA          67                   34
#> 132539:          Big 12        NA          66                   35
#> 132540:          Big 12        NA          63                   31
#>         home_period_points_2 home_period_points_3 home_period_points_4
#>                        <int>                <int>                <int>
#>      1:                   29                   NA                   NA
#>      2:                   37                   NA                   NA
#>      3:                   38                   NA                   NA
#>      4:                   25                   NA                   NA
#>      5:                   47                   NA                   NA
#>     ---                                                               
#> 132536:                   44                   10                   NA
#> 132537:                   27                   NA                   NA
#> 132538:                   33                   NA                   NA
#> 132539:                   31                   NA                   NA
#> 132540:                   32                   NA                   NA
#>         home_period_points_5 home_period_points_6 home_winner away_team_id
#>                        <int>                <int>      <lgcl>        <int>
#>      1:                   NA                   NA       FALSE          163
#>      2:                   NA                   NA       FALSE            5
#>      3:                   NA                   NA       FALSE          157
#>      4:                   NA                   NA        TRUE          354
#>      5:                   NA                   NA        TRUE           80
#>     ---                                                                   
#> 132536:                   NA                   NA       FALSE          312
#> 132537:                   NA                   NA       FALSE           87
#> 132538:                   NA                   NA       FALSE          113
#> 132539:                   NA                   NA       FALSE          188
#> 132540:                   NA                   NA       FALSE           87
#>                  away_team away_conference_id away_conference away_seed
#>                     <char>              <int>          <char>     <int>
#>      1:            Memphis                 13            CUSA        NA
#>      2:            Alabama                 24             SEC        NA
#>      3:          Marquette                 13            CUSA        NA
#>      4:           Winthrop                  9       Big South        NA
#>      5: Eastern Washington                  8         Big Sky        NA
#>     ---                                                                
#> 132536:                UCF                  6          Big 12        NA
#> 132537:            Florida                 24             SEC        NA
#> 132538:            Houston                  6          Big 12        NA
#> 132539:           Nebraska                 10         Big Ten        NA
#> 132540:            Florida                 24             SEC        NA
#>         away_points away_period_points_1 away_period_points_2
#>               <int>                <int>                <int>
#>      1:          70                   43                   27
#>      2:          68                   33                   35
#>      3:          73                   37                   36
#>      4:          50                   30                   20
#>      5:          55                   25                   30
#>     ---                                                      
#> 132536:         104                   48                   40
#> 132537:          79                   38                   41
#> 132538:          70                   28                   42
#> 132539:          77                   35                   42
#> 132540:          65                   28                   37
#>         away_period_points_3 away_period_points_4 away_period_points_5
#>                        <int>                <int>                <int>
#>      1:                   NA                   NA                   NA
#>      2:                   NA                   NA                   NA
#>      3:                   NA                   NA                   NA
#>      4:                   NA                   NA                   NA
#>      5:                   NA                   NA                   NA
#>     ---                                                               
#> 132536:                   16                   NA                   NA
#> 132537:                   NA                   NA                   NA
#> 132538:                   NA                   NA                   NA
#> 132539:                   NA                   NA                   NA
#> 132540:                   NA                   NA                   NA
#>         away_period_points_6 away_winner excitement venue_id
#>                        <int>      <lgcl>      <num>    <int>
#>      1:                   NA        TRUE         NA      371
#>      2:                   NA        TRUE         NA      371
#>      3:                   NA        TRUE         NA      371
#>      4:                   NA       FALSE         NA      142
#>      5:                   NA       FALSE         NA      142
#>     ---                                                     
#> 132536:                   NA        TRUE       12.3       13
#> 132537:                   NA        TRUE        8.6      629
#> 132538:                   NA        TRUE        6.2      629
#> 132539:                   NA        TRUE        7.0       13
#> 132540:                   NA        TRUE        7.7      629
#>                         venue        city  state home_period_points_7
#>                        <char>      <char> <char>                <int>
#>      1: Madison Square Garden    New York     NY                   NA
#>      2: Madison Square Garden    New York     NY                   NA
#>      3: Madison Square Garden    New York     NY                   NA
#>      4:           Kohl Center     Madison     WI                   NA
#>      5:           Kohl Center     Madison     WI                   NA
#>     ---                                                              
#> 132536:        T-Mobile Arena   Las Vegas     NV                   NA
#> 132537:             Alamodome San Antonio     TX                   NA
#> 132538:             Alamodome San Antonio     TX                   NA
#> 132539:        T-Mobile Arena   Las Vegas     NV                   NA
#> 132540:             Alamodome San Antonio     TX                   NA
#>         away_period_points_7 home_period_points_8 away_period_points_8
#>                        <int>                <int>                <int>
#>      1:                   NA                   NA                   NA
#>      2:                   NA                   NA                   NA
#>      3:                   NA                   NA                   NA
#>      4:                   NA                   NA                   NA
#>      5:                   NA                   NA                   NA
#>     ---                                                               
#> 132536:                   NA                   NA                   NA
#> 132537:                   NA                   NA                   NA
#> 132538:                   NA                   NA                   NA
#> 132539:                   NA                   NA                   NA
#> 132540:                   NA                   NA                   NA
# }