Load betting information for cbb games

Pull in vegas odds (overs, unders, spreads, moneylines, etc.) for college basketball games.

Usage

load_lines(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of betting information for all games for the specified season(s).

See also

Examples

# \donttest{
load_lines()
#>       game_id season season_type               start_date home_team_id
#>         <int>  <int>      <char>                   <char>        <int>
#>    1:      27   2025     regular 2024-11-04T12:00:00.000Z          273
#>    2:      29   2025     regular 2024-11-04T15:00:00.000Z          261
#>    3:      31   2025     regular 2024-11-04T16:00:00.000Z          281
#>    4:      35   2025     regular 2024-11-04T17:00:00.000Z          107
#>    5:      36   2025     regular 2024-11-04T17:00:00.000Z          142
#>   ---                                                                 
#> 5436:  209747   2025  postseason 2025-04-05T20:00:00.000Z          338
#> 5437:  209731   2025  postseason 2025-04-05T22:09:00.000Z           16
#> 5438:  209732   2025  postseason 2025-04-06T01:28:00.000Z           72
#> 5439:  209749   2025  postseason 2025-04-06T21:30:00.000Z          312
#> 5440:  209751   2025  postseason 2025-04-08T00:50:00.000Z          113
#>          home_team_name home_conference_short_name home_score away_team_id
#>                  <char>                     <char>      <int>        <int>
#>    1: Southern Illinois                        MVC         80           46
#>    2:       Santa Clara                        WCC         85          252
#>    3:          Stanford                        ACC         85           68
#>    4:           Harvard                        Ivy         79          156
#>    5:           Liberty                       CUSA         83          335
#>   ---                                                                     
#> 5436:         Villanova                   Big East         98          312
#> 5437:            Auburn                        SEC         73           87
#> 5438:              Duke                        ACC         67          113
#> 5439:               UCF                     Big 12         66          188
#> 5440:           Houston                     Big 12         63           87
#>       away_team_name away_conference_short_name away_score provider spread
#>               <char>                     <char>      <int>   <char>  <num>
#>    1:     Charleston                        CAA         90 ESPN BET    3.5
#>    2:    Saint Louis                       A-10         78 ESPN BET   -1.5
#>    3:         Denver                     Summit         62 ESPN BET  -17.5
#>    4:         Marist                       MAAC         66 ESPN BET   -2.5
#>    5:     Valparaiso                        MVC         63 ESPN BET  -10.5
#>   ---                                                                     
#> 5436:            UCF                     Big 12        104 ESPN BET   -3.5
#> 5437:        Florida                        SEC         79 ESPN BET    2.5
#> 5438:        Houston                     Big 12         70 ESPN BET   -4.5
#> 5439:       Nebraska                    Big Ten         77 ESPN BET    3.5
#> 5440:        Florida                        SEC         65 ESPN BET    1.5
#>       over_under home_moneyline away_moneyline spread_open over_under_open
#>            <num>          <int>          <int>       <num>           <num>
#>    1:      146.5            145           -170          NA              NA
#>    2:      157.5           -130            110          NA              NA
#>    3:      147.5          -3000           1200          NA              NA
#>    4:      131.5           -145            125          NA              NA
#>    5:      138.5           -750            475          NA              NA
#>   ---                                                                     
#> 5436:      156.5             NA             NA        -1.5           155.5
#> 5437:      158.5             NA             NA         2.5           159.5
#> 5438:      136.5             NA             NA        -4.5           135.5
#> 5439:      160.5             NA             NA         3.5           158.5
#> 5440:      140.5             NA             NA         1.5           140.5

load_lines(2024)
#>       game_id season season_type               start_date home_team_id
#>         <int>  <int>      <char>                   <char>        <int>
#>    1:    5906   2024     regular 2023-11-06T16:30:00.000Z          223
#>    2:    5887   2024     regular 2023-11-06T18:30:00.000Z          292
#>    3:    5882   2024     regular 2023-11-06T18:30:00.000Z          236
#>    4:    5885   2024     regular 2023-11-06T18:30:00.000Z          314
#>    5:    5968   2024     regular 2023-11-06T19:00:00.000Z          286
#>   ---                                                                 
#> 5307:   12124   2024  postseason 2024-04-02T21:35:00.000Z          263
#> 5308:   12125   2024  postseason 2024-04-04T19:00:00.000Z          263
#> 5309:   12126   2024  postseason 2024-04-06T18:09:00.000Z          236
#> 5310:   12127   2024  postseason 2024-04-06T20:51:00.000Z          314
#> 5311:   12128   2024  postseason 2024-04-08T21:20:00.000Z          314
#>       home_team_name home_conference_short_name home_score away_team_id
#>               <char>                     <char>      <int>        <int>
#>    1:         Oregon                     Pac-12         82           98
#>    2:      Tennessee                        SEC         80          291
#>    3:         Purdue                    Big Ten         98          256
#>    4:          UConn                   Big East         95          206
#>    5:       Syracuse                        ACC         83          190
#>   ---                                                                  
#> 5307:     Seton Hall                   Big East         84           98
#> 5308:     Seton Hall                   Big East         79          122
#> 5309:         Purdue                    Big Ten         63          185
#> 5310:          UConn                   Big East         86            5
#> 5311:          UConn                   Big East         75          236
#>         away_team_name away_conference_short_name away_score provider spread
#>                 <char>                     <char>      <int>   <char>  <num>
#>    1:          Georgia                        SEC         71 ESPN BET   -6.5
#>    2:   Tennessee Tech                        OVC         42 ESPN BET  -32.5
#>    3:          Samford                      SoCon         45 ESPN BET  -19.5
#>    4: Northern Arizona                    Big Sky         52 ESPN BET  -26.5
#>    5:    New Hampshire                   Am. East         72 ESPN BET  -17.5
#>   ---                                                                       
#> 5307:          Georgia                        SEC         67 ESPN BET   -5.5
#> 5308:    Indiana State                        MVC         77 ESPN BET    3.5
#> 5309:         NC State                        ACC         50 ESPN BET   -9.5
#> 5310:          Alabama                        SEC         72 ESPN BET  -10.5
#> 5311:           Purdue                    Big Ten         60 ESPN BET   -6.5
#>       over_under home_moneyline away_moneyline spread_open over_under_open
#>            <num>          <int>          <int>      <lgcl>          <lgcl>
#>    1:      140.5           -260            210          NA              NA
#>    2:      134.5         -20000          10000          NA              NA
#>    3:      143.5          -4000           1800          NA              NA
#>    4:      142.5          -5000           2500          NA              NA
#>    5:      148.5          -2500           1300          NA              NA
#>   ---                                                                     
#> 5307:      145.5           -210            180          NA              NA
#> 5308:      159.5            135           -160          NA              NA
#> 5309:      146.5           -375            295          NA              NA
#> 5310:      159.5           -550            400          NA              NA
#> 5311:      142.5           -285            230          NA              NA

load_lines(TRUE)
#>         game_id season season_type               start_date home_team_id
#>           <int>  <int>      <char>                   <char>        <int>
#>      1:   71898   2013     regular 2012-11-09T13:00:00.000Z          113
#>      2:   71898   2013     regular 2012-11-09T13:00:00.000Z          113
#>      3:   71899   2013     regular 2012-11-09T14:00:00.000Z          115
#>      4:   71899   2013     regular 2012-11-09T14:00:00.000Z          115
#>      5:   71895   2013     regular 2012-11-09T17:00:00.000Z           20
#>     ---                                                                 
#> 158612:  209747   2025  postseason 2025-04-05T20:00:00.000Z          338
#> 158613:  209731   2025  postseason 2025-04-05T22:09:00.000Z           16
#> 158614:  209732   2025  postseason 2025-04-06T01:28:00.000Z           72
#> 158615:  209749   2025  postseason 2025-04-06T21:30:00.000Z          312
#> 158616:  209751   2025  postseason 2025-04-08T00:50:00.000Z          113
#>          home_team_name home_conference_short_name home_score away_team_id
#>                  <char>                     <char>      <int>        <int>
#>      1:         Houston                       CUSA         81           85
#>      2:         Houston                       CUSA         81           85
#>      3: IU Indianapolis                     Summit         67          332
#>      4: IU Indianapolis                     Summit         67          332
#>      5:          Baylor                     Big 12         99          141
#>     ---                                                                   
#> 158612:       Villanova                   Big East         98          312
#> 158613:          Auburn                        SEC         73           87
#> 158614:            Duke                        ACC         67          113
#> 158615:             UCF                     Big 12         66          188
#> 158616:         Houston                     Big 12         63           87
#>         away_team_name away_conference_short_name away_score     provider
#>                 <char>                     <char>      <int>       <char>
#>      1:    Florida A&M                       MEAC         76 teamrankings
#>      2:    Florida A&M                       MEAC         76   numberfire
#>      3:    Utah Valley                 Great West         54 teamrankings
#>      4:    Utah Valley                 Great West         54   numberfire
#>      5:         Lehigh                    Patriot         77 teamrankings
#>     ---                                                                  
#> 158612:            UCF                     Big 12        104     ESPN BET
#> 158613:        Florida                        SEC         79     ESPN BET
#> 158614:        Houston                     Big 12         70     ESPN BET
#> 158615:       Nebraska                    Big Ten         77     ESPN BET
#> 158616:        Florida                        SEC         65     ESPN BET
#>         spread over_under home_moneyline away_moneyline spread_open
#>          <num>      <num>          <int>          <int>       <num>
#>      1:     NA         NA             NA             NA          NA
#>      2:     NA         NA             NA             NA          NA
#>      3:     NA         NA             NA             NA          NA
#>      4:     NA         NA             NA             NA          NA
#>      5:     NA         NA             NA             NA          NA
#>     ---                                                            
#> 158612:   -3.5      156.5             NA             NA        -1.5
#> 158613:    2.5      158.5             NA             NA         2.5
#> 158614:   -4.5      136.5             NA             NA        -4.5
#> 158615:    3.5      160.5             NA             NA         3.5
#> 158616:    1.5      140.5             NA             NA         1.5
#>         over_under_open
#>                   <num>
#>      1:              NA
#>      2:              NA
#>      3:              NA
#>      4:              NA
#>      5:              NA
#>     ---                
#> 158612:           155.5
#> 158613:           159.5
#> 158614:           135.5
#> 158615:           158.5
#> 158616:           140.5
# }