Load media for cbb games

Pull in broadcast information for all games in a given season

Usage

load_media(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of broadcast information for all games for the specified season(s).

See also

Examples

# \donttest{
try({ # avoid cran errors
load_media()
})
#>       game_id season season_label season_type               start_date
#>         <int>  <int>       <char>      <char>                   <char>
#>    1:      24   2025     20242025     regular 2024-11-04T11:00:00.000Z
#>    2:      26   2025     20242025     regular 2024-11-04T12:00:00.000Z
#>    3:      25   2025     20242025     regular 2024-11-04T12:00:00.000Z
#>    4:      28   2025     20242025     regular 2024-11-04T12:30:00.000Z
#>    5:      29   2025     20242025     regular 2024-11-04T15:00:00.000Z
#>   ---                                                                 
#> 6698:  209747   2025     20242025  postseason 2025-04-05T20:00:00.000Z
#> 6699:  209731   2025     20242025  postseason 2025-04-05T22:09:00.000Z
#> 6700:  209732   2025     20242025  postseason 2025-04-06T01:28:00.000Z
#> 6701:  209749   2025     20242025  postseason 2025-04-06T21:30:00.000Z
#> 6702:  209751   2025     20242025  postseason 2025-04-08T00:50:00.000Z
#>       start_time_tbd home_team_id          home_team home_conference_short_name
#>               <lgcl>        <int>             <char>                     <char>
#>    1:          FALSE          115    IU Indianapolis                    Horizon
#>    2:          FALSE          324       UT Arlington                        WAC
#>    3:          FALSE          171   Middle Tennessee                       CUSA
#>    4:          FALSE          128 Jacksonville State                       CUSA
#>    5:          FALSE          261        Santa Clara                        WCC
#>   ---                                                                          
#> 6698:          FALSE          338          Villanova                   Big East
#> 6699:          FALSE           16             Auburn                        SEC
#> 6700:          FALSE           72               Duke                        ACC
#> 6701:          FALSE          312                UCF                     Big 12
#> 6702:          FALSE          113            Houston                     Big 12
#>       away_team_id   away_team away_conference_short_name neutral_site
#>              <int>      <char>                     <char>       <lgcl>
#>    1:          492 IU Columbus                       <NA>        FALSE
#>    2:          494  UNT Dallas                       <NA>        FALSE
#>    3:          493  Oglethorpe                       <NA>        FALSE
#>    4:          495    LaGrange                       <NA>        FALSE
#>    5:          252 Saint Louis                       A-10         TRUE
#>   ---                                                                 
#> 6698:          312         UCF                     Big 12         TRUE
#> 6699:           87     Florida                        SEC         TRUE
#> 6700:          113     Houston                     Big 12         TRUE
#> 6701:          188    Nebraska                    Big Ten         TRUE
#> 6702:           87     Florida                        SEC         TRUE
#>       conference_game game_type
#>                <lgcl>    <char>
#>    1:           FALSE       STD
#>    2:           FALSE       STD
#>    3:           FALSE       STD
#>    4:           FALSE       STD
#>    5:           FALSE    TRNMNT
#>   ---                          
#> 6698:           FALSE    TRNMNT
#> 6699:           FALSE    TRNMNT
#> 6700:           FALSE    TRNMNT
#> 6701:           FALSE    TRNMNT
#> 6702:           FALSE    TRNMNT
#>                                                  game_notes broadcast_type
#>                                                      <char>         <char>
#>    1:                                                  <NA>      Streaming
#>    2:                                                  <NA>      Streaming
#>    3:                                                  <NA>      Streaming
#>    4:                                                  <NA>      Streaming
#>    5:                                  Field of 68 Showcase      Streaming
#>   ---                                                                     
#> 6698:                 College Basketball Crown - Semifinals             TV
#> 6699:            Men's Basketball Championship - Final Four             TV
#> 6700:            Men's Basketball Championship - Final Four             TV
#> 6701:            College Basketball Crown Championship Game             TV
#> 6702: Men's Basketball Championship - National Championship             TV
#>       broadcast_name
#>               <char>
#>    1:          ESPN+
#>    2:          ESPN+
#>    3:          ESPN+
#>    4:          ESPN+
#>    5:        YouTube
#>   ---               
#> 6698:            FOX
#> 6699:            CBS
#> 6700:            CBS
#> 6701:            FOX
#> 6702:            CBS
# }