Load cbb player box score stats

Pull in statistics aggregated by player and game for specified season(s)

This data includes both traditional box score stats (such as points, OREB, AST) as well as more advanced stats (such as USG%, ORTG). For more information on each stat, see the data dictionary dictionary_player_box_scores.

Usage

load_player_box_scores(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of player box scores.

See also

Examples

# \donttest{
load_player_box_scores()
#>        game_id season season_label season_type tournament
#>          <int>  <int>       <char>      <char>     <char>
#>     1:      24   2025     20242025     regular       <NA>
#>     2:      24   2025     20242025     regular       <NA>
#>     3:      24   2025     20242025     regular       <NA>
#>     4:      24   2025     20242025     regular       <NA>
#>     5:      24   2025     20242025     regular       <NA>
#>    ---                                                   
#> 63705:  209751   2025     20242025  postseason       NCAA
#> 63706:  209751   2025     20242025  postseason       NCAA
#> 63707:  209751   2025     20242025  postseason       NCAA
#> 63708:  209751   2025     20242025  postseason       NCAA
#> 63709:  209751   2025     20242025  postseason       NCAA
#>                      start_date start_time_tbd conference_game neutral_site
#>                          <char>         <lgcl>          <lgcl>       <lgcl>
#>     1: 2024-11-04T11:00:00.000Z          FALSE           FALSE        FALSE
#>     2: 2024-11-04T11:00:00.000Z          FALSE           FALSE        FALSE
#>     3: 2024-11-04T11:00:00.000Z          FALSE           FALSE        FALSE
#>     4: 2024-11-04T11:00:00.000Z          FALSE           FALSE        FALSE
#>     5: 2024-11-04T11:00:00.000Z          FALSE           FALSE        FALSE
#>    ---                                                                     
#> 63705: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#> 63706: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#> 63707: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#> 63708: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#> 63709: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#>        is_home game_type                                            game_notes
#>         <lgcl>    <char>                                                <char>
#>     1:   FALSE       STD                                                  <NA>
#>     2:   FALSE       STD                                                  <NA>
#>     3:   FALSE       STD                                                  <NA>
#>     4:   FALSE       STD                                                  <NA>
#>     5:   FALSE       STD                                                  <NA>
#>    ---                                                                        
#> 63705:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#> 63706:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#> 63707:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#> 63708:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#> 63709:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#>        team_id   team_name team_conference_short_name team_seed opponent_id
#>          <int>      <char>                     <char>     <int>       <int>
#>     1:     492 IU Columbus                       <NA>        NA         115
#>     2:     492 IU Columbus                       <NA>        NA         115
#>     3:     492 IU Columbus                       <NA>        NA         115
#>     4:     492 IU Columbus                       <NA>        NA         115
#>     5:     492 IU Columbus                       <NA>        NA         115
#>    ---                                                                     
#> 63705:      87     Florida                        SEC        NA         113
#> 63706:      87     Florida                        SEC        NA         113
#> 63707:      87     Florida                        SEC        NA         113
#> 63708:      87     Florida                        SEC        NA         113
#> 63709:      87     Florida                        SEC        NA         113
#>          opponent_name opponent_conference_short_name opponent_seed
#>                 <char>                         <char>         <int>
#>     1: IU Indianapolis                        Horizon            NA
#>     2: IU Indianapolis                        Horizon            NA
#>     3: IU Indianapolis                        Horizon            NA
#>     4: IU Indianapolis                        Horizon            NA
#>     5: IU Indianapolis                        Horizon            NA
#>    ---                                                             
#> 63705:         Houston                         Big 12            NA
#> 63706:         Houston                         Big 12            NA
#> 63707:         Houston                         Big 12            NA
#> 63708:         Houston                         Big 12            NA
#> 63709:         Houston                         Big 12            NA
#>        game_minutes game_pace athlete_id athlete_source_id       athlete_name
#>               <int>     <num>      <int>            <char>             <char>
#>     1:           40        73        704           5246619      Gavin Dowling
#>     2:           40        73        705           5246612           Owen Law
#>     3:           40        73        706           5246607  Damontae Thompson
#>     4:           40        73        707           5246604     Nathan Pickett
#>     5:           40        73        708           5246603      Deonta Luster
#>    ---                                                                       
#> 63705:           40        67    4255910           4702656      Alijah Martin
#> 63706:           40        67    4255909           4896372 Walter Clayton Jr.
#> 63707:           40        67    4255908           4897262       Will Richard
#> 63708:           40        67    4255907           5174971    Rueben Chinyelu
#> 63709:           40        67    4255906           5174657        Alex Condon
#>        athlete_position starter ejected minutes points turnovers fouls assists
#>                  <char>  <lgcl>  <lgcl>   <int>  <int>     <int> <int>   <int>
#>     1:              ATH    TRUE   FALSE      31      9         2     3       4
#>     2:              ATH    TRUE   FALSE      21     13         1     2       0
#>     3:              ATH    TRUE   FALSE      25     10         3     1       1
#>     4:              ATH    TRUE   FALSE      22      0         2     0       0
#>     5:              ATH    TRUE   FALSE      17      0         0     4       1
#>    ---                                                                        
#> 63705:                G    TRUE   FALSE      34      7         1     0       0
#> 63706:                G    TRUE   FALSE      35     11         3     2       7
#> 63707:                G    TRUE   FALSE      33     18         2     2       2
#> 63708:                C    TRUE   FALSE      19      3         0     2       0
#> 63709:                F    TRUE   FALSE      28     12         4     3       0
#>        steals blocks game_score offensive_rating defensive_rating net_rating
#>         <int>  <int>      <num>            <num>            <num>      <num>
#>     1:      0      1        7.4            101.6            138.6      -37.0
#>     2:      2      0        7.4             92.8            127.9      -35.1
#>     3:      0      0        4.3             89.8            143.8      -54.0
#>     4:      0      0       -4.9              4.6            140.7     -136.1
#>     5:      0      0       -5.8              6.6            144.5     -137.9
#>    ---                                                                      
#> 63705:      0      0        0.5             68.6            108.4      -39.8
#> 63706:      1      1        9.5             97.4             94.2        3.2
#> 63707:      2      1       17.1            132.6             85.7       46.9
#> 63708:      0      1        3.8            123.4             95.5       27.9
#> 63709:      4      0        9.0             82.6             75.9        6.7
#>        usage effective_field_goal_pct true_shooting_pct assists_turnover_ratio
#>        <num>                    <num>             <num>                  <num>
#>     1: 0.160                    0.750             0.580                    2.0
#>     2: 0.314                    0.542             0.542                    0.0
#>     3: 0.212                    0.714             0.672                    0.3
#>     4: 0.184                    0.000             0.000                    0.0
#>     5: 0.209                    0.000             0.000                     NA
#>    ---                                                                        
#> 63705: 0.186                    0.250             0.322                    0.0
#> 63706: 0.224                    0.350             0.468                    2.3
#> 63707: 0.222                    0.700             0.765                    1.0
#> 63708: 0.081                    0.500             0.521                     NA
#> 63709: 0.264                    0.556             0.607                    0.0
#>        free_throw_rate offensive_rebound_pct field_goals_made
#>                  <num>                 <num>            <int>
#>     1:           0.667                 0.400                4
#>     2:           0.000                 0.000                5
#>     3:           0.143                 0.000                5
#>     4:           0.000                 0.333                0
#>     5:           0.000                    NA                0
#>    ---                                                       
#> 63705:           0.200                 1.000                2
#> 63706:           0.400                 0.000                3
#> 63707:           0.400                 0.125                5
#> 63708:           1.000                 0.400                1
#> 63709:           0.222                 0.143                5
#>        field_goals_attempted field_goals_pct two_point_field_goals_made
#>                        <int>           <num>                      <int>
#>     1:                     6           0.667                          3
#>     2:                    12           0.417                          2
#>     3:                     7           0.714                          5
#>     4:                     6           0.000                          0
#>     5:                     7           0.000                          0
#>    ---                                                                 
#> 63705:                    10           0.200                          1
#> 63706:                    10           0.300                          2
#> 63707:                    10           0.500                          1
#> 63708:                     2           0.500                          1
#> 63709:                     9           0.556                          5
#>        two_point_field_goals_attempted two_point_field_goals_pct
#>                                  <int>                     <num>
#>     1:                               5                     0.600
#>     2:                               7                     0.286
#>     3:                               7                     0.714
#>     4:                               2                     0.000
#>     5:                               4                     0.000
#>    ---                                                          
#> 63705:                               4                     0.250
#> 63706:                               3                     0.667
#> 63707:                               3                     0.333
#> 63708:                               2                     0.500
#> 63709:                               8                     0.625
#>        three_point_field_goals_made three_point_field_goals_attempted
#>                               <int>                             <int>
#>     1:                            1                                 1
#>     2:                            3                                 5
#>     3:                            0                                 0
#>     4:                            0                                 4
#>     5:                            0                                 3
#>    ---                                                               
#> 63705:                            1                                 6
#> 63706:                            1                                 7
#> 63707:                            4                                 7
#> 63708:                            0                                 0
#> 63709:                            0                                 1
#>        three_point_field_goals_pct free_throws_made free_throws_attempted
#>                              <num>            <int>                 <int>
#>     1:                       1.000                0                     4
#>     2:                       0.600                0                     0
#>     3:                       0.000                0                     1
#>     4:                       0.000                0                     0
#>     5:                       0.000                0                     0
#>    ---                                                                   
#> 63705:                       0.167                2                     2
#> 63706:                       0.143                4                     4
#> 63707:                       0.571                4                     4
#> 63708:                       0.000                1                     2
#> 63709:                       0.000                2                     2
#>        free_throws_pct rebounds_offensive rebounds_defensive rebounds_total
#>                  <num>              <int>              <int>          <int>
#>     1:             0.0                  2                  3              5
#>     2:             0.0                  0                  2              2
#>     3:             0.0                  0                  1              1
#>     4:             0.0                  1                  2              3
#>     5:             0.0                  0                  0              0
#>    ---                                                                     
#> 63705:             1.0                  1                  0              1
#> 63706:             1.0                  0                  5              5
#> 63707:             1.0                  1                  7              8
#> 63708:             0.5                  2                  3              5
#> 63709:             1.0                  1                  6              7

load_player_box_scores(2024)
#>        game_id season season_label season_type tournament
#>          <int>  <int>       <char>      <char>     <char>
#>     1:    5901   2024     20232024     regular       <NA>
#>     2:    5901   2024     20232024     regular       <NA>
#>     3:    5901   2024     20232024     regular       <NA>
#>     4:    5901   2024     20232024     regular       <NA>
#>     5:    5901   2024     20232024     regular       <NA>
#>    ---                                                   
#> 62928:   12128   2024     20232024  postseason       NCAA
#> 62929:   12128   2024     20232024  postseason       NCAA
#> 62930:   12128   2024     20232024  postseason       NCAA
#> 62931:   12128   2024     20232024  postseason       NCAA
#> 62932:   12128   2024     20232024  postseason       NCAA
#>                      start_date start_time_tbd conference_game neutral_site
#>                          <char>         <lgcl>          <lgcl>       <lgcl>
#>     1: 2023-11-06T11:00:00.000Z          FALSE           FALSE        FALSE
#>     2: 2023-11-06T11:00:00.000Z          FALSE           FALSE        FALSE
#>     3: 2023-11-06T11:00:00.000Z          FALSE           FALSE        FALSE
#>     4: 2023-11-06T11:00:00.000Z          FALSE           FALSE        FALSE
#>     5: 2023-11-06T11:00:00.000Z          FALSE           FALSE        FALSE
#>    ---                                                                     
#> 62928: 2024-04-08T21:20:00.000Z          FALSE           FALSE         TRUE
#> 62929: 2024-04-08T21:20:00.000Z          FALSE           FALSE         TRUE
#> 62930: 2024-04-08T21:20:00.000Z          FALSE           FALSE         TRUE
#> 62931: 2024-04-08T21:20:00.000Z          FALSE           FALSE         TRUE
#> 62932: 2024-04-08T21:20:00.000Z          FALSE           FALSE         TRUE
#>        is_home game_type                                            game_notes
#>         <lgcl>    <char>                                                <char>
#>     1:   FALSE       STD                                                  <NA>
#>     2:   FALSE       STD                                                  <NA>
#>     3:   FALSE       STD                                                  <NA>
#>     4:   FALSE       STD                                                  <NA>
#>     5:   FALSE       STD                                                  <NA>
#>    ---                                                                        
#> 62928:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#> 62929:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#> 62930:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#> 62931:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#> 62932:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#>        team_id team_name team_conference_short_name team_seed opponent_id
#>          <int>    <char>                     <char>     <int>       <int>
#>     1:     831  Spalding                       <NA>        NA         115
#>     2:     831  Spalding                       <NA>        NA         115
#>     3:     831  Spalding                       <NA>        NA         115
#>     4:     831  Spalding                       <NA>        NA         115
#>     5:     831  Spalding                       <NA>        NA         115
#>    ---                                                                   
#> 62928:     236    Purdue                    Big Ten         1         314
#> 62929:     236    Purdue                    Big Ten         1         314
#> 62930:     236    Purdue                    Big Ten         1         314
#> 62931:     236    Purdue                    Big Ten         1         314
#> 62932:     236    Purdue                    Big Ten         1         314
#>          opponent_name opponent_conference_short_name opponent_seed
#>                 <char>                         <char>         <int>
#>     1: IU Indianapolis                        Horizon            NA
#>     2: IU Indianapolis                        Horizon            NA
#>     3: IU Indianapolis                        Horizon            NA
#>     4: IU Indianapolis                        Horizon            NA
#>     5: IU Indianapolis                        Horizon            NA
#>    ---                                                             
#> 62928:           UConn                       Big East             1
#> 62929:           UConn                       Big East             1
#> 62930:           UConn                       Big East             1
#> 62931:           UConn                       Big East             1
#> 62932:           UConn                       Big East             1
#>        game_minutes game_pace athlete_id athlete_source_id       athlete_name
#>               <int>     <num>      <int>            <char>             <char>
#>     1:           40        69     124589           5118156 Devonte McCampbell
#>     2:           40        69     124590           5118155    Anthony Dillard
#>     3:           40        69     124591           5118152     Andre Crawford
#>     4:           40        69     124592           4905941     Jacoby Johnson
#>     5:           40        69     124593           5118151     Caden Bradford
#>    ---                                                                       
#> 62928:           40        61     322544           4431753       Mason Gillis
#> 62929:           40        61     322545           4894452       Myles Colvin
#> 62930:           40        61     322546           4702334       Chase Martin
#> 62931:           40        61     322547           4702333     Carson Barrett
#> 62932:           40        61     322548           4433151       Ethan Morton
#>        athlete_position starter ejected minutes points turnovers fouls assists
#>                  <char>  <lgcl>  <lgcl>   <int>  <int>     <int> <int>   <int>
#>     1:                F    TRUE   FALSE      29      6         1     4       2
#>     2:                F    TRUE   FALSE      32     19         3     1       0
#>     3:                F    TRUE   FALSE      28     17         3     1       2
#>     4:                G    TRUE   FALSE      32     11         6     3       2
#>     5:                G    TRUE   FALSE      16      2         1     2       1
#>    ---                                                                        
#> 62928:                F   FALSE   FALSE      24      0         0     1       0
#> 62929:                G   FALSE   FALSE      12      0         1     0       0
#> 62930:                G   FALSE   FALSE       1      0         0     0       0
#> 62931:                G   FALSE   FALSE       1      0         0     0       0
#> 62932:                G   FALSE   FALSE       1      0         0     0       0
#>        steals blocks game_score offensive_rating defensive_rating net_rating
#>         <int>  <int>      <num>            <num>            <num>      <num>
#>     1:      1      0        5.0            108.4             99.1        9.3
#>     2:      2      0       13.2            106.3             98.4        7.9
#>     3:      0      1       12.6            113.0            103.6        9.4
#>     4:      2      1        2.8             75.4            101.7      -26.3
#>     5:      0      0        0.6             87.2            112.1      -24.9
#>    ---                                                                      
#> 62928:      0      0        0.2             44.9            126.9      -82.0
#> 62929:      0      0       -1.8              0.0            121.7     -121.7
#> 62930:      0      0        0.0               NA               NA         NA
#> 62931:      0      0        0.0               NA               NA         NA
#> 62932:      0      0        0.0               NA               NA         NA
#>        usage effective_field_goal_pct true_shooting_pct assists_turnover_ratio
#>        <num>                    <num>             <num>                  <num>
#>     1: 0.114                     0.50             0.500                    2.0
#>     2: 0.306                     0.50             0.535                    0.0
#>     3: 0.307                     0.50             0.559                    0.7
#>     4: 0.249                     0.45             0.506                    0.3
#>     5: 0.089                     0.50             0.500                    1.0
#>    ---                                                                        
#> 62928: 0.048                     0.00             0.000                     NA
#> 62929: 0.144                     0.00             0.000                    0.0
#> 62930:    NA                       NA                NA                     NA
#> 62931:    NA                       NA                NA                     NA
#> 62932:    NA                       NA                NA                     NA
#>        free_throw_rate offensive_rebound_pct field_goals_made
#>                  <num>                 <num>            <int>
#>     1:           0.000                 0.167                3
#>     2:           0.250                 0.500                8
#>     3:           0.385                 0.500                6
#>     4:           0.200                 0.000                4
#>     5:           0.000                 1.000                1
#>    ---                                                       
#> 62928:           0.000                 0.500                0
#> 62929:           0.000                 0.000                0
#> 62930:              NA                    NA                0
#> 62931:              NA                    NA                0
#> 62932:              NA                    NA                0
#>        field_goals_attempted field_goals_pct two_point_field_goals_made
#>                        <int>           <num>                      <int>
#>     1:                     6           0.500                          3
#>     2:                    16           0.500                          8
#>     3:                    13           0.462                          5
#>     4:                    10           0.400                          3
#>     5:                     2           0.500                          1
#>    ---                                                                 
#> 62928:                     2           0.000                          0
#> 62929:                     2           0.000                          0
#> 62930:                     0           0.000                          0
#> 62931:                     0           0.000                          0
#> 62932:                     0           0.000                          0
#>        two_point_field_goals_attempted two_point_field_goals_pct
#>                                  <int>                     <num>
#>     1:                               6                     0.500
#>     2:                              15                     0.533
#>     3:                              10                     0.500
#>     4:                               9                     0.333
#>     5:                               1                     1.000
#>    ---                                                          
#> 62928:                               1                     0.000
#> 62929:                               0                     0.000
#> 62930:                               0                     0.000
#> 62931:                               0                     0.000
#> 62932:                               0                     0.000
#>        three_point_field_goals_made three_point_field_goals_attempted
#>                               <int>                             <int>
#>     1:                            0                                 0
#>     2:                            0                                 1
#>     3:                            1                                 3
#>     4:                            1                                 1
#>     5:                            0                                 1
#>    ---                                                               
#> 62928:                            0                                 1
#> 62929:                            0                                 2
#> 62930:                            0                                 0
#> 62931:                            0                                 0
#> 62932:                            0                                 0
#>        three_point_field_goals_pct free_throws_made free_throws_attempted
#>                              <num>            <int>                 <int>
#>     1:                       0.000                0                     0
#>     2:                       0.000                3                     4
#>     3:                       0.333                4                     5
#>     4:                       1.000                2                     2
#>     5:                       0.000                0                     0
#>    ---                                                                   
#> 62928:                       0.000                0                     0
#> 62929:                       0.000                0                     0
#> 62930:                       0.000                0                     0
#> 62931:                       0.000                0                     0
#> 62932:                       0.000                0                     0
#>        free_throws_pct rebounds_offensive rebounds_defensive rebounds_total
#>                  <num>              <int>              <int>          <int>
#>     1:            0.00                  1                  5              6
#>     2:            0.75                  4                  4              8
#>     3:            0.80                  4                  4              8
#>     4:            1.00                  0                  1              1
#>     5:            0.00                  1                  0              1
#>    ---                                                                     
#> 62928:            0.00                  2                  2              4
#> 62929:            0.00                  0                  2              2
#> 62930:            0.00                  0                  0              0
#> 62931:            0.00                  0                  0              0
#> 62932:            0.00                  0                  0              0

load_player_box_scores(TRUE)
#>          game_id season season_label season_type tournament
#>            <int>  <int>       <char>      <char>     <char>
#>       1:  131858   2003     20022003     regular       <NA>
#>       2:  131858   2003     20022003     regular       <NA>
#>       3:  131858   2003     20022003     regular       <NA>
#>       4:  131858   2003     20022003     regular       <NA>
#>       5:  131858   2003     20022003     regular       <NA>
#>      ---                                                   
#> 1170922:  209751   2025     20242025  postseason       NCAA
#> 1170923:  209751   2025     20242025  postseason       NCAA
#> 1170924:  209751   2025     20242025  postseason       NCAA
#> 1170925:  209751   2025     20242025  postseason       NCAA
#> 1170926:  209751   2025     20242025  postseason       NCAA
#>                        start_date start_time_tbd conference_game neutral_site
#>                            <char>         <lgcl>          <lgcl>       <lgcl>
#>       1: 2003-03-10T00:00:00.000Z          FALSE            TRUE        FALSE
#>       2: 2003-03-10T00:00:00.000Z          FALSE            TRUE        FALSE
#>       3: 2003-03-10T00:00:00.000Z          FALSE            TRUE        FALSE
#>       4: 2003-03-10T00:00:00.000Z          FALSE            TRUE        FALSE
#>       5: 2003-03-10T00:00:00.000Z          FALSE            TRUE        FALSE
#>      ---                                                                     
#> 1170922: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#> 1170923: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#> 1170924: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#> 1170925: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#> 1170926: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#>          is_home game_type
#>           <lgcl>    <char>
#>       1:   FALSE       STD
#>       2:   FALSE       STD
#>       3:   FALSE       STD
#>       4:   FALSE       STD
#>       5:   FALSE       STD
#>      ---                  
#> 1170922:   FALSE    TRNMNT
#> 1170923:   FALSE    TRNMNT
#> 1170924:   FALSE    TRNMNT
#> 1170925:   FALSE    TRNMNT
#> 1170926:   FALSE    TRNMNT
#>                                                     game_notes team_id
#>                                                         <char>   <int>
#>       1:                                    WEST COAST CONF TR     253
#>       2:                                    WEST COAST CONF TR     253
#>       3:                                    WEST COAST CONF TR     253
#>       4:                                    WEST COAST CONF TR     253
#>       5:                                    WEST COAST CONF TR     253
#>      ---                                                              
#> 1170922: Men's Basketball Championship - National Championship      87
#> 1170923: Men's Basketball Championship - National Championship      87
#> 1170924: Men's Basketball Championship - National Championship      87
#> 1170925: Men's Basketball Championship - National Championship      87
#> 1170926: Men's Basketball Championship - National Championship      87
#>             team_name team_conference_short_name team_seed opponent_id
#>                <char>                     <char>     <int>       <int>
#>       1: Saint Mary's                        WCC        NA         102
#>       2: Saint Mary's                        WCC        NA         102
#>       3: Saint Mary's                        WCC        NA         102
#>       4: Saint Mary's                        WCC        NA         102
#>       5: Saint Mary's                        WCC        NA         102
#>      ---                                                              
#> 1170922:      Florida                        SEC        NA         113
#> 1170923:      Florida                        SEC        NA         113
#> 1170924:      Florida                        SEC        NA         113
#> 1170925:      Florida                        SEC        NA         113
#> 1170926:      Florida                        SEC        NA         113
#>          opponent_name opponent_conference_short_name opponent_seed
#>                 <char>                         <char>         <int>
#>       1:       Gonzaga                            WCC            NA
#>       2:       Gonzaga                            WCC            NA
#>       3:       Gonzaga                            WCC            NA
#>       4:       Gonzaga                            WCC            NA
#>       5:       Gonzaga                            WCC            NA
#>      ---                                                           
#> 1170922:       Houston                         Big 12            NA
#> 1170923:       Houston                         Big 12            NA
#> 1170924:       Houston                         Big 12            NA
#> 1170925:       Houston                         Big 12            NA
#> 1170926:       Houston                         Big 12            NA
#>          game_minutes game_pace athlete_id athlete_source_id       athlete_name
#>                 <int>     <num>      <int>            <char>             <char>
#>       1:           40        NA    3770993              3598         Tyler Herr
#>       2:           40        NA    3770992              3686   Jonathan Sanders
#>       3:           40        NA    3770991              6591  Samuel Saint-Jean
#>       4:           40        NA    3770990              6593        Adam Caporn
#>       5:           40        NA    3770989              6594          Bill Enea
#>      ---                                                                       
#> 1170922:           40        67    4255910           4702656      Alijah Martin
#> 1170923:           40        67    4255909           4896372 Walter Clayton Jr.
#> 1170924:           40        67    4255908           4897262       Will Richard
#> 1170925:           40        67    4255907           5174971    Rueben Chinyelu
#> 1170926:           40        67    4255906           5174657        Alex Condon
#>          athlete_position starter ejected minutes points turnovers fouls
#>                    <char>  <lgcl>  <lgcl>   <int>  <int>     <int> <int>
#>       1:                G   FALSE   FALSE      20     10         1     0
#>       2:                G   FALSE   FALSE      28      2         1     3
#>       3:                G   FALSE   FALSE      23      2         3     4
#>       4:                G   FALSE   FALSE      31      3         3     3
#>       5:                G   FALSE   FALSE       2      0         0     0
#>      ---                                                                
#> 1170922:                G    TRUE   FALSE      34      7         1     0
#> 1170923:                G    TRUE   FALSE      35     11         3     2
#> 1170924:                G    TRUE   FALSE      33     18         2     2
#> 1170925:                C    TRUE   FALSE      19      3         0     2
#> 1170926:                F    TRUE   FALSE      28     12         4     3
#>          assists steals blocks game_score offensive_rating defensive_rating
#>            <int>  <int>  <int>      <num>            <num>            <num>
#>       1:       0      0      0        7.3            149.4               NA
#>       2:       3      0      0        0.4             69.8               NA
#>       3:       2      0      0       -3.0             45.3               NA
#>       4:       1      1      0       -1.7             45.8               NA
#>       5:       0      0      0       -0.7               NA               NA
#>      ---                                                                   
#> 1170922:       0      0      0        0.5             68.6            108.4
#> 1170923:       7      1      1        9.5             97.4             94.2
#> 1170924:       2      2      1       17.1            132.6             85.7
#> 1170925:       0      0      1        3.8            123.4             95.5
#> 1170926:       0      4      0        9.0             82.6             75.9
#>          net_rating usage effective_field_goal_pct true_shooting_pct
#>               <num> <num>                    <num>             <num>
#>       1:         NA 0.173                    1.000             0.940
#>       2:         NA 0.098                    0.250             0.250
#>       3:         NA 0.140                    0.500             0.347
#>       4:         NA 0.147                    0.250             0.282
#>       5:         NA    NA                    0.000             0.000
#>      ---                                                            
#> 1170922:      -39.8 0.186                    0.250             0.322
#> 1170923:        3.2 0.224                    0.350             0.468
#> 1170924:       46.9 0.222                    0.700             0.765
#> 1170925:       27.9 0.081                    0.500             0.521
#> 1170926:        6.7 0.264                    0.556             0.607
#>          assists_turnover_ratio free_throw_rate offensive_rebound_pct
#>                           <num>           <num>                 <num>
#>       1:                    0.0           0.750                 0.000
#>       2:                    3.0           0.000                 0.000
#>       3:                    0.7           1.000                    NA
#>       4:                    0.3           0.750                 0.500
#>       5:                     NA           0.000                    NA
#>      ---                                                             
#> 1170922:                    0.0           0.200                 1.000
#> 1170923:                    2.3           0.400                 0.000
#> 1170924:                    1.0           0.400                 0.125
#> 1170925:                     NA           1.000                 0.400
#> 1170926:                    0.0           0.222                 0.143
#>          field_goals_made field_goals_attempted field_goals_pct
#>                     <int>                 <int>           <num>
#>       1:                3                     4           0.750
#>       2:                1                     4           0.250
#>       3:                1                     2           0.500
#>       4:                1                     4           0.250
#>       5:                0                     1           0.000
#>      ---                                                       
#> 1170922:                2                    10           0.200
#> 1170923:                3                    10           0.300
#> 1170924:                5                    10           0.500
#> 1170925:                1                     2           0.500
#> 1170926:                5                     9           0.556
#>          two_point_field_goals_made two_point_field_goals_attempted
#>                               <int>                           <int>
#>       1:                          1                               1
#>       2:                          1                               3
#>       3:                          1                               2
#>       4:                          1                               3
#>       5:                          0                               1
#>      ---                                                           
#> 1170922:                          1                               4
#> 1170923:                          2                               3
#> 1170924:                          1                               3
#> 1170925:                          1                               2
#> 1170926:                          5                               8
#>          two_point_field_goals_pct three_point_field_goals_made
#>                              <num>                        <int>
#>       1:                     1.000                            2
#>       2:                     0.333                            0
#>       3:                     0.500                            0
#>       4:                     0.333                            0
#>       5:                     0.000                            0
#>      ---                                                       
#> 1170922:                     0.250                            1
#> 1170923:                     0.667                            1
#> 1170924:                     0.333                            4
#> 1170925:                     0.500                            0
#> 1170926:                     0.625                            0
#>          three_point_field_goals_attempted three_point_field_goals_pct
#>                                      <int>                       <num>
#>       1:                                 3                       0.667
#>       2:                                 1                       0.000
#>       3:                                 0                       0.000
#>       4:                                 1                       0.000
#>       5:                                 0                       0.000
#>      ---                                                              
#> 1170922:                                 6                       0.167
#> 1170923:                                 7                       0.143
#> 1170924:                                 7                       0.571
#> 1170925:                                 0                       0.000
#> 1170926:                                 1                       0.000
#>          free_throws_made free_throws_attempted free_throws_pct
#>                     <int>                 <int>           <num>
#>       1:                2                     3           0.667
#>       2:                0                     0           0.000
#>       3:                0                     2           0.000
#>       4:                1                     3           0.333
#>       5:                0                     0           0.000
#>      ---                                                       
#> 1170922:                2                     2           1.000
#> 1170923:                4                     4           1.000
#> 1170924:                4                     4           1.000
#> 1170925:                1                     2           0.500
#> 1170926:                2                     2           1.000
#>          rebounds_offensive rebounds_defensive rebounds_total
#>                       <int>              <int>          <int>
#>       1:                  0                  1              1
#>       2:                  0                  3              3
#>       3:                  0                  0              0
#>       4:                  1                  1              2
#>       5:                  0                  0              0
#>      ---                                                     
#> 1170922:                  1                  0              1
#> 1170923:                  0                  5              5
#> 1170924:                  1                  7              8
#> 1170925:                  2                  3              5
#> 1170926:                  1                  6              7
# }