Load cbb player stats

Pull in statistics aggregated by player for specified season(s)

This data includes both traditional box score stats (such as points, OREB, AST) as well as more advanced stats (such as USG%, ORTG). For more information on each stat, see the data dictionary dictionary_player_stats.

Usage

load_player_stats(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of player statistics for college basketball players

See also

Examples

# \donttest{
load_player_stats()
#>       season season_label team_id                 team conference athlete_id
#>        <int>       <char>   <int>               <char>     <char>      <int>
#>    1:   2025     20242025       1    Abilene Christian        WAC       6045
#>    2:   2025     20242025       1    Abilene Christian        WAC       6046
#>    3:   2025     20242025       1    Abilene Christian        WAC       6047
#>    4:   2025     20242025       1    Abilene Christian        WAC       6048
#>    5:   2025     20242025       1    Abilene Christian        WAC       6049
#>   ---                                                                       
#> 9813:   2025     20242025     830 Washington Adventist       <NA>      15168
#> 9814:   2025     20242025     830 Washington Adventist       <NA>      15169
#> 9815:   2025     20242025     830 Washington Adventist       <NA>      15170
#> 9816:   2025     20242025     830 Washington Adventist       <NA>      15171
#> 9817:   2025     20242025     830 Washington Adventist       <NA>      15172
#>       athlete_source_id               name position games starts minutes points
#>                  <char>             <char>   <char> <int>  <int>   <int>  <int>
#>    1:           4896505   Leonardo Bettiol        F    32     31     828    419
#>    2:           5177704       Nasir DeGruy        G    32      9     579    117
#>    3:           5106274     Quion Williams        G    30     30     888    411
#>    4:           4897572     Joseph Venzant        G     8      6     157     13
#>    5:           4701572 Hunter Jack Madden        G    32     29     883    285
#>   ---                                                                          
#> 9813:           5264286     Emanuel Hylton      ATH     1      1      34     15
#> 9814:           5264285  Jayden Rowe-Perry      ATH     1      1      29      8
#> 9815:           5264300       Mateo Cataid      ATH     1      0       3      0
#> 9816:           5264299        James Quinn      ATH     1      0      23      4
#> 9817:           5264294       Felix Mcghee      ATH     1      0      34     10
#>       turnovers fouls assists steals blocks offensive_rating defensive_rating
#>           <int> <int>   <int>  <int>  <int>            <num>            <num>
#>    1:        61   111      37     33     13            107.3            103.0
#>    2:        49    59      42     42      3             76.6            100.4
#>    3:        77    69      75     35     22            102.0            103.1
#>    4:         8    17       4      7      1             70.1            110.9
#>    5:        65    45      68     41      4             98.1            104.6
#>   ---                                                                        
#> 9813:         4     1       1      1      0             59.1            127.3
#> 9814:         4     4       0      2      0             46.8            121.2
#> 9815:         1     1       1      0      0               NA            130.9
#> 9816:         3     2       3      0      0             75.2            129.7
#> 9817:         5     1       0      2      0             78.0            124.2
#>       net_rating porpag usage assists_turnover_ratio offensive_rebound_pct
#>            <num>  <num> <num>                  <num>                 <num>
#>    1:        4.3    2.1  26.0                   0.61                  38.1
#>    2:      -23.8   -0.5  16.3                   0.86                  15.7
#>    3:       -1.1    1.7  26.2                   0.97                  32.7
#>    4:      -40.8   -0.4   7.1                   0.50                  12.5
#>    5:       -6.5    0.9  19.5                   1.05                  25.3
#>   ---                                                                     
#> 9813:      -68.2   -5.4  33.6                   0.25                  20.0
#> 9814:      -74.4   -5.5  28.1                   0.00                  16.7
#> 9815:         NA     NA  14.8                   1.00                    NA
#> 9816:      -54.5   -0.5  11.3                   1.00                  25.0
#> 9817:      -46.2   -0.8  14.8                   0.00                   0.0
#>       free_throw_rate effective_field_goal_pct true_shooting_pct
#>                 <num>                    <num>             <num>
#>    1:            40.9                     52.6               0.6
#>    2:            21.3                     38.2               0.4
#>    3:            35.7                     49.2               0.5
#>    4:           120.0                     40.0               0.4
#>    5:            27.7                     44.6               0.5
#>   ---                                                           
#> 9813:            20.0                     32.5               0.3
#> 9814:           100.0                     15.0               0.3
#> 9815:              NA                       NA                NA
#> 9816:           100.0                     75.0               0.7
#> 9817:            60.0                     70.0               0.8
#>       field_goals_made field_goals_attempted field_goals_pct
#>                  <int>                 <int>           <num>
#>    1:              164                   313            52.4
#>    2:               46                   127            36.2
#>    3:              155                   333            46.5
#>    4:                4                    10            40.0
#>    5:               85                   249            34.1
#>   ---                                                       
#> 9813:                5                    20            25.0
#> 9814:                1                    10            10.0
#> 9815:                0                     0             0.0
#> 9816:                1                     2            50.0
#> 9817:                3                     5            60.0
#>       two_point_field_goals_made two_point_field_goals_attempted
#>                            <int>                           <int>
#>    1:                        163                             304
#>    2:                         41                              98
#>    3:                        137                             279
#>    4:                          4                              10
#>    5:                         33                              78
#>   ---                                                           
#> 9813:                          2                              11
#> 9814:                          0                               6
#> 9815:                          0                               0
#> 9816:                          0                               1
#> 9817:                          2                               2
#>       two_point_field_goals_pct three_point_field_goals_made
#>                           <num>                        <int>
#>    1:                      53.6                            1
#>    2:                      41.8                            5
#>    3:                      49.1                           18
#>    4:                      40.0                            0
#>    5:                      42.3                           52
#>   ---                                                       
#> 9813:                      18.2                            3
#> 9814:                       0.0                            1
#> 9815:                       0.0                            0
#> 9816:                       0.0                            1
#> 9817:                     100.0                            1
#>       three_point_field_goals_attempted three_point_field_goals_pct
#>                                   <int>                       <num>
#>    1:                                 9                        11.1
#>    2:                                29                        17.2
#>    3:                                54                        33.3
#>    4:                                 0                         0.0
#>    5:                               171                        30.4
#>   ---                                                              
#> 9813:                                 9                        33.3
#> 9814:                                 4                        25.0
#> 9815:                                 0                         0.0
#> 9816:                                 1                       100.0
#> 9817:                                 3                        33.3
#>       free_throws_made free_throws_attempted free_throws_pct rebounds_offensive
#>                  <int>                 <int>           <num>              <int>
#>    1:               90                   128            70.3                 48
#>    2:               20                    27            74.1                  8
#>    3:               83                   119            69.7                 50
#>    4:                5                    12            41.7                  4
#>    5:               63                    69            91.3                 19
#>   ---                                                                          
#> 9813:                2                     4            50.0                  1
#> 9814:                5                    10            50.0                  1
#> 9815:                0                     0             0.0                  0
#> 9816:                1                     2            50.0                  1
#> 9817:                3                     3           100.0                  0
#>       rebounds_defensive rebounds_total win_shares_offensive
#>                    <int>          <int>                <num>
#>    1:                 78            126                  1.5
#>    2:                 43             51                 -0.7
#>    3:                103            153                  1.1
#>    4:                 28             32                 -0.1
#>    5:                 56             75                  0.5
#>   ---                                                       
#> 9813:                  4              5                 -0.2
#> 9814:                  5              6                 -0.2
#> 9815:                  0              0                   NA
#> 9816:                  3              4                  0.0
#> 9817:                  4              4                  0.0
#>       win_shares_defensive win_shares_total win_shares_total_per40
#>                      <num>            <num>                  <num>
#>    1:                  1.2              2.7                  0.130
#>    2:                  1.0              0.3                  0.021
#>    3:                  1.2              2.3                  0.104
#>    4:                  0.1              0.0                  0.000
#>    5:                  1.1              1.6                  0.072
#>   ---                                                             
#> 9813:                  0.0             -0.2                 -0.235
#> 9814:                  0.0             -0.2                 -0.276
#> 9815:                  0.0               NA                     NA
#> 9816:                  0.0              0.0                  0.000
#> 9817:                  0.0              0.0                  0.000

load_player_stats(2024)
#>       season season_label team_id                         team conference
#>        <int>       <char>   <int>                       <char>     <char>
#>    1:   2024     20232024       1            Abilene Christian        WAC
#>    2:   2024     20232024       1            Abilene Christian        WAC
#>    3:   2024     20232024       1            Abilene Christian        WAC
#>    4:   2024     20232024       1            Abilene Christian        WAC
#>    5:   2024     20232024       1            Abilene Christian        WAC
#>   ---                                                                    
#> 9872:   2024     20232024     972 Indiana University Northwest       <NA>
#> 9873:   2024     20232024     972 Indiana University Northwest       <NA>
#> 9874:   2024     20232024     972 Indiana University Northwest       <NA>
#> 9875:   2024     20232024     972 Indiana University Northwest       <NA>
#> 9876:   2024     20232024     972 Indiana University Northwest       <NA>
#>       athlete_id athlete_source_id               name position games starts
#>            <int>            <char>             <char>   <char> <int>  <int>
#>    1:        747           4896862          Ali Dibba        G    31     30
#>    2:       4056           5177703     Kavion McClain        G    34     30
#>    3:       6045           4896505   Leonardo Bettiol        F    31     22
#>    4:       6046           5177704       Nasir DeGruy        G    34      6
#>    5:       6049           4701572 Hunter Jack Madden        G    34     34
#>   ---                                                                      
#> 9872:      19824           5192074     Ladaion Barnes        F     1      0
#> 9873:      19825           5192077      Adam Eastland        G     1      0
#> 9874:      19826           5192073       Kionne Layne        G     1      0
#> 9875:      19827           5192070    Brian Hernandez        G     1      0
#> 9876:      19828           5192068      Nicky Flesher        G     1      0
#>       minutes points turnovers fouls assists steals blocks offensive_rating
#>         <int>  <int>     <int> <int>   <int>  <int>  <int>            <num>
#>    1:     985    482        72    65      36     43      7            102.5
#>    2:     919    211        68    56     115     41      2             97.7
#>    3:     667    303        41   111      29     12      6            111.9
#>    4:     483    131        31    37      36     17      5             93.2
#>    5:    1108    400        54    67      68     39      1            109.6
#>   ---                                                                      
#> 9872:      21      3         0     4       2      0      0             95.1
#> 9873:       2      0         0     0       0      0      0             22.1
#> 9874:      12      2         0     1       1      1      1             50.6
#> 9875:      16      3         0     4       2      1      0             65.7
#> 9876:      19      9         0     1       0      0      0            137.7
#>       defensive_rating net_rating porpag usage assists_turnover_ratio
#>                  <num>      <num>  <num> <num>                  <num>
#>    1:            110.1       -7.6    1.9  26.3                   0.50
#>    2:            112.3      -14.6    0.6  14.5                   1.69
#>    3:            107.8        4.1    1.9  22.7                   0.71
#>    4:            111.5      -18.3    0.2  17.4                   1.16
#>    5:            113.3       -3.7    2.1  18.8                   1.26
#>   ---                                                                
#> 9872:            122.6      -27.5    0.2   9.8                     NA
#> 9873:            134.4     -112.3   -0.9  41.3                     NA
#> 9874:            116.8      -66.2   -1.8  24.1                     NA
#> 9875:            122.8      -57.1   -1.2  20.7                     NA
#> 9876:            134.3        3.4    2.4  15.2                     NA
#>       offensive_rebound_pct free_throw_rate effective_field_goal_pct
#>                       <num>           <num>                    <num>
#>    1:                  22.0            43.1                     50.0
#>    2:                  10.6            25.9                     49.4
#>    3:                  39.7            48.3                     55.9
#>    4:                  22.0            31.6                     44.0
#>    5:                  21.6            41.7                     49.8
#>   ---                                                               
#> 9872:                  20.0             0.0                     30.0
#> 9873:                 100.0             0.0                      0.0
#> 9874:                  50.0             0.0                     14.3
#> 9875:                  50.0             0.0                     18.8
#> 9876:                    NA             0.0                     64.3
#>       true_shooting_pct field_goals_made field_goals_attempted field_goals_pct
#>                   <num>            <int>                 <int>           <num>
#>    1:               0.6              175                   367            47.7
#>    2:               0.5               78                   174            44.8
#>    3:               0.6              118                   211            55.9
#>    4:               0.5               49                   117            41.9
#>    5:               0.6              118                   300            39.3
#>   ---                                                                         
#> 9872:               0.3                1                     5            20.0
#> 9873:               0.0                0                     2             0.0
#> 9874:               0.1                1                     7            14.3
#> 9875:               0.2                1                     8            12.5
#> 9876:               0.6                3                     7            42.9
#>       two_point_field_goals_made two_point_field_goals_attempted
#>                            <int>                           <int>
#>    1:                        158                             318
#>    2:                         62                             121
#>    3:                        118                             211
#>    4:                         44                              98
#>    5:                         55                             138
#>   ---                                                           
#> 9872:                          0                               2
#> 9873:                          0                               1
#> 9874:                          1                               6
#> 9875:                          0                               2
#> 9876:                          0                               1
#>       two_point_field_goals_pct three_point_field_goals_made
#>                           <num>                        <int>
#>    1:                      49.7                           17
#>    2:                      51.2                           16
#>    3:                      55.9                            0
#>    4:                      44.9                            5
#>    5:                      39.9                           63
#>   ---                                                       
#> 9872:                       0.0                            1
#> 9873:                       0.0                            0
#> 9874:                      16.7                            0
#> 9875:                       0.0                            1
#> 9876:                       0.0                            3
#>       three_point_field_goals_attempted three_point_field_goals_pct
#>                                   <int>                       <num>
#>    1:                                49                        34.7
#>    2:                                53                        30.2
#>    3:                                 0                         0.0
#>    4:                                19                        26.3
#>    5:                               162                        38.9
#>   ---                                                              
#> 9872:                                 3                        33.3
#> 9873:                                 1                         0.0
#> 9874:                                 1                         0.0
#> 9875:                                 6                        16.7
#> 9876:                                 6                        50.0
#>       free_throws_made free_throws_attempted free_throws_pct rebounds_offensive
#>                  <int>                 <int>           <num>              <int>
#>    1:              115                   158            72.8                 31
#>    2:               39                    45            86.7                  7
#>    3:               67                   102            65.7                 62
#>    4:               28                    37            75.7                 13
#>    5:              101                   125            80.8                 22
#>   ---                                                                          
#> 9872:                0                     0             0.0                  1
#> 9873:                0                     0             0.0                  1
#> 9874:                0                     0             0.0                  1
#> 9875:                0                     0             0.0                  1
#> 9876:                0                     0             0.0                  0
#>       rebounds_defensive rebounds_total win_shares_offensive
#>                    <int>          <int>                <num>
#>    1:                110            141                  1.4
#>    2:                 59             66                  0.4
#>    3:                 94            156                  1.5
#>    4:                 46             59                  0.1
#>    5:                 80            102                  1.9
#>   ---                                                       
#> 9872:                  4              5                  0.0
#> 9873:                  0              1                  0.0
#> 9874:                  1              2                  0.0
#> 9875:                  1              2                  0.0
#> 9876:                  0              0                  0.1
#>       win_shares_defensive win_shares_total win_shares_total_per40
#>                      <num>            <num>                  <num>
#>    1:                  0.7              2.1                  0.085
#>    2:                  0.5              0.9                  0.039
#>    3:                  0.6              2.1                  0.126
#>    4:                  0.3              0.4                  0.033
#>    5:                  0.5              2.4                  0.087
#>   ---                                                             
#> 9872:                  0.0              0.0                  0.000
#> 9873:                  0.0              0.0                  0.000
#> 9874:                  0.0              0.0                  0.000
#> 9875:                  0.0              0.0                  0.000
#> 9876:                  0.0              0.1                  0.211

load_player_stats(TRUE)
#>         season season_label team_id                 team conference athlete_id
#>          <int>       <char>   <int>               <char>     <char>      <int>
#>      1:   2003     20022003     102              Gonzaga        WCC      89361
#>      2:   2003     20022003     102              Gonzaga        WCC      90765
#>      3:   2003     20022003     102              Gonzaga        WCC      90766
#>      4:   2003     20022003     102              Gonzaga        WCC      91830
#>      5:   2003     20022003     102              Gonzaga        WCC      92102
#>     ---                                                                       
#> 157937:   2025     20242025     830 Washington Adventist       <NA>      15168
#> 157938:   2025     20242025     830 Washington Adventist       <NA>      15169
#> 157939:   2025     20242025     830 Washington Adventist       <NA>      15170
#> 157940:   2025     20242025     830 Washington Adventist       <NA>      15171
#> 157941:   2025     20242025     830 Washington Adventist       <NA>      15172
#>         athlete_source_id              name position games starts minutes
#>                    <char>            <char>   <char> <int>  <int>   <int>
#>      1:              8021 Dustin Villepigue        F     1      0       1
#>      2:              6585      Ronny Turiaf        F     1      0      17
#>      3:              6580  Brian Michaelson        G     1      0       1
#>      4:              8020    Joshua Reisman        G     1      0       5
#>      5:              3546     Kyle Bankhead      G-F     1      0      30
#>     ---                                                                  
#> 157937:           5264286    Emanuel Hylton      ATH     1      1      34
#> 157938:           5264285 Jayden Rowe-Perry      ATH     1      1      29
#> 157939:           5264300      Mateo Cataid      ATH     1      0       3
#> 157940:           5264299       James Quinn      ATH     1      0      23
#> 157941:           5264294      Felix Mcghee      ATH     1      0      34
#>         points turnovers fouls assists steals blocks offensive_rating
#>          <int>     <int> <int>   <int>  <int>  <int>            <num>
#>      1:      0         0     0       0      0      0               NA
#>      2:      8         0     4       0      2      2             89.3
#>      3:      0         0     0       0      0      0               NA
#>      4:      2         0     0       0      1      0            114.5
#>      5:      5         0     3       5      0      0            145.2
#>     ---                                                              
#> 157937:     15         4     1       1      1      0             59.1
#> 157938:      8         4     4       0      2      0             46.8
#> 157939:      0         1     1       1      0      0               NA
#> 157940:      4         3     2       3      0      0             75.2
#> 157941:     10         5     1       0      2      0             78.0
#>         defensive_rating net_rating porpag usage assists_turnover_ratio
#>                    <num>      <num>  <num> <num>                  <num>
#>      1:               NA         NA     NA   0.0                     NA
#>      2:               NA         NA    0.1  37.7                     NA
#>      3:               NA         NA     NA   0.0                     NA
#>      4:               NA         NA    0.5  21.4                     NA
#>      5:               NA         NA    2.5   8.9                     NA
#>     ---                                                                
#> 157937:            127.3      -68.2   -5.4  33.6                   0.25
#> 157938:            121.2      -74.4   -5.5  28.1                   0.00
#> 157939:            130.9         NA     NA  14.8                   1.00
#> 157940:            129.7      -54.5   -0.5  11.3                   1.00
#> 157941:            124.2      -46.2   -0.8  14.8                   0.00
#>         offensive_rebound_pct free_throw_rate effective_field_goal_pct
#>                         <num>           <num>                    <num>
#>      1:                   0.0              NA                       NA
#>      2:                  60.0               0                     33.3
#>      3:                    NA              NA                       NA
#>      4:                   0.0               0                     50.0
#>      5:                   0.0               0                     50.0
#>     ---                                                               
#> 157937:                  20.0              20                     32.5
#> 157938:                  16.7             100                     15.0
#> 157939:                    NA              NA                       NA
#> 157940:                  25.0             100                     75.0
#> 157941:                   0.0              60                     70.0
#>         true_shooting_pct field_goals_made field_goals_attempted
#>                     <num>            <int>                 <int>
#>      1:                NA                0                     0
#>      2:               0.3                4                    12
#>      3:                NA                0                     0
#>      4:               0.5                1                     2
#>      5:               0.5                2                     5
#>     ---                                                         
#> 157937:               0.3                5                    20
#> 157938:               0.3                1                    10
#> 157939:                NA                0                     0
#> 157940:               0.7                1                     2
#> 157941:               0.8                3                     5
#>         field_goals_pct two_point_field_goals_made
#>                   <num>                      <int>
#>      1:             0.0                          0
#>      2:            33.3                          4
#>      3:             0.0                          0
#>      4:            50.0                          1
#>      5:            40.0                          1
#>     ---                                           
#> 157937:            25.0                          2
#> 157938:            10.0                          0
#> 157939:             0.0                          0
#> 157940:            50.0                          0
#> 157941:            60.0                          2
#>         two_point_field_goals_attempted two_point_field_goals_pct
#>                                   <int>                     <num>
#>      1:                               0                       0.0
#>      2:                              12                      33.3
#>      3:                               0                       0.0
#>      4:                               2                      50.0
#>      5:                               2                      50.0
#>     ---                                                          
#> 157937:                              11                      18.2
#> 157938:                               6                       0.0
#> 157939:                               0                       0.0
#> 157940:                               1                       0.0
#> 157941:                               2                     100.0
#>         three_point_field_goals_made three_point_field_goals_attempted
#>                                <int>                             <int>
#>      1:                            0                                 0
#>      2:                            0                                 0
#>      3:                            0                                 0
#>      4:                            0                                 0
#>      5:                            1                                 3
#>     ---                                                               
#> 157937:                            3                                 9
#> 157938:                            1                                 4
#> 157939:                            0                                 0
#> 157940:                            1                                 1
#> 157941:                            1                                 3
#>         three_point_field_goals_pct free_throws_made free_throws_attempted
#>                               <num>            <int>                 <int>
#>      1:                         0.0                0                     0
#>      2:                         0.0                0                     0
#>      3:                         0.0                0                     0
#>      4:                         0.0                0                     0
#>      5:                        33.3                0                     0
#>     ---                                                                   
#> 157937:                        33.3                2                     4
#> 157938:                        25.0                5                    10
#> 157939:                         0.0                0                     0
#> 157940:                       100.0                1                     2
#> 157941:                        33.3                3                     3
#>         free_throws_pct rebounds_offensive rebounds_defensive rebounds_total
#>                   <num>              <int>              <int>          <int>
#>      1:               0                  0                  1              1
#>      2:               0                  3                  2              5
#>      3:               0                  0                  0              0
#>      4:               0                  0                  1              1
#>      5:               0                  0                  3              3
#>     ---                                                                     
#> 157937:              50                  1                  4              5
#> 157938:              50                  1                  5              6
#> 157939:               0                  0                  0              0
#> 157940:              50                  1                  3              4
#> 157941:             100                  0                  4              4
#>         win_shares_offensive win_shares_defensive win_shares_total
#>                        <num>                <num>            <num>
#>      1:                   NA                   NA               NA
#>      2:                   NA                   NA               NA
#>      3:                   NA                   NA               NA
#>      4:                   NA                   NA               NA
#>      5:                   NA                   NA               NA
#>     ---                                                           
#> 157937:                 -0.2                    0             -0.2
#> 157938:                 -0.2                    0             -0.2
#> 157939:                   NA                    0               NA
#> 157940:                  0.0                    0              0.0
#> 157941:                  0.0                    0              0.0
#>         win_shares_total_per40
#>                          <num>
#>      1:                     NA
#>      2:                     NA
#>      3:                     NA
#>      4:                     NA
#>      5:                     NA
#>     ---                       
#> 157937:                 -0.235
#> 157938:                 -0.276
#> 157939:                     NA
#> 157940:                  0.000
#> 157941:                  0.000
# }