Load cbb play by play data

Grabs entire seasons of play-by-play data from the collegebasketballdata.com API for specified season(s)

Includes a roughly complete record of every tracked play in for the games in the specified seasons.

For many games with clean player substitution data, the identities of the players on the floor are included. Sometimes, substitution data is unclean, and this results in too many or too fewer players listed as being on the floor at a given time.

A primary player is identified for some plays as participants_*_1. For example, in an unassisted shot, the ID of the player taking the shot is participants_id_1. If two players are involved in a play, the initiating player is listed as participants_id_1 and the receiving player is listed as participants_id_2. If a player passes the ball to another player who makes a shot, the assisting player is listed as participants_id_1 and the shooting player is listed as participants_id_2.

Shot location data is provided where available. build vignette showing how to use shot location data

The win probability for the home team at the time of each play is provided as win_probability. These are predictions provided by the collegebasketballdata.com API, not built into the package.

Usage

load_plays(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of play-by-play data from all D1 college basketball games.

See also

Examples

# \donttest{
load_plays(2024)
#>          play_id on_floor_athlete_id_1 on_floor_athlete_id_2
#>            <int>                 <int>                 <int>
#>       1: 1160702                    47                  2938
#>       2: 1160703                    47                  2938
#>       3: 1160704                    47                  2938
#>       4: 1160705                    47                  2938
#>       5: 1160706                    47                  2938
#>      ---                                                    
#> 2101077: 3165694                   210                   369
#> 2101078: 3165695                   210                   369
#> 2101079: 3165696                   210                   369
#> 2101080: 3165697                  8146                 15262
#> 2101081: 3165698                  8146                 15262
#>          on_floor_athlete_id_3 on_floor_athlete_id_4 on_floor_athlete_id_5
#>                          <int>                 <int>                 <int>
#>       1:                 15214                    48                 15207
#>       2:                 15214                    48                 15207
#>       3:                 15214                    48                 15207
#>       4:                 15214                    48                 15207
#>       5:                 15214                    48                 15207
#>      ---                                                                  
#> 2101077:                 15235                 15258                  7792
#> 2101078:                 15235                 15258                  7792
#> 2101079:                 15235                 15258                  7792
#> 2101080:                  7797                  7794                  7802
#> 2101081:                  7797                  7794                  7802
#>          on_floor_athlete_id_6 on_floor_athlete_id_7 on_floor_athlete_id_8
#>                          <int>                 <int>                 <int>
#>       1:                    51                    63                 15206
#>       2:                    51                    63                 15206
#>       3:                    51                    63                 15206
#>       4:                    51                    63                 15206
#>       5:                    51                    63                 15206
#>      ---                                                                  
#> 2101077:                 15259                 15260                 15261
#> 2101078:                 15259                 15260                 15261
#> 2101079:                 15259                 15260                 15261
#> 2101080:                  1486                   373                  6789
#> 2101081:                  1486                   373                  6789
#>          on_floor_athlete_id_9 on_floor_athlete_id_10 on_floor_athlete_id_11
#>                          <int>                  <int>                  <int>
#>       1:                 15205                   2940                     NA
#>       2:                 15205                   2940                     NA
#>       3:                 15205                   2940                     NA
#>       4:                 15205                   2940                     NA
#>       5:                 15205                   2940                     NA
#>      ---                                                                    
#> 2101077:                   368                  15234                     NA
#> 2101078:                   368                  15234                     NA
#> 2101079:                   368                  15234                     NA
#> 2101080:                 15236                    210                     NA
#> 2101081:                 15236                    210                     NA
#>          on_floor_athlete_id_12 on_floor_athlete_id_13 on_floor_athlete_id_14
#>                           <int>                  <int>                  <int>
#>       1:                     NA                     NA                     NA
#>       2:                     NA                     NA                     NA
#>       3:                     NA                     NA                     NA
#>       4:                     NA                     NA                     NA
#>       5:                     NA                     NA                     NA
#>      ---                                                                     
#> 2101077:                     NA                     NA                     NA
#> 2101078:                     NA                     NA                     NA
#> 2101079:                     NA                     NA                     NA
#> 2101080:                     NA                     NA                     NA
#> 2101081:                     NA                     NA                     NA
#>          on_floor_athlete_name_1 on_floor_athlete_name_2
#>                           <char>                  <char>
#>       1:            KJ Adams Jr.           Po'Boigh King
#>       2:            KJ Adams Jr.           Po'Boigh King
#>       3:            KJ Adams Jr.           Po'Boigh King
#>       4:            KJ Adams Jr.           Po'Boigh King
#>       5:            KJ Adams Jr.           Po'Boigh King
#>      ---                                                
#> 2101077:            Mason Gillis          Fletcher Loyer
#> 2101078:            Mason Gillis          Fletcher Loyer
#> 2101079:            Mason Gillis          Fletcher Loyer
#> 2101080:     Apostolos Roumoglou           Andrew Hurley
#> 2101081:     Apostolos Roumoglou           Andrew Hurley
#>          on_floor_athlete_name_3 on_floor_athlete_name_4
#>                           <char>                  <char>
#>       1:      Kevin McCullar Jr.        Hunter Dickinson
#>       2:      Kevin McCullar Jr.        Hunter Dickinson
#>       3:      Kevin McCullar Jr.        Hunter Dickinson
#>       4:      Kevin McCullar Jr.        Hunter Dickinson
#>       5:      Kevin McCullar Jr.        Hunter Dickinson
#>      ---                                                
#> 2101077:             Lance Jones         Donovan Clingan
#> 2101078:             Lance Jones         Donovan Clingan
#> 2101079:             Lance Jones         Donovan Clingan
#> 2101080:          Jaylin Stewart               Solo Ball
#> 2101081:          Jaylin Stewart               Solo Ball
#>          on_floor_athlete_name_5 on_floor_athlete_name_6
#>                           <char>                  <char>
#>       1:      Fred Cleveland Jr.       Dajuan Harris Jr.
#>       2:      Fred Cleveland Jr.       Dajuan Harris Jr.
#>       3:      Fred Cleveland Jr.       Dajuan Harris Jr.
#>       4:      Fred Cleveland Jr.       Dajuan Harris Jr.
#>       5:      Fred Cleveland Jr.       Dajuan Harris Jr.
#>      ---                                                
#> 2101077:            Alex Karaban          Stephon Castle
#> 2101078:            Alex Karaban          Stephon Castle
#> 2101079:            Alex Karaban          Stephon Castle
#> 2101080:             Jayden Ross            Chase Martin
#> 2101081:             Jayden Ross            Chase Martin
#>          on_floor_athlete_name_7 on_floor_athlete_name_8
#>                           <char>                  <char>
#>       1:         Elmarko Jackson          Cobey Harraway
#>       2:         Elmarko Jackson          Cobey Harraway
#>       3:         Elmarko Jackson          Cobey Harraway
#>       4:         Elmarko Jackson          Cobey Harraway
#>       5:         Elmarko Jackson          Cobey Harraway
#>      ---                                                
#> 2101077:          Tristen Newton             Cam Spencer
#> 2101078:          Tristen Newton             Cam Spencer
#> 2101079:          Tristen Newton             Cam Spencer
#> 2101080:             Caleb Furst            Ethan Morton
#> 2101081:             Caleb Furst            Ethan Morton
#>          on_floor_athlete_name_9 on_floor_athlete_name_10
#>                           <char>                   <char>
#>       1:       Emmanuel Izunabor           Keishon Porter
#>       2:       Emmanuel Izunabor           Keishon Porter
#>       3:       Emmanuel Izunabor           Keishon Porter
#>       4:       Emmanuel Izunabor           Keishon Porter
#>       5:       Emmanuel Izunabor           Keishon Porter
#>      ---                                                 
#> 2101077:            Braden Smith                Zach Edey
#> 2101078:            Braden Smith                Zach Edey
#> 2101079:            Braden Smith                Zach Edey
#> 2101080:          Carson Barrett             Mason Gillis
#> 2101081:          Carson Barrett             Mason Gillis
#>          on_floor_athlete_name_11 on_floor_athlete_name_12
#>                            <char>                   <char>
#>       1:                     <NA>                     <NA>
#>       2:                     <NA>                     <NA>
#>       3:                     <NA>                     <NA>
#>       4:                     <NA>                     <NA>
#>       5:                     <NA>                     <NA>
#>      ---                                                  
#> 2101077:                     <NA>                     <NA>
#> 2101078:                     <NA>                     <NA>
#> 2101079:                     <NA>                     <NA>
#> 2101080:                     <NA>                     <NA>
#> 2101081:                     <NA>                     <NA>
#>          on_floor_athlete_name_13 on_floor_athlete_name_14 on_floor_team_name_1
#>                            <char>                   <char>               <char>
#>       1:                     <NA>                     <NA>               Kansas
#>       2:                     <NA>                     <NA>               Kansas
#>       3:                     <NA>                     <NA>               Kansas
#>       4:                     <NA>                     <NA>               Kansas
#>       5:                     <NA>                     <NA>               Kansas
#>      ---                                                                       
#> 2101077:                     <NA>                     <NA>               Purdue
#> 2101078:                     <NA>                     <NA>               Purdue
#> 2101079:                     <NA>                     <NA>               Purdue
#> 2101080:                     <NA>                     <NA>                UConn
#> 2101081:                     <NA>                     <NA>                UConn
#>            on_floor_team_name_2 on_floor_team_name_3 on_floor_team_name_4
#>                          <char>               <char>               <char>
#>       1: North Carolina Central               Kansas               Kansas
#>       2: North Carolina Central               Kansas               Kansas
#>       3: North Carolina Central               Kansas               Kansas
#>       4: North Carolina Central               Kansas               Kansas
#>       5: North Carolina Central               Kansas               Kansas
#>      ---                                                                 
#> 2101077:                 Purdue               Purdue                UConn
#> 2101078:                 Purdue               Purdue                UConn
#> 2101079:                 Purdue               Purdue                UConn
#> 2101080:                  UConn                UConn                UConn
#> 2101081:                  UConn                UConn                UConn
#>            on_floor_team_name_5 on_floor_team_name_6 on_floor_team_name_7
#>                          <char>               <char>               <char>
#>       1: North Carolina Central               Kansas               Kansas
#>       2: North Carolina Central               Kansas               Kansas
#>       3: North Carolina Central               Kansas               Kansas
#>       4: North Carolina Central               Kansas               Kansas
#>       5: North Carolina Central               Kansas               Kansas
#>      ---                                                                 
#> 2101077:                  UConn                UConn                UConn
#> 2101078:                  UConn                UConn                UConn
#> 2101079:                  UConn                UConn                UConn
#> 2101080:                  UConn               Purdue               Purdue
#> 2101081:                  UConn               Purdue               Purdue
#>            on_floor_team_name_8   on_floor_team_name_9  on_floor_team_name_10
#>                          <char>                 <char>                 <char>
#>       1: North Carolina Central North Carolina Central North Carolina Central
#>       2: North Carolina Central North Carolina Central North Carolina Central
#>       3: North Carolina Central North Carolina Central North Carolina Central
#>       4: North Carolina Central North Carolina Central North Carolina Central
#>       5: North Carolina Central North Carolina Central North Carolina Central
#>      ---                                                                     
#> 2101077:                  UConn                 Purdue                 Purdue
#> 2101078:                  UConn                 Purdue                 Purdue
#> 2101079:                  UConn                 Purdue                 Purdue
#> 2101080:                 Purdue                 Purdue                 Purdue
#> 2101081:                 Purdue                 Purdue                 Purdue
#>          on_floor_team_name_11 on_floor_team_name_12 on_floor_team_name_13
#>                         <char>                <char>                <char>
#>       1:                  <NA>                  <NA>                  <NA>
#>       2:                  <NA>                  <NA>                  <NA>
#>       3:                  <NA>                  <NA>                  <NA>
#>       4:                  <NA>                  <NA>                  <NA>
#>       5:                  <NA>                  <NA>                  <NA>
#>      ---                                                                  
#> 2101077:                  <NA>                  <NA>                  <NA>
#> 2101078:                  <NA>                  <NA>                  <NA>
#> 2101079:                  <NA>                  <NA>                  <NA>
#> 2101080:                  <NA>                  <NA>                  <NA>
#> 2101081:                  <NA>                  <NA>                  <NA>
#>          on_floor_team_name_14 shooter_athlete_id shooter_athlete_name
#>                         <char>              <int>               <char>
#>       1:                  <NA>              15214   Kevin McCullar Jr.
#>       2:                  <NA>                 NA                 <NA>
#>       3:                  <NA>                 48     Hunter Dickinson
#>       4:                  <NA>                 NA                 <NA>
#>       5:                  <NA>                 48     Hunter Dickinson
#>      ---                                                              
#> 2101077:                  <NA>              15258      Donovan Clingan
#> 2101078:                  <NA>                368         Braden Smith
#> 2101079:                  <NA>                 NA                 <NA>
#> 2101080:                  <NA>                 NA                 <NA>
#> 2101081:                  <NA>                 NA                 <NA>
#>          shot_made    shot_range shot_assisted assist_athlete_id
#>             <lgcl>        <char>        <lgcl>             <int>
#>       1:     FALSE three_pointer         FALSE                NA
#>       2:        NA          <NA>            NA                NA
#>       3:     FALSE           rim         FALSE                NA
#>       4:        NA          <NA>            NA                NA
#>       5:      TRUE           rim         FALSE                NA
#>      ---                                                        
#> 2101077:      TRUE           rim          TRUE              7792
#> 2101078:     FALSE three_pointer         FALSE                NA
#> 2101079:        NA          <NA>            NA                NA
#> 2101080:        NA          <NA>            NA                NA
#> 2101081:        NA          <NA>            NA                NA
#>          assist_athlete_name shot_x_coordinate shot_y_coordinate
#>                       <char>             <num>             <int>
#>       1:                <NA>             780.2               465
#>       2:                <NA>                NA                NA
#>       3:                <NA>             883.6               250
#>       4:                <NA>                NA                NA
#>       5:                <NA>             883.6               250
#>      ---                                                        
#> 2101077:        Alex Karaban                NA                NA
#> 2101078:                <NA>                NA                NA
#> 2101079:                <NA>                NA                NA
#> 2101080:                <NA>                NA                NA
#> 2101081:                <NA>                NA                NA
#>          primary_participant_athlete_id secondary_participant_athlete_id
#>                                   <int>                            <int>
#>       1:                          15214                               NA
#>       2:                             48                               NA
#>       3:                             48                               NA
#>       4:                             48                               NA
#>       5:                             48                               NA
#>      ---                                                                
#> 2101077:                           7792                            15258
#> 2101078:                            368                               NA
#> 2101079:                             NA                               NA
#> 2101080:                             NA                               NA
#> 2101081:                             NA                               NA
#>          participants_id_3 participants_id_4 primary_participant_athlete_name
#>                      <int>             <int>                           <char>
#>       1:                NA                NA               Kevin McCullar Jr.
#>       2:                NA                NA                 Hunter Dickinson
#>       3:                NA                NA                 Hunter Dickinson
#>       4:                NA                NA                 Hunter Dickinson
#>       5:                NA                NA                 Hunter Dickinson
#>      ---                                                                     
#> 2101077:                NA                NA                     Alex Karaban
#> 2101078:                NA                NA                     Braden Smith
#> 2101079:                NA                NA                             <NA>
#> 2101080:                NA                NA                             <NA>
#> 2101081:                NA                NA                             <NA>
#>          secondary_participant_athlete_name participants_name_3
#>                                      <char>              <char>
#>       1:                               <NA>                <NA>
#>       2:                               <NA>                <NA>
#>       3:                               <NA>                <NA>
#>       4:                               <NA>                <NA>
#>       5:                               <NA>                <NA>
#>      ---                                                       
#> 2101077:                    Donovan Clingan                <NA>
#> 2101078:                               <NA>                <NA>
#> 2101079:                               <NA>                <NA>
#> 2101080:                               <NA>                <NA>
#> 2101081:                               <NA>                <NA>
#>          participants_name_4 game_id game_source_id          game_start_date
#>                       <char>   <int>         <char>                   <char>
#>       1:                <NA>    5880      401575451 2023-11-06T20:00:00.000Z
#>       2:                <NA>    5880      401575451 2023-11-06T20:00:00.000Z
#>       3:                <NA>    5880      401575451 2023-11-06T20:00:00.000Z
#>       4:                <NA>    5880      401575451 2023-11-06T20:00:00.000Z
#>       5:                <NA>    5880      401575451 2023-11-06T20:00:00.000Z
#>      ---                                                                    
#> 2101077:                <NA>   12128      401638645 2024-04-08T21:20:00.000Z
#> 2101078:                <NA>   12128      401638645 2024-04-08T21:20:00.000Z
#> 2101079:                <NA>   12128      401638645 2024-04-08T21:20:00.000Z
#> 2101080:                <NA>   12128      401638645 2024-04-08T21:20:00.000Z
#> 2101081:                <NA>   12128      401638645 2024-04-08T21:20:00.000Z
#>          season season_type game_type tournament     play_source_id
#>           <int>      <char>    <char>     <char>             <char>
#>       1:   2024     regular       STD       <NA> 401575451101805401
#>       2:   2024     regular       STD       <NA> 401575451101806001
#>       3:   2024     regular       STD       <NA> 401575451101806002
#>       4:   2024     regular       STD       <NA> 401575451101806201
#>       5:   2024     regular       STD       <NA> 401575451101806202
#>      ---                                                           
#> 2101077:   2024  postseason    TRNMNT       NCAA 401638645102995401
#> 2101078:   2024  postseason    TRNMNT       NCAA 401638645102996201
#> 2101079:   2024  postseason    TRNMNT       NCAA 401638645102996202
#> 2101080:   2024  postseason    TRNMNT       NCAA 401638645102999301
#> 2101081:   2024  postseason    TRNMNT       NCAA 401638645102999901
#>                   play_type is_home_team team_id team_name
#>                      <char>       <lgcl>   <int>    <char>
#>       1:           JumpShot         TRUE     131    Kansas
#>       2:  Offensive Rebound         TRUE     131    Kansas
#>       3:          LayUpShot         TRUE     131    Kansas
#>       4:  Offensive Rebound         TRUE     131    Kansas
#>       5:          LayUpShot         TRUE     131    Kansas
#>      ---                                                  
#> 2101077:          LayUpShot         TRUE     314     UConn
#> 2101078:           JumpShot        FALSE     236    Purdue
#> 2101079:  Defensive Rebound         TRUE     314     UConn
#> 2101080: Lost Ball Turnover         TRUE     314     UConn
#> 2101081:           End Game           NA      NA      <NA>
#>          conference_short_name team_seed opponent_team_id
#>                         <char>     <int>            <int>
#>       1:                Big 12        NA              199
#>       2:                Big 12        NA              199
#>       3:                Big 12        NA              199
#>       4:                Big 12        NA              199
#>       5:                Big 12        NA              199
#>      ---                                                 
#> 2101077:              Big East         1              236
#> 2101078:               Big Ten         1              314
#> 2101079:              Big East         1              236
#> 2101080:              Big East         1              236
#> 2101081:                  <NA>        NA               NA
#>                   opponent_name opponent_conference_short_name opponent_seed
#>                          <char>                         <char>         <int>
#>       1: North Carolina Central                           MEAC            NA
#>       2: North Carolina Central                           MEAC            NA
#>       3: North Carolina Central                           MEAC            NA
#>       4: North Carolina Central                           MEAC            NA
#>       5: North Carolina Central                           MEAC            NA
#>      ---                                                                    
#> 2101077:                 Purdue                        Big Ten             1
#> 2101078:                  UConn                       Big East             1
#> 2101079:                 Purdue                        Big Ten             1
#> 2101080:                 Purdue                        Big Ten             1
#> 2101081:                   <NA>                           <NA>            NA
#>          home_score away_score home_win_probability period  clock
#>               <int>      <int>                <num>  <int> <char>
#>       1:          0          0                0.641      1  19:45
#>       2:          0          0                0.635      1  19:39
#>       3:          0          0                0.635      1  19:39
#>       4:          0          0                0.635      1  19:37
#>       5:          2          0                0.694      1  19:37
#>      ---                                                         
#> 2101077:         75         60                1.000      2   0:45
#> 2101078:         75         60                1.000      2   0:37
#> 2101079:         75         60                1.000      2   0:37
#> 2101080:         75         60                1.000      2   0:06
#> 2101081:         75         60                1.000      2   0:00
#>          seconds_remaining scoring_play shooting_play score_value
#>                      <int>       <lgcl>        <lgcl>       <int>
#>       1:              1185        FALSE          TRUE           3
#>       2:              1179        FALSE         FALSE           0
#>       3:              1179        FALSE          TRUE           2
#>       4:              1177        FALSE         FALSE           0
#>       5:              1177         TRUE          TRUE           2
#>      ---                                                         
#> 2101077:                45         TRUE          TRUE           2
#> 2101078:                37        FALSE          TRUE           3
#> 2101079:                37        FALSE         FALSE           0
#> 2101080:                 6        FALSE         FALSE           0
#> 2101081:                 0        FALSE         FALSE           0
#>                        wall_clock
#>                            <char>
#>       1: 2023-11-06T20:03:19.000Z
#>       2: 2023-11-06T20:03:19.000Z
#>       3: 2023-11-06T20:03:19.000Z
#>       4: 2023-11-06T20:03:19.000Z
#>       5: 2023-11-06T20:03:19.000Z
#>      ---                         
#> 2101077: 2024-04-08T23:28:07.000Z
#> 2101078: 2024-04-08T23:28:17.000Z
#> 2101079: 2024-04-08T23:28:17.000Z
#> 2101080: 2024-04-08T23:29:58.000Z
#> 2101081: 2024-04-08T23:29:58.000Z
#>                                                      play_text
#>                                                         <char>
#>       1:         Kevin McCullar Jr. missed Three Point Jumper.
#>       2:                   Hunter Dickinson Offensive Rebound.
#>       3:                        Hunter Dickinson missed Layup.
#>       4:                   Hunter Dickinson Offensive Rebound.
#>       5:                          Hunter Dickinson made Layup.
#>      ---                                                      
#> 2101077: Donovan Clingan made Layup. Assisted by Alex Karaban.
#> 2101078:               Braden Smith missed Three Point Jumper.
#> 2101079:                              UConn Defensive Rebound.
#> 2101080:                                      UConn  Turnover.
#> 2101081:                                           End of Game
# }