Load ranking data

Loads information about college basketball rankings, polls, etc.

At present, only the AP Top 25 and Coaches Polls are available from this function. Ranking, first place votes, and over points are available for each week of the season dating back to 1949.

Usage

load_rankings()

Value

a dataframe of information about college basketball rankings and polls

See also

Examples

# \donttest{
load_rankings()
#>        season season_type  week                poll_date poll_type team_id
#>         <int>      <char> <int>                   <char>    <char>   <int>
#>     1:   2025     regular     1 2024-10-14T00:00:00.000Z AP Top 25     131
#>     2:   2025     regular     1 2024-10-14T00:00:00.000Z AP Top 25       5
#>     3:   2025     regular     1 2024-10-14T00:00:00.000Z AP Top 25     314
#>     4:   2025     regular     1 2024-10-14T00:00:00.000Z AP Top 25     113
#>     5:   2025     regular     1 2024-10-14T00:00:00.000Z AP Top 25     125
#>    ---                                                                    
#> 55807:   1949     regular     7 1949-02-28T00:00:00.000Z AP Top 25     425
#> 55808:   1949     regular     7 1949-02-28T00:00:00.000Z AP Top 25     338
#> 55809:   1949     regular     7 1949-02-28T00:00:00.000Z AP Top 25     360
#> 55810:   1949     regular     7 1949-02-28T00:00:00.000Z AP Top 25     216
#> 55811:   1949     regular     7 1949-02-28T00:00:00.000Z AP Top 25      12
#>         team_name conference_short_name ranking points first_place_votes
#>            <char>                <char>   <int>  <int>             <int>
#>     1:     Kansas                Big 12       1   1449                30
#>     2:    Alabama                   SEC       2   1428                14
#>     3:      UConn              Big East       3   1345                11
#>     4:    Houston                Big 12       4   1343                 4
#>     5: Iowa State                Big 12       5   1177                 0
#>    ---                                                                  
#> 55807:    Hamline                  <NA>      16     NA                NA
#> 55808:  Villanova                Indep.      17     NA                NA
#> 55809:       Yale                   Ivy      18     NA                NA
#> 55810: Ohio State               Big Ten      19     NA                NA
#> 55811:   Arkansas                  <NA>      20     NA                NA
# }