Load cbb team rosters

Pull in rosters for college basketball teams for specified season(s)

Usage

load_rosters(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of rosters for college basketball teams in the specified season(s).

See also

Examples

# \donttest{
load_rosters()
#>        team_id team_source_id                team conference_short_name season
#>          <int>         <char>              <char>                <char>  <int>
#>     1:     161            113       Massachusetts                  A-10   2025
#>     2:     161            113       Massachusetts                  A-10   2025
#>     3:     161            113       Massachusetts                  A-10   2025
#>     4:     161            113       Massachusetts                  A-10   2025
#>     5:     161            113       Massachusetts                  A-10   2025
#>    ---                                                                        
#> 15233:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#> 15234:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#> 15235:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#> 15236:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#> 15237:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#>        athlete_id athlete_source_id           athlete_name athlete_first_name
#>             <int>            <char>                 <char>             <char>
#>     1:       2882           5176382       Malek Abdelgowad              Malek
#>     2:       2883           5105982          Daniel Rivera             Daniel
#>     3:       2884           5105551 Daniel Hankins-Sanford             Daniel
#>     4:       2885           5174611           Jaylen Curry             Jaylen
#>     5:       2886           4683733        Rahsool Diggins            Rahsool
#>    ---                                                                       
#> 15233:       9314           4907749     Tarek Abdel-Kireem              Tarek
#> 15234:       9315           4907748           Abeal Ayalew              Abeal
#> 15235:       9316           4907745            Matt Kohler               Matt
#> 15236:       9317           4907744            Devyn Jones              Devyn
#> 15237:       9318           4907742          Jayden Priddy             Jayden
#>        athlete_last_name athlete_jersey_number athlete_position
#>                   <char>                <char>           <char>
#>     1:        Abdelgowad                    25          Forward
#>     2:            Rivera                     5          Forward
#>     3:   Hankins-Sanford                     1          Forward
#>     4:             Curry                     0            Guard
#>     5:           Diggins                     7            Guard
#>    ---                                                         
#> 15233:      Abdel-Kireem                  <NA>          Athlete
#> 15234:            Ayalew                  <NA>          Athlete
#> 15235:            Kohler                  <NA>          Athlete
#> 15236:             Jones                  <NA>          Athlete
#> 15237:            Priddy                  <NA>          Athlete
#>        athlete_height_inches athlete_weight_pounds athlete_hometown_city
#>                        <int>                 <int>                <char>
#>     1:                    82                   220                 Cairo
#>     2:                    78                   215              San Juan
#>     3:                    80                   236             Charlotte
#>     4:                    72                   168             Charlotte
#>     5:                    74                   190          Philadelphia
#>    ---                                                                  
#> 15233:                    NA                    NA                  <NA>
#> 15234:                    NA                    NA                  <NA>
#> 15235:                    NA                    NA                  <NA>
#> 15236:                    NA                    NA                  <NA>
#> 15237:                    NA                    NA                  <NA>
#>        athlete_hometown_state athlete_hometown_country
#>                        <char>                   <lgcl>
#>     1:                  Egypt                       NA
#>     2:            Puerto Rico                       NA
#>     3:                     NC                       NA
#>     4:                     NC                       NA
#>     5:                     PA                       NA
#>    ---                                                
#> 15233:                   <NA>                       NA
#> 15234:                   <NA>                       NA
#> 15235:                   <NA>                       NA
#> 15236:                   <NA>                       NA
#> 15237:                   <NA>                       NA
#>        athlete_hometown_latitude athlete_hometown_longitude
#>                            <num>                      <num>
#>     1:                        NA                         NA
#>     2:                        NA                         NA
#>     3:                  35.22721                  -80.84308
#>     4:                  35.22721                  -80.84308
#>     5:                  39.95272                  -75.16353
#>    ---                                                     
#> 15233:                        NA                         NA
#> 15234:                        NA                         NA
#> 15235:                        NA                         NA
#> 15236:                        NA                         NA
#> 15237:                        NA                         NA
#>        athlete_hometown_county_fips athlete_date_of_birth athlete_start_season
#>                              <char>                <char>                <int>
#>     1:                         <NA>                  <NA>                 2025
#>     2:                         <NA>                  <NA>                 2025
#>     3:                        37119                  <NA>                 2024
#>     4:                        37119                  <NA>                 2024
#>     5:                        42101                  <NA>                 2023
#>    ---                                                                        
#> 15233:                         <NA>                  <NA>                 2023
#> 15234:                         <NA>                  <NA>                 2022
#> 15235:                         <NA>                  <NA>                 2022
#> 15236:                         <NA>                  <NA>                 2022
#> 15237:                         <NA>                  <NA>                 2023
#>        athlete_end_season
#>                     <int>
#>     1:               2025
#>     2:               2025
#>     3:               2025
#>     4:               2025
#>     5:               2025
#>    ---                   
#> 15233:               2025
#> 15234:               2025
#> 15235:               2025
#> 15236:               2025
#> 15237:               2025

load_rosters(2024)
#>        team_id team_source_id                team conference_short_name season
#>          <int>         <char>              <char>                <char>  <int>
#>     1:     161            113       Massachusetts                  A-10   2024
#>     2:     161            113       Massachusetts                  A-10   2024
#>     3:     161            113       Massachusetts                  A-10   2024
#>     4:     161            113       Massachusetts                  A-10   2024
#>     5:     161            113       Massachusetts                  A-10   2024
#>    ---                                                                        
#> 14126:     616           2985 Concordia-Ann Arbor                  <NA>   2024
#> 14127:     616           2985 Concordia-Ann Arbor                  <NA>   2024
#> 14128:     616           2985 Concordia-Ann Arbor                  <NA>   2024
#> 14129:     616           2985 Concordia-Ann Arbor                  <NA>   2024
#> 14130:     616           2985 Concordia-Ann Arbor                  <NA>   2024
#>        athlete_id athlete_source_id           athlete_name athlete_first_name
#>             <int>            <char>                 <char>             <char>
#>     1:       2631           5174612       Robert Davis Jr.             Robert
#>     2:       2884           5105551 Daniel Hankins-Sanford             Daniel
#>     3:       2885           5174611           Jaylen Curry             Jaylen
#>     4:       2886           4683733        Rahsool Diggins            Rahsool
#>     5:       2889           5174615         Tarique Foster            Tarique
#>    ---                                                                       
#> 14126:       9314           4907749     Tarek Abdel-Kireem              Tarek
#> 14127:       9315           4907748           Abeal Ayalew              Abeal
#> 14128:       9316           4907745            Matt Kohler               Matt
#> 14129:       9317           4907744            Devyn Jones              Devyn
#> 14130:       9318           4907742          Jayden Priddy             Jayden
#>        athlete_last_name athlete_jersey_number athlete_position
#>                   <char>                <char>           <char>
#>     1:         Davis Jr.                     4            Guard
#>     2:   Hankins-Sanford                     1          Forward
#>     3:             Curry                     0            Guard
#>     4:           Diggins                     7            Guard
#>     5:            Foster                    12          Forward
#>    ---                                                         
#> 14126:      Abdel-Kireem                  <NA>          Athlete
#> 14127:            Ayalew                  <NA>          Athlete
#> 14128:            Kohler                  <NA>          Athlete
#> 14129:             Jones                  <NA>          Athlete
#> 14130:            Priddy                  <NA>          Athlete
#>        athlete_height_inches athlete_weight_pounds athlete_hometown_city
#>                        <int>                 <int>                <char>
#>     1:                    78                   185               Detroit
#>     2:                    80                   236             Charlotte
#>     3:                    72                   168             Charlotte
#>     4:                    74                   190          Philadelphia
#>     5:                    80                   185                 Bronx
#>    ---                                                                  
#> 14126:                    NA                    NA                  <NA>
#> 14127:                    NA                    NA                  <NA>
#> 14128:                    NA                    NA                  <NA>
#> 14129:                    NA                    NA                  <NA>
#> 14130:                    NA                    NA                  <NA>
#>        athlete_hometown_state athlete_hometown_country
#>                        <char>                   <lgcl>
#>     1:                     MI                       NA
#>     2:                     NC                       NA
#>     3:                     NC                       NA
#>     4:                     PA                       NA
#>     5:                     NY                       NA
#>    ---                                                
#> 14126:                   <NA>                       NA
#> 14127:                   <NA>                       NA
#> 14128:                   <NA>                       NA
#> 14129:                   <NA>                       NA
#> 14130:                   <NA>                       NA
#>        athlete_hometown_latitude athlete_hometown_longitude
#>                            <num>                      <num>
#>     1:                  42.33155                  -83.04664
#>     2:                  35.22721                  -80.84308
#>     3:                  35.22721                  -80.84308
#>     4:                  39.95272                  -75.16353
#>     5:                  40.84665                  -73.87859
#>    ---                                                     
#> 14126:                        NA                         NA
#> 14127:                        NA                         NA
#> 14128:                        NA                         NA
#> 14129:                        NA                         NA
#> 14130:                        NA                         NA
#>        athlete_hometown_county_fips athlete_date_of_birth athlete_start_season
#>                              <char>                <char>                <int>
#>     1:                        26163                  <NA>                 2024
#>     2:                        37119                  <NA>                 2024
#>     3:                        37119                  <NA>                 2024
#>     4:                        42101                  <NA>                 2023
#>     5:                        36005                  <NA>                 2024
#>    ---                                                                        
#> 14126:                         <NA>                  <NA>                 2023
#> 14127:                         <NA>                  <NA>                 2022
#> 14128:                         <NA>                  <NA>                 2022
#> 14129:                         <NA>                  <NA>                 2022
#> 14130:                         <NA>                  <NA>                 2023
#>        athlete_end_season
#>                     <int>
#>     1:               2024
#>     2:               2025
#>     3:               2025
#>     4:               2025
#>     5:               2025
#>    ---                   
#> 14126:               2025
#> 14127:               2025
#> 14128:               2025
#> 14129:               2025
#> 14130:               2025

load_rosters(TRUE)
#>         team_id team_source_id                team conference_short_name season
#>           <int>         <char>              <char>                <char>  <int>
#>      1:     161            113       Massachusetts                  A-10   2003
#>      2:     161            113       Massachusetts                  A-10   2003
#>      3:     161            113       Massachusetts                  A-10   2003
#>      4:     161            113       Massachusetts                  A-10   2003
#>      5:     161            113       Massachusetts                  A-10   2003
#>     ---                                                                        
#> 217184:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#> 217185:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#> 217186:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#> 217187:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#> 217188:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#>         athlete_id athlete_source_id       athlete_name athlete_first_name
#>              <int>            <char>             <char>             <char>
#>      1:      88819             11811      Jeff Viggiano               Jeff
#>      2:      91926              5264   Anthony Anderson            Anthony
#>      3:      91929              5262     Brennan Martin            Brennan
#>      4:      92301             11809           Gabe Lee               Gabe
#>      5:      92303              5265     Paco Kotaridis               Paco
#>     ---                                                                   
#> 217184:       9314           4907749 Tarek Abdel-Kireem              Tarek
#> 217185:       9315           4907748       Abeal Ayalew              Abeal
#> 217186:       9316           4907745        Matt Kohler               Matt
#> 217187:       9317           4907744        Devyn Jones              Devyn
#> 217188:       9318           4907742      Jayden Priddy             Jayden
#>         athlete_last_name athlete_jersey_number athlete_position
#>                    <char>                <char>           <char>
#>      1:          Viggiano                   22           Forward
#>      2:          Anderson                   12             Guard
#>      3:            Martin                  <NA>          Forward
#>      4:               Lee                   2             Center
#>      5:         Kotaridis                   14             Guard
#>     ---                                                         
#> 217184:      Abdel-Kireem                  <NA>          Athlete
#> 217185:            Ayalew                  <NA>          Athlete
#> 217186:            Kohler                  <NA>          Athlete
#> 217187:             Jones                  <NA>          Athlete
#> 217188:            Priddy                  <NA>          Athlete
#>         athlete_height_inches athlete_weight_pounds  athlete_hometown_city
#>                         <int>                 <int>                 <char>
#>      1:                    78                   215               Suffield
#>      2:                    71                   170                   Lynn
#>      3:                    78                   190 Rancho Santa Margarita
#>      4:                    81                   210                Camilla
#>      5:                    73                   190                 Athens
#>     ---                                                                   
#> 217184:                    NA                    NA                   <NA>
#> 217185:                    NA                    NA                   <NA>
#> 217186:                    NA                    NA                   <NA>
#> 217187:                    NA                    NA                   <NA>
#> 217188:                    NA                    NA                   <NA>
#>         athlete_hometown_state athlete_hometown_country
#>                         <char>                   <lgcl>
#>      1:                     CT                       NA
#>      2:                     MA                       NA
#>      3:                     CA                       NA
#>      4:                     GA                       NA
#>      5:                 Greece                       NA
#>     ---                                                
#> 217184:                   <NA>                       NA
#> 217185:                   <NA>                       NA
#> 217186:                   <NA>                       NA
#> 217187:                   <NA>                       NA
#> 217188:                   <NA>                       NA
#>         athlete_hometown_latitude athlete_hometown_longitude
#>                             <num>                      <num>
#>      1:                  41.98169                  -72.65066
#>      2:                  42.46676                  -70.94949
#>      3:                  33.64088                 -117.60293
#>      4:                  31.23129                  -84.21037
#>      5:                        NA                         NA
#>     ---                                                     
#> 217184:                        NA                         NA
#> 217185:                        NA                         NA
#> 217186:                        NA                         NA
#> 217187:                        NA                         NA
#> 217188:                        NA                         NA
#>         athlete_hometown_county_fips athlete_date_of_birth athlete_start_season
#>                               <char>                <char>                <int>
#>      1:                        09003            2003-09-17                 2003
#>      2:                        25009                  <NA>                 2003
#>      3:                        06059                  <NA>                 2003
#>      4:                        13205            2003-09-17                 2003
#>      5:                         <NA>                  <NA>                 2003
#>     ---                                                                        
#> 217184:                         <NA>                  <NA>                 2023
#> 217185:                         <NA>                  <NA>                 2022
#> 217186:                         <NA>                  <NA>                 2022
#> 217187:                         <NA>                  <NA>                 2022
#> 217188:                         <NA>                  <NA>                 2023
#>         athlete_end_season
#>                      <int>
#>      1:               2006
#>      2:               2005
#>      3:               2005
#>      4:               2004
#>      5:               2004
#>     ---                   
#> 217184:               2025
#> 217185:               2025
#> 217186:               2025
#> 217187:               2025
#> 217188:               2025
# }