Load cbb teams stats

Pull in statistics aggregated by team for for specified season(s)

Usage

load_team_stats(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of team statistics for college basketball teams.

See also

Examples

# \donttest{
try({ # avoid cran errors
load_team_stats()
})
#>      season season_label team_id            team_name conference_short_name
#>       <int>       <char>   <int>               <char>                <char>
#>   1:   2025     20242025       1    Abilene Christian                   WAC
#>   2:   2025     20242025       2            Air Force         Mountain West
#>   3:   2025     20242025       3                Akron                   MAC
#>   4:   2025     20242025       4          Alabama A&M                  SWAC
#>   5:   2025     20242025       5              Alabama                   SEC
#>  ---                                                                       
#> 696:   2025     20242025     826         Medgar Evers                  <NA>
#> 697:   2025     20242025     827              Bowdoin                  <NA>
#> 698:   2025     20242025     828          Bowie State                  <NA>
#> 699:   2025     20242025     829            Morehouse                  <NA>
#> 700:   2025     20242025     830 Washington Adventist                  <NA>
#>      games  wins losses minutes  pace assists blocks steals possessions
#>      <int> <int>  <int>   <int> <num>   <int>  <int>  <int>       <int>
#>   1:    32    16     16    1310  66.4     421     94    331        2178
#>   2:    32     4     28    1310  61.1     428     90    181        2004
#>   3:    35    28      7    1415  69.1     617    130    267        2441
#>   4:    32    10     22    1320  61.5     432    131    266        2037
#>   5:    37    28      9    1510  74.5     634    167    228        2812
#>  ---                                                                   
#> 696:     1     0      1      40  76.5       7      3      6          76
#> 697:     1     0      1      40  62.0      15      6      6          62
#> 698:     1     1      0      40  67.0      16      2      6          67
#> 699:     1     1      0      40  76.5      13      1      6          77
#> 700:     1     0      1      40  83.5       8      1      8          84
#>      true_shooting offensive_rating field_goals_made field_goals_attempted
#>              <num>            <num>            <int>                 <int>
#>   1:          52.9            103.8              819                  1845
#>   2:          53.1             98.9              680                  1616
#>   3:          58.4            120.5             1060                  2259
#>   4:          51.0            114.7              795                  1966
#>   5:          60.3            119.3             1147                  2371
#>  ---                                                                      
#> 696:          33.8             67.1               20                    68
#> 697:          50.5            104.8               25                    60
#> 698:          60.2            113.4               30                    60
#> 699:          54.0            102.6               28                    67
#> 700:          30.3             50.0               12                    60
#>      field_goals_pct two_point_field_goals_made two_point_field_goals_attempted
#>                <num>                      <int>                           <int>
#>   1:           0.444                        673                            1362
#>   2:           0.421                        421                             833
#>   3:           0.469                        684                            1222
#>   4:           0.404                        538                            1159
#>   5:           0.484                        757                            1267
#>  ---                                                                           
#> 696:           0.294                         18                              54
#> 697:           0.417                         14                              28
#> 698:           0.500                         19                              36
#> 699:           0.418                         13                              33
#> 700:           0.200                          6                              37
#>      two_point_field_goals_pct three_point_field_goals_made
#>                          <num>                        <int>
#>   1:                     0.494                          146
#>   2:                     0.505                          259
#>   3:                     0.560                          376
#>   4:                     0.464                          257
#>   5:                     0.597                          390
#>  ---                                                       
#> 696:                     0.333                            2
#> 697:                     0.500                           11
#> 698:                     0.528                           11
#> 699:                     0.394                           15
#> 700:                     0.162                            6
#>      three_point_field_goals_attempted three_point_field_goals_pct
#>                                  <int>                       <num>
#>   1:                               483                       0.302
#>   2:                               783                       0.331
#>   3:                              1037                       0.363
#>   4:                               807                       0.318
#>   5:                              1104                       0.353
#>  ---                                                              
#> 696:                                14                       0.143
#> 697:                                32                       0.344
#> 698:                                24                       0.458
#> 699:                                34                       0.441
#> 700:                                23                       0.261
#>      free_throws_made free_throws_attempted free_throws_pct rebounds_offensive
#>                 <int>                 <int>           <num>              <int>
#>   1:              476                   666           0.715                356
#>   2:              363                   572           0.635                236
#>   3:              445                   590           0.754                414
#>   4:              490                   738           0.664                447
#>   5:              671                   932           0.720                459
#>  ---                                                                          
#> 696:                9                    17           0.529                 12
#> 697:                4                    10           0.400                  8
#> 698:                5                     7           0.714                  4
#> 699:                8                    14           0.571                  9
#> 700:               12                    21           0.571                  8
#>      rebounds_defensive rebounds_total turnovers team_turnovers fouls
#>                   <int>          <int>     <int>          <int> <int>
#>   1:                727           1083       476             14   671
#>   2:                720            956       432             43   571
#>   3:                934           1348       414             13   623
#>   4:                725           1172       513             27   687
#>   5:               1108           1567       469             23   702
#>  ---                                                                 
#> 696:                 15             27        12              0    12
#> 697:                 23             31         6              0    15
#> 698:                 22             26         9              1    24
#> 699:                 30             39        12              0    18
#> 700:                 23             31        21              1    18
#>      technical_fouls flagrant_fouls points points_in_paint points_off_turnovers
#>                <int>          <int>  <int>           <int>                <int>
#>   1:               9              0   2260            1164                  581
#>   2:               1              0   1982             768                  336
#>   3:               8              0   2941            1208                  555
#>   4:               8              0   2337             970                  491
#>   5:               1              0   3355            1432                  453
#>  ---                                                                           
#> 696:               0              0     51              30                    8
#> 697:               0              0     65              22                    9
#> 698:               1              0     76              32                   21
#> 699:               0              0     79              22                   13
#> 700:               0              0     42               6                   12
#>      points_fast_break effective_field_goal_pct assists_turnover_ratio
#>                  <int>                    <num>                  <num>
#>   1:               359                    0.483                    0.2
#>   2:               174                    0.501                    0.2
#>   3:               462                    0.552                    0.2
#>   4:               428                    0.470                    0.3
#>   5:               500                    0.566                    0.2
#>  ---                                                                  
#> 696:                14                    0.309                    0.2
#> 697:                 2                    0.508                    0.1
#> 698:                14                    0.592                    0.1
#> 699:                17                    0.530                    0.2
#> 700:                 3                    0.250                    0.3
#>      offensive_rebound_pct free_throw_rate opponent_assists opponent_blocks
#>                      <num>           <num>            <int>           <int>
#>   1:                 0.329           0.361              383             143
#>   2:                 0.247           0.354              410             103
#>   3:                 0.307           0.261              418              88
#>   4:                 0.381           0.375              484              97
#>   5:                 0.293           0.393              480             153
#>  ---                                                                       
#> 696:                 0.444           0.250               25              10
#> 697:                 0.258           0.167               13               1
#> 698:                 0.154           0.117               15               1
#> 699:                 0.231           0.209               16               2
#> 700:                 0.258           0.350               24               6
#>      opponent_steals opponent_possessions opponent_true_shooting
#>                <int>                <int>                  <num>
#>   1:             259                 2172                   55.2
#>   2:             217                 1999                   58.0
#>   3:             246                 2445                   53.3
#>   4:             289                 2025                   57.5
#>   5:             288                 2813                   52.0
#>  ---                                                            
#> 696:               8                   77                   59.1
#> 697:               5                   62                   58.0
#> 698:               2                   67                   59.0
#> 699:               6                   76                   52.0
#> 700:              15                   83                   69.2
#>      opponent_offensive_rating opponent_field_goals_made
#>                          <num>                     <int>
#>   1:                     103.5                       747
#>   2:                     117.9                       845
#>   3:                     107.4                       933
#>   4:                     124.5                       834
#>   5:                     106.9                      1074
#>  ---                                                    
#> 696:                     139.0                        41
#> 697:                     114.5                        25
#> 698:                     109.0                        23
#> 699:                     100.0                        25
#> 700:                     126.5                        38
#>      opponent_field_goals_attempted opponent_field_goals_pct
#>                               <int>                    <num>
#>   1:                           1675                    0.446
#>   2:                           1758                    0.481
#>   3:                           2157                    0.433
#>   4:                           1816                    0.459
#>   5:                           2523                    0.426
#>  ---                                                        
#> 696:                             83                    0.494
#> 697:                             52                    0.481
#> 698:                             50                    0.460
#> 699:                             63                    0.397
#> 700:                             64                    0.594
#>      opponent_two_point_field_goals_made
#>                                    <int>
#>   1:                                 566
#>   2:                                 614
#>   3:                                 670
#>   4:                                 574
#>   5:                                 809
#>  ---                                    
#> 696:                                  29
#> 697:                                  17
#> 698:                                  13
#> 699:                                  16
#> 700:                                  27
#>      opponent_two_point_field_goals_attempted
#>                                         <int>
#>   1:                                     1098
#>   2:                                     1121
#>   3:                                     1333
#>   4:                                     1141
#>   5:                                     1650
#>  ---                                         
#> 696:                                       50
#> 697:                                       31
#> 698:                                       27
#> 699:                                       28
#> 700:                                       40
#>      opponent_two_point_field_goals_pct opponent_three_point_field_goals_made
#>                                   <num>                                 <int>
#>   1:                              0.515                                   181
#>   2:                              0.548                                   231
#>   3:                              0.503                                   263
#>   4:                              0.503                                   260
#>   5:                              0.490                                   265
#>  ---                                                                         
#> 696:                              0.580                                    12
#> 697:                              0.548                                     8
#> 698:                              0.481                                    10
#> 699:                              0.571                                     9
#> 700:                              0.675                                    11
#>      opponent_three_point_field_goals_attempted
#>                                           <int>
#>   1:                                        577
#>   2:                                        637
#>   3:                                        824
#>   4:                                        675
#>   5:                                        873
#>  ---                                           
#> 696:                                         33
#> 697:                                         21
#> 698:                                         23
#> 699:                                         35
#> 700:                                         24
#>      opponent_three_point_field_goals_pct opponent_free_throws_made
#>                                     <num>                     <int>
#>   1:                                0.314                       574
#>   2:                                0.363                       436
#>   3:                                0.319                       496
#>   4:                                0.385                       591
#>   5:                                0.304                       594
#>  ---                                                               
#> 696:                                0.364                        13
#> 697:                                0.381                        13
#> 698:                                0.435                        17
#> 699:                                0.257                        17
#> 700:                                0.458                        18
#>      opponent_free_throws_attempted opponent_free_throws_pct
#>                               <int>                    <num>
#>   1:                            825                    0.696
#>   2:                            623                    0.700
#>   3:                            692                    0.717
#>   4:                            853                    0.693
#>   5:                            837                    0.710
#>  ---                                                        
#> 696:                             17                    0.765
#> 697:                             21                    0.619
#> 698:                             27                    0.630
#> 699:                             23                    0.739
#> 700:                             27                    0.667
#>      opponent_rebounds_offensive opponent_rebounds_defensive
#>                            <int>                       <int>
#>   1:                         312                         762
#>   2:                         294                         802
#>   3:                         375                         850
#>   4:                         370                         828
#>   5:                         462                         902
#>  ---                                                        
#> 696:                          28                          41
#> 697:                           9                          30
#> 698:                          12                          26
#> 699:                          10                          33
#> 700:                           9                          44
#>      opponent_rebounds_total opponent_turnovers opponent_team_turnovers
#>                        <int>              <int>                   <int>
#>   1:                    1074                547                      28
#>   2:                    1096                327                      17
#>   3:                    1225                444                      16
#>   4:                    1198                514                      40
#>   5:                    1364                382                      12
#>  ---                                                                   
#> 696:                      69                 13                       0
#> 697:                      39                  8                       0
#> 698:                      38                 16                       1
#> 699:                      43                 12                       0
#> 700:                      53                 15                       0
#>      opponent_fouls opponent_technical_fouls opponent_flagrant_fouls
#>               <int>                    <int>                   <int>
#>   1:            570                        5                       0
#>   2:            558                        3                       0
#>   3:            578                        3                       0
#>   4:            599                        6                       0
#>   5:            759                        6                       0
#>  ---                                                                
#> 696:             18                        0                       0
#> 697:              9                        0                       0
#> 698:             14                        1                       0
#> 699:             15                        0                       0
#> 700:             20                        0                       0
#>      opponent_points opponent_points_in_paint opponent_points_off_turnovers
#>                <int>                    <int>                         <int>
#>   1:            2249                     1032                           459
#>   2:            2357                     1108                           535
#>   3:            2625                     1140                           419
#>   4:            2521                     1040                           555
#>   5:            3007                     1336                           459
#>  ---                                                                       
#> 696:             107                       56                            11
#> 697:              71                       26                             9
#> 698:              73                       22                            12
#> 699:              76                       28                            13
#> 700:             105                       46                            30
#>      opponent_points_fast_break opponent_effective_field_goal_pct
#>                           <int>                             <num>
#>   1:                        306                             0.500
#>   2:                        302                             0.546
#>   3:                        362                             0.494
#>   4:                        441                             0.531
#>   5:                        346                             0.478
#>  ---                                                             
#> 696:                         14                             0.566
#> 697:                          6                             0.558
#> 698:                         10                             0.560
#> 699:                         21                             0.468
#> 700:                         41                             0.680
#>      opponent_assists_turnover_ratio opponent_offensive_rebound_pct
#>                                <num>                          <num>
#>   1:                             0.3                          0.291
#>   2:                             0.2                          0.268
#>   3:                             0.2                          0.306
#>   4:                             0.3                          0.309
#>   5:                             0.1                          0.339
#>  ---                                                               
#> 696:                             0.2                          0.406
#> 697:                             0.1                          0.231
#> 698:                             0.2                          0.316
#> 699:                             0.2                          0.233
#> 700:                             0.2                          0.170
#>      opponent_free_throw_rate
#>                         <num>
#>   1:                    0.493
#>   2:                    0.354
#>   3:                    0.321
#>   4:                    0.470
#>   5:                    0.332
#>  ---                         
#> 696:                    0.205
#> 697:                    0.404
#> 698:                    0.540
#> 699:                    0.365
#> 700:                    0.422
# }