Load cbb teams

Pull in information on all teams that played in a given season.

Usage

load_teams(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of all college basketball teams tracked by collegebasketballdata.com for a given season(s).

See also

Examples

# \donttest{
try({ # avoid cran errors
load_teams()
})
#>      team_id team_source_id          team_name   team_mascot team_abbreviation
#>        <int>         <char>             <char>        <char>            <char>
#>   1:     264           2561              Siena        Saints               SIE
#>   2:      87             57            Florida        Gators               FLA
#>   3:     124           2294               Iowa      Hawkeyes              IOWA
#>   4:     148            309          Louisiana Ragin' Cajuns                UL
#>   5:     236           2509             Purdue  Boilermakers               PUR
#>  ---                                                                          
#> 360:     147           2344           Longwood       Lancers              LONG
#> 361:     346           2698       West Georgia        Wolves               WGA
#> 362:     165           2385         Mercyhurst        Lakers              MERC
#> 363:     270           2571 South Dakota State   Jackrabbits              SDST
#> 364:     137           2325           La Salle     Explorers               LAS
#>                   team_display_name team_short_name team_primary_color
#>                              <char>          <char>             <char>
#>   1:                   Siena Saints           Siena             037961
#>   2:                 Florida Gators         Florida             0021a5
#>   3:                  Iowa Hawkeyes            Iowa             000000
#>   4:        Louisiana Ragin' Cajuns       Louisiana             ce181e
#>   5:            Purdue Boilermakers          Purdue             000000
#>  ---                                                                  
#> 360:               Longwood Lancers        Longwood             003273
#> 361:            West Georgia Wolves    West Georgia             0033a1
#> 362:              Mercyhurst Lakers      Mercyhurst             000000
#> 363: South Dakota State Jackrabbits     S Dakota St             0033a0
#> 364:             La Salle Explorers        La Salle             003356
#>      team_secondary_color current_venue_id                  venue_name
#>                    <char>            <int>                      <char>
#>   1:               eea60f               71                   MVP Arena
#>   2:               fa4616              240 Stephen C. O'Connell Center
#>   3:               fcd116              144        Carver-Hawkeye Arena
#>   4:               ffffff              127                   Cajundome
#>   5:               cfb991               11                Mackey Arena
#>  ---                                                                  
#> 360:               9ea2a3              105     Joan Perry Brock Center
#> 361:               db1a21              375                The Coliseum
#> 362:                 <NA>              300  Mercyhurst Athletic Center
#> 363:               ffd100              287    First Bank & Trust Arena
#> 364:               ffce00               94           John Glaser Arena
#>          venue_city venue_state conference_id conference_short_name season
#>              <char>      <char>         <int>                <char>  <int>
#>   1:         Albany          NY            16                  MAAC   2025
#>   2:    Gainesville          FL            24                   SEC   2025
#>   3:      Iowa City          IA            10               Big Ten   2025
#>   4:      Lafayette          LA            29              Sun Belt   2025
#>   5: West Lafayette          IN            10               Big Ten   2025
#>  ---                                                                      
#> 360:      Farmville          VA             9             Big South   2025
#> 361:     Carrollton          GA             3                  ASUN   2025
#> 362:           Erie          PA            21                   NEC   2025
#> 363:      Brookings          SD            28                Summit   2025
#> 364:   Philadelphia          PA             1                  A-10   2025
# }