Load college basketball games for specified season(s)

Pull in all college basketball games tracked by collegebasketballdata.com

Returns data for all NCAA D1 men's basketball games for the specified seasons. Includes contextual information, such as where the game took place and who it is between, as well as high level box score data, such as the scores of each team.

This will return both games in the past and scheduled games. If a game is scheduled but its opponents have not yet been determined (i.e., the national championship game of the upcoming season), it will not appear in this dataset.

Usage

load_games(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of all games in the specified season(s).

See also

Examples

# \donttest{
load_games()
#>       game_id game_source_id season_label season season_type tournament
#>         <int>         <char>       <char>  <int>      <char>     <char>
#>    1:      24      401727499     20242025   2025     regular       <NA>
#>    2:      27      401715872     20242025   2025     regular       <NA>
#>    3:      25      401726400     20242025   2025     regular       <NA>
#>    4:      26      401725796     20242025   2025     regular       <NA>
#>    5:      28      401726395     20242025   2025     regular       <NA>
#>   ---                                                                  
#> 6296:  209747      401758633     20242025   2025  postseason       <NA>
#> 6297:  209731      401746081     20242025   2025  postseason       NCAA
#> 6298:  209732      401746080     20242025   2025  postseason       NCAA
#> 6299:  209749      401758636     20242025   2025  postseason       <NA>
#> 6300:  209751      401746082     20242025   2025  postseason       NCAA
#>                     start_date start_time_tbd neutral_site conference_game
#>                         <char>         <lgcl>       <lgcl>          <lgcl>
#>    1: 2024-11-04T11:00:00.000Z          FALSE        FALSE           FALSE
#>    2: 2024-11-04T12:00:00.000Z          FALSE         TRUE           FALSE
#>    3: 2024-11-04T12:00:00.000Z          FALSE        FALSE           FALSE
#>    4: 2024-11-04T12:00:00.000Z          FALSE        FALSE           FALSE
#>    5: 2024-11-04T12:30:00.000Z          FALSE        FALSE           FALSE
#>   ---                                                                     
#> 6296: 2025-04-05T20:00:00.000Z          FALSE         TRUE           FALSE
#> 6297: 2025-04-05T22:09:00.000Z          FALSE         TRUE           FALSE
#> 6298: 2025-04-06T01:28:00.000Z          FALSE         TRUE           FALSE
#> 6299: 2025-04-06T21:30:00.000Z          FALSE         TRUE           FALSE
#> 6300: 2025-04-08T00:50:00.000Z          FALSE         TRUE           FALSE
#>       game_type status                                            game_notes
#>          <char> <char>                                                <char>
#>    1:       STD  final                                                  <NA>
#>    2:    TRNMNT  final                                  Field of 68 Showcase
#>    3:       STD  final                                                  <NA>
#>    4:       STD  final                                                  <NA>
#>    5:       STD  final                                                  <NA>
#>   ---                                                                       
#> 6296:    TRNMNT  final                 College Basketball Crown - Semifinals
#> 6297:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 6298:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 6299:    TRNMNT  final            College Basketball Crown Championship Game
#> 6300:    TRNMNT  final Men's Basketball Championship - National Championship
#>       attendance home_team_id     home_team_name home_conference_id
#>            <int>        <int>             <char>              <int>
#>    1:       3716          115    IU Indianapolis                 14
#>    2:          0          273  Southern Illinois                 19
#>    3:       7501          171   Middle Tennessee                 13
#>    4:       1001          324       UT Arlington                 30
#>    5:       2055          128 Jacksonville State                 13
#>   ---                                                              
#> 6296:       2972          338          Villanova                  7
#> 6297:      68252           16             Auburn                 24
#> 6298:      68252           72               Duke                  2
#> 6299:       3314          312                UCF                  6
#> 6300:      66602          113            Houston                  6
#>       home_conference_short_name home_seed home_points home_period_points_1
#>                           <char>     <int>       <int>                <int>
#>    1:                    Horizon        NA         100                   48
#>    2:                        MVC        NA          80                   42
#>    3:                       CUSA        NA          97                   51
#>    4:                        WAC        NA          90                   48
#>    5:                       CUSA        NA         112                   55
#>   ---                                                                      
#> 6296:                   Big East        NA          98                   44
#> 6297:                        SEC        NA          73                   46
#> 6298:                        ACC        NA          67                   34
#> 6299:                     Big 12        NA          66                   35
#> 6300:                     Big 12        NA          63                   31
#>       home_period_points_2 home_period_points_3 home_period_points_4
#>                      <int>                <int>                <int>
#>    1:                   52                   NA                   NA
#>    2:                   38                   NA                   NA
#>    3:                   46                   NA                   NA
#>    4:                   42                   NA                   NA
#>    5:                   57                   NA                   NA
#>   ---                                                               
#> 6296:                   44                   10                   NA
#> 6297:                   27                   NA                   NA
#> 6298:                   33                   NA                   NA
#> 6299:                   31                   NA                   NA
#> 6300:                   32                   NA                   NA
#>       home_period_points_5 home_period_points_6 home_period_points_7
#>                      <int>                <int>                <int>
#>    1:                   NA                   NA                   NA
#>    2:                   NA                   NA                   NA
#>    3:                   NA                   NA                   NA
#>    4:                   NA                   NA                   NA
#>    5:                   NA                   NA                   NA
#>   ---                                                               
#> 6296:                   NA                   NA                   NA
#> 6297:                   NA                   NA                   NA
#> 6298:                   NA                   NA                   NA
#> 6299:                   NA                   NA                   NA
#> 6300:                   NA                   NA                   NA
#>       home_winner away_team_id away_team_name away_conference_id
#>            <lgcl>        <int>         <char>              <int>
#>    1:        TRUE          492    IU Columbus                 NA
#>    2:       FALSE           46     Charleston                 12
#>    3:        TRUE          493     Oglethorpe                 NA
#>    4:        TRUE          494     UNT Dallas                 NA
#>    5:        TRUE          495       LaGrange                 NA
#>   ---                                                           
#> 6296:       FALSE          312            UCF                  6
#> 6297:       FALSE           87        Florida                 24
#> 6298:       FALSE          113        Houston                  6
#> 6299:       FALSE          188       Nebraska                 10
#> 6300:       FALSE           87        Florida                 24
#>       away_conference_short_name away_seed away_points away_period_points_1
#>                           <char>     <int>       <int>                <int>
#>    1:                       <NA>        NA          44                   21
#>    2:                        CAA        NA          90                   45
#>    3:                       <NA>        NA          51                   22
#>    4:                       <NA>        NA          55                   25
#>    5:                       <NA>        NA          48                   20
#>   ---                                                                      
#> 6296:                     Big 12        NA         104                   48
#> 6297:                        SEC        NA          79                   38
#> 6298:                     Big 12        NA          70                   28
#> 6299:                    Big Ten        NA          77                   35
#> 6300:                        SEC        NA          65                   28
#>       away_period_points_2 away_period_points_3 away_period_points_4
#>                      <int>                <int>                <int>
#>    1:                   23                   NA                   NA
#>    2:                   45                   NA                   NA
#>    3:                   29                   NA                   NA
#>    4:                   30                   NA                   NA
#>    5:                   28                   NA                   NA
#>   ---                                                               
#> 6296:                   40                   16                   NA
#> 6297:                   41                   NA                   NA
#> 6298:                   42                   NA                   NA
#> 6299:                   42                   NA                   NA
#> 6300:                   37                   NA                   NA
#>       away_period_points_5 away_period_points_6 away_period_points_7
#>                      <int>                <int>                <int>
#>    1:                   NA                   NA                   NA
#>    2:                   NA                   NA                   NA
#>    3:                   NA                   NA                   NA
#>    4:                   NA                   NA                   NA
#>    5:                   NA                   NA                   NA
#>   ---                                                               
#> 6296:                   NA                   NA                   NA
#> 6297:                   NA                   NA                   NA
#> 6298:                   NA                   NA                   NA
#> 6299:                   NA                   NA                   NA
#> 6300:                   NA                   NA                   NA
#>       away_winner excitement venue_id                 venue   venue_city
#>            <lgcl>      <num>    <int>                <char>       <char>
#>    1:       FALSE        0.6       19      Corteva Coliseum Indianapolis
#>    2:        TRUE        7.2       22      Sanford Pentagon  Sioux Falls
#>    3:       FALSE        0.6       20         Murphy Center Murfreesboro
#>    4:       FALSE        1.5       21   College Park Center    Arlington
#>    5:       FALSE        1.0       23 Pete Mathews Coliseum Jacksonville
#>   ---                                                                   
#> 6296:        TRUE       12.3       13        T-Mobile Arena    Las Vegas
#> 6297:        TRUE        8.6      629             Alamodome  San Antonio
#> 6298:        TRUE        6.2      629             Alamodome  San Antonio
#> 6299:        TRUE        7.0       13        T-Mobile Arena    Las Vegas
#> 6300:        TRUE        7.7      629             Alamodome  San Antonio
#>       venue_state
#>            <char>
#>    1:          IN
#>    2:          SD
#>    3:          TN
#>    4:          TX
#>    5:          AL
#>   ---            
#> 6296:          NV
#> 6297:          TX
#> 6298:          TX
#> 6299:          NV
#> 6300:          TX

load_games(2024)
#>       game_id game_source_id season_label season season_type tournament
#>         <int>         <char>       <char>  <int>      <char>     <char>
#>    1:    5901      401594209     20232024   2024     regular       <NA>
#>    2:    5902      401610539     20232024   2024     regular       <NA>
#>    3:    5903      401591565     20232024   2024     regular       <NA>
#>    4:    5904      401604410     20232024   2024     regular       <NA>
#>    5:    5905      401598014     20232024   2024     regular       <NA>
#>   ---                                                                  
#> 6245:   12124      401641122     20232024   2024  postseason        NIT
#> 6246:   12125      401641124     20232024   2024  postseason        NIT
#> 6247:   12126      401638643     20232024   2024  postseason       NCAA
#> 6248:   12127      401638644     20232024   2024  postseason       NCAA
#> 6249:   12128      401638645     20232024   2024  postseason       NCAA
#>                     start_date start_time_tbd neutral_site conference_game
#>                         <char>         <lgcl>       <lgcl>          <lgcl>
#>    1: 2023-11-06T11:00:00.000Z          FALSE        FALSE           FALSE
#>    2: 2023-11-06T11:30:00.000Z          FALSE        FALSE           FALSE
#>    3: 2023-11-06T12:00:00.000Z          FALSE        FALSE           FALSE
#>    4: 2023-11-06T12:30:00.000Z          FALSE        FALSE           FALSE
#>    5: 2023-11-06T13:00:00.000Z          FALSE        FALSE           FALSE
#>   ---                                                                     
#> 6245: 2024-04-02T21:35:00.000Z          FALSE         TRUE           FALSE
#> 6246: 2024-04-04T19:00:00.000Z          FALSE         TRUE           FALSE
#> 6247: 2024-04-06T18:09:00.000Z          FALSE         TRUE           FALSE
#> 6248: 2024-04-06T20:51:00.000Z          FALSE         TRUE           FALSE
#> 6249: 2024-04-08T21:20:00.000Z          FALSE         TRUE           FALSE
#>       game_type status                                            game_notes
#>          <char> <char>                                                <char>
#>    1:       STD  final                                                  <NA>
#>    2:       STD  final                                                  <NA>
#>    3:       STD  final                                                  <NA>
#>    4:       STD  final                                                  <NA>
#>    5:       STD  final                                                  <NA>
#>   ---                                                                       
#> 6245:    TRNMNT  final                                       NIT - Semifinal
#> 6246:    TRNMNT  final                                    NIT - Championship
#> 6247:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 6248:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 6249:    TRNMNT  final Men's Basketball Championship - National Championship
#>       attendance home_team_id    home_team_name home_conference_id
#>            <int>        <int>            <char>              <int>
#>    1:       4867          115   IU Indianapolis                 14
#>    2:       2147          110           Hofstra                 12
#>    3:       3321          302              Troy                 29
#>    4:       3250          282 Stephen F. Austin                 27
#>    5:        350          232  Prairie View A&M                 25
#>   ---                                                             
#> 6245:          0          263        Seton Hall                  7
#> 6246:       9100          263        Seton Hall                  7
#> 6247:          0          236            Purdue                 10
#> 6248:      74720          314             UConn                  7
#> 6249:      74423          314             UConn                  7
#>       home_conference_short_name home_seed home_points home_period_points_1
#>                           <char>     <int>       <int>                <int>
#>    1:                    Horizon        NA          70                   33
#>    2:                        CAA        NA         101                   44
#>    3:                   Sun Belt        NA          92                   38
#>    4:                  Southland        NA          96                   54
#>    5:                       SWAC        NA          89                   43
#>   ---                                                                      
#> 6245:                   Big East        NA          84                   42
#> 6246:                   Big East        NA          79                   39
#> 6247:                    Big Ten         1          63                   35
#> 6248:                   Big East         1          86                   44
#> 6249:                   Big East         1          75                   36
#>       home_period_points_2 home_period_points_3 home_period_points_4
#>                      <int>                <int>                <int>
#>    1:                   37                   NA                   NA
#>    2:                   57                   NA                   NA
#>    3:                   54                   NA                   NA
#>    4:                   42                   NA                   NA
#>    5:                   46                   NA                   NA
#>   ---                                                               
#> 6245:                   42                   NA                   NA
#> 6246:                   40                   NA                   NA
#> 6247:                   28                   NA                   NA
#> 6248:                   42                   NA                   NA
#> 6249:                   39                   NA                   NA
#>       home_period_points_5 home_period_points_6 home_period_points_7
#>                      <int>                <int>                <int>
#>    1:                   NA                   NA                   NA
#>    2:                   NA                   NA                   NA
#>    3:                   NA                   NA                   NA
#>    4:                   NA                   NA                   NA
#>    5:                   NA                   NA                   NA
#>   ---                                                               
#> 6245:                   NA                   NA                   NA
#> 6246:                   NA                   NA                   NA
#> 6247:                   NA                   NA                   NA
#> 6248:                   NA                   NA                   NA
#> 6249:                   NA                   NA                   NA
#>       home_winner away_team_id             away_team_name away_conference_id
#>            <lgcl>        <int>                     <char>              <int>
#>    1:        TRUE          831                   Spalding                 NA
#>    2:        TRUE          677 Saint Joseph's Long Island                 NA
#>    3:        TRUE          797            Fort Lauderdale                 NA
#>    4:        TRUE          832             North American                 NA
#>    5:        TRUE          614           Kansas Christian                 NA
#>   ---                                                                       
#> 6245:        TRUE           98                    Georgia                 24
#> 6246:        TRUE          122              Indiana State                 19
#> 6247:        TRUE          185                   NC State                  2
#> 6248:        TRUE            5                    Alabama                 24
#> 6249:        TRUE          236                     Purdue                 10
#>       away_conference_short_name away_seed away_points away_period_points_1
#>                           <char>     <int>       <int>                <int>
#>    1:                       <NA>        NA          63                   25
#>    2:                       <NA>        NA          48                   19
#>    3:                       <NA>        NA          47                   18
#>    4:                       <NA>        NA          68                   22
#>    5:                       <NA>        NA          66                   33
#>   ---                                                                      
#> 6245:                        SEC        NA          67                   25
#> 6246:                        MVC        NA          77                   39
#> 6247:                        ACC        11          50                   29
#> 6248:                        SEC         4          72                   40
#> 6249:                    Big Ten         1          60                   30
#>       away_period_points_2 away_period_points_3 away_period_points_4
#>                      <int>                <int>                <int>
#>    1:                   38                   NA                   NA
#>    2:                   29                   NA                   NA
#>    3:                   29                   NA                   NA
#>    4:                   46                   NA                   NA
#>    5:                   33                   NA                   NA
#>   ---                                                               
#> 6245:                   42                   NA                   NA
#> 6246:                   38                   NA                   NA
#> 6247:                   21                   NA                   NA
#> 6248:                   32                   NA                   NA
#> 6249:                   30                   NA                   NA
#>       away_period_points_5 away_period_points_6 away_period_points_7
#>                      <int>                <int>                <int>
#>    1:                   NA                   NA                   NA
#>    2:                   NA                   NA                   NA
#>    3:                   NA                   NA                   NA
#>    4:                   NA                   NA                   NA
#>    5:                   NA                   NA                   NA
#>   ---                                                               
#> 6245:                   NA                   NA                   NA
#> 6246:                   NA                   NA                   NA
#> 6247:                   NA                   NA                   NA
#> 6248:                   NA                   NA                   NA
#> 6249:                   NA                   NA                   NA
#>       away_winner excitement venue_id
#>            <lgcl>      <num>    <int>
#>    1:       FALSE        2.7       19
#>    2:       FALSE        1.3       27
#>    3:       FALSE        1.2       65
#>    4:       FALSE        0.9      154
#>    5:       FALSE        2.2       56
#>   ---                                
#> 6245:       FALSE        1.3       40
#> 6246:       FALSE        7.2       40
#> 6247:       FALSE        2.3      485
#> 6248:       FALSE        4.5      485
#> 6249:       FALSE        3.0      485
#>                                             venue   venue_city venue_state
#>                                            <char>       <char>      <char>
#>    1:                            Corteva Coliseum Indianapolis          IN
#>    2: David S. Mack Sports and Exhibition Complex    Hempstead          NY
#>    3:                                Trojan Arena         Troy          AL
#>    4:                 William R. Johnson Coliseum  Nacogdoches          TX
#>    5:                      William Nicks Building Prairie View          TX
#>   ---                                                                     
#> 6245:                           Hinkle Fieldhouse Indianapolis          IN
#> 6246:                           Hinkle Fieldhouse Indianapolis          IN
#> 6247:                          State Farm Stadium     Glendale          AZ
#> 6248:                          State Farm Stadium     Glendale          AZ
#> 6249:                          State Farm Stadium     Glendale          AZ

load_games(TRUE)
#>         game_id game_source_id season_label season season_type tournament
#>           <int>         <char>       <char>  <int>      <char>     <char>
#>      1:  127127      223180183     20022003   2003     regular       <NA>
#>      2:  127126      223180201     20022003   2003     regular       <NA>
#>      3:  127129      223190222     20022003   2003     regular       <NA>
#>      4:  127130      223192459     20022003   2003     regular       <NA>
#>      5:  127131      223190275     20022003   2003     regular       <NA>
#>     ---                                                                  
#> 132536:  209747      401758633     20242025   2025  postseason       <NA>
#> 132537:  209731      401746081     20242025   2025  postseason       NCAA
#> 132538:  209732      401746080     20242025   2025  postseason       NCAA
#> 132539:  209749      401758636     20242025   2025  postseason       <NA>
#> 132540:  209751      401746082     20242025   2025  postseason       NCAA
#>                       start_date start_time_tbd neutral_site conference_game
#>                           <char>         <lgcl>       <lgcl>          <lgcl>
#>      1: 2002-11-14T19:00:00.000Z          FALSE        FALSE           FALSE
#>      2: 2002-11-14T21:30:00.000Z          FALSE        FALSE           FALSE
#>      3: 2002-11-15T18:30:00.000Z          FALSE        FALSE           FALSE
#>      4: 2002-11-15T19:00:00.000Z          FALSE        FALSE           FALSE
#>      5: 2002-11-15T21:00:00.000Z          FALSE        FALSE           FALSE
#>     ---                                                                     
#> 132536: 2025-04-05T20:00:00.000Z          FALSE         TRUE           FALSE
#> 132537: 2025-04-05T22:09:00.000Z          FALSE         TRUE           FALSE
#> 132538: 2025-04-06T01:28:00.000Z          FALSE         TRUE           FALSE
#> 132539: 2025-04-06T21:30:00.000Z          FALSE         TRUE           FALSE
#> 132540: 2025-04-08T00:50:00.000Z          FALSE         TRUE           FALSE
#>         game_type status                                            game_notes
#>            <char> <char>                                                <char>
#>      1:       STD  final                                                  <NA>
#>      2:       STD  final                                                  <NA>
#>      3:       STD  final                                                  <NA>
#>      4:       STD  final                                                  <NA>
#>      5:       STD  final                                                  <NA>
#>     ---                                                                       
#> 132536:    TRNMNT  final                 College Basketball Crown - Semifinals
#> 132537:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 132538:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 132539:    TRNMNT  final            College Basketball Crown Championship Game
#> 132540:    TRNMNT  final Men's Basketball Championship - National Championship
#>         attendance home_team_id    home_team_name home_conference_id
#>              <int>        <int>            <char>              <int>
#>      1:          0          286          Syracuse                  7
#>      2:       8826          217          Oklahoma                  6
#>      3:          0          338         Villanova                  7
#>      4:          0          208 Northern Illinois                 17
#>      5:      16374          355         Wisconsin                 10
#>     ---                                                             
#> 132536:       2972          338         Villanova                  7
#> 132537:      68252           16            Auburn                 24
#> 132538:      68252           72              Duke                  2
#> 132539:       3314          312               UCF                  6
#> 132540:      66602          113           Houston                  6
#>         home_conference_short_name home_seed home_points home_period_points_1
#>                             <char>     <int>       <int>                <int>
#>      1:                   Big East        NA          63                   34
#>      2:                     Big 12        NA          62                   25
#>      3:                   Big East        NA          61                   23
#>      4:                        MAC        NA          56                   31
#>      5:                    Big Ten        NA          81                   34
#>     ---                                                                      
#> 132536:                   Big East        NA          98                   44
#> 132537:                        SEC        NA          73                   46
#> 132538:                        ACC        NA          67                   34
#> 132539:                     Big 12        NA          66                   35
#> 132540:                     Big 12        NA          63                   31
#>         home_period_points_2 home_period_points_3 home_period_points_4
#>                        <int>                <int>                <int>
#>      1:                   29                   NA                   NA
#>      2:                   37                   NA                   NA
#>      3:                   38                   NA                   NA
#>      4:                   25                   NA                   NA
#>      5:                   47                   NA                   NA
#>     ---                                                               
#> 132536:                   44                   10                   NA
#> 132537:                   27                   NA                   NA
#> 132538:                   33                   NA                   NA
#> 132539:                   31                   NA                   NA
#> 132540:                   32                   NA                   NA
#>         home_period_points_5 home_period_points_6 home_period_points_7
#>                        <int>                <int>                <int>
#>      1:                   NA                   NA                   NA
#>      2:                   NA                   NA                   NA
#>      3:                   NA                   NA                   NA
#>      4:                   NA                   NA                   NA
#>      5:                   NA                   NA                   NA
#>     ---                                                               
#> 132536:                   NA                   NA                   NA
#> 132537:                   NA                   NA                   NA
#> 132538:                   NA                   NA                   NA
#> 132539:                   NA                   NA                   NA
#> 132540:                   NA                   NA                   NA
#>         home_winner away_team_id     away_team_name away_conference_id
#>              <lgcl>        <int>             <char>              <int>
#>      1:       FALSE          163            Memphis                 13
#>      2:       FALSE            5            Alabama                 24
#>      3:       FALSE          157          Marquette                 13
#>      4:        TRUE          354           Winthrop                  9
#>      5:        TRUE           80 Eastern Washington                  8
#>     ---                                                               
#> 132536:       FALSE          312                UCF                  6
#> 132537:       FALSE           87            Florida                 24
#> 132538:       FALSE          113            Houston                  6
#> 132539:       FALSE          188           Nebraska                 10
#> 132540:       FALSE           87            Florida                 24
#>         away_conference_short_name away_seed away_points away_period_points_1
#>                             <char>     <int>       <int>                <int>
#>      1:                       CUSA        NA          70                   43
#>      2:                        SEC        NA          68                   33
#>      3:                       CUSA        NA          73                   37
#>      4:                  Big South        NA          50                   30
#>      5:                    Big Sky        NA          55                   25
#>     ---                                                                      
#> 132536:                     Big 12        NA         104                   48
#> 132537:                        SEC        NA          79                   38
#> 132538:                     Big 12        NA          70                   28
#> 132539:                    Big Ten        NA          77                   35
#> 132540:                        SEC        NA          65                   28
#>         away_period_points_2 away_period_points_3 away_period_points_4
#>                        <int>                <int>                <int>
#>      1:                   27                   NA                   NA
#>      2:                   35                   NA                   NA
#>      3:                   36                   NA                   NA
#>      4:                   20                   NA                   NA
#>      5:                   30                   NA                   NA
#>     ---                                                               
#> 132536:                   40                   16                   NA
#> 132537:                   41                   NA                   NA
#> 132538:                   42                   NA                   NA
#> 132539:                   42                   NA                   NA
#> 132540:                   37                   NA                   NA
#>         away_period_points_5 away_period_points_6 away_period_points_7
#>                        <int>                <int>                <int>
#>      1:                   NA                   NA                   NA
#>      2:                   NA                   NA                   NA
#>      3:                   NA                   NA                   NA
#>      4:                   NA                   NA                   NA
#>      5:                   NA                   NA                   NA
#>     ---                                                               
#> 132536:                   NA                   NA                   NA
#> 132537:                   NA                   NA                   NA
#> 132538:                   NA                   NA                   NA
#> 132539:                   NA                   NA                   NA
#> 132540:                   NA                   NA                   NA
#>         away_winner excitement venue_id                 venue  venue_city
#>              <lgcl>      <num>    <int>                <char>      <char>
#>      1:        TRUE         NA      371 Madison Square Garden    New York
#>      2:        TRUE         NA      371 Madison Square Garden    New York
#>      3:        TRUE         NA      371 Madison Square Garden    New York
#>      4:       FALSE         NA      142           Kohl Center     Madison
#>      5:       FALSE         NA      142           Kohl Center     Madison
#>     ---                                                                  
#> 132536:        TRUE       12.3       13        T-Mobile Arena   Las Vegas
#> 132537:        TRUE        8.6      629             Alamodome San Antonio
#> 132538:        TRUE        6.2      629             Alamodome San Antonio
#> 132539:        TRUE        7.0       13        T-Mobile Arena   Las Vegas
#> 132540:        TRUE        7.7      629             Alamodome San Antonio
#>         venue_state home_period_points_8 away_period_points_8
#>              <char>                <int>                <int>
#>      1:          NY                   NA                   NA
#>      2:          NY                   NA                   NA
#>      3:          NY                   NA                   NA
#>      4:          WI                   NA                   NA
#>      5:          WI                   NA                   NA
#>     ---                                                      
#> 132536:          NV                   NA                   NA
#> 132537:          TX                   NA                   NA
#> 132538:          TX                   NA                   NA
#> 132539:          NV                   NA                   NA
#> 132540:          TX                   NA                   NA
# }