Load college basketball games for specified season(s)

Pull in all college basketball games tracked by collegebasketballdata.com

Returns data for all NCAA D1 men's basketball games for the specified seasons. Includes contextual information, such as where the game took place and who it is between, as well as high level box score data, such as the scores of each team.

This will return both games in the past and scheduled games. If a game is scheduled but its opponents have not yet been determined (i.e., the national championship game of the upcoming season), it will not appear in this dataset.

Usage

load_games(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of all games in the specified season(s).

See also

Examples

# \donttest{
try({ # avoid cran errors
load_games()
})
#>       game_id game_source_id season_label season season_type tournament
#>         <int>         <char>       <char>  <int>      <char>     <char>
#>    1:      24      401727499     20242025   2025     regular       <NA>
#>    2:      27      401715872     20242025   2025     regular       <NA>
#>    3:      25      401726400     20242025   2025     regular       <NA>
#>    4:      26      401725796     20242025   2025     regular       <NA>
#>    5:      28      401726395     20242025   2025     regular       <NA>
#>   ---                                                                  
#> 6296:  209747      401758633     20242025   2025  postseason       <NA>
#> 6297:  209731      401746081     20242025   2025  postseason       NCAA
#> 6298:  209732      401746080     20242025   2025  postseason       NCAA
#> 6299:  209749      401758636     20242025   2025  postseason       <NA>
#> 6300:  209751      401746082     20242025   2025  postseason       NCAA
#>                     start_date start_time_tbd neutral_site conference_game
#>                         <char>         <lgcl>       <lgcl>          <lgcl>
#>    1: 2024-11-04T11:00:00.000Z          FALSE        FALSE           FALSE
#>    2: 2024-11-04T12:00:00.000Z          FALSE         TRUE           FALSE
#>    3: 2024-11-04T12:00:00.000Z          FALSE        FALSE           FALSE
#>    4: 2024-11-04T12:00:00.000Z          FALSE        FALSE           FALSE
#>    5: 2024-11-04T12:30:00.000Z          FALSE        FALSE           FALSE
#>   ---                                                                     
#> 6296: 2025-04-05T20:00:00.000Z          FALSE         TRUE           FALSE
#> 6297: 2025-04-05T22:09:00.000Z          FALSE         TRUE           FALSE
#> 6298: 2025-04-06T01:28:00.000Z          FALSE         TRUE           FALSE
#> 6299: 2025-04-06T21:30:00.000Z          FALSE         TRUE           FALSE
#> 6300: 2025-04-08T00:50:00.000Z          FALSE         TRUE           FALSE
#>       game_type status                                            game_notes
#>          <char> <char>                                                <char>
#>    1:       STD  final                                                  <NA>
#>    2:    TRNMNT  final                                  Field of 68 Showcase
#>    3:       STD  final                                                  <NA>
#>    4:       STD  final                                                  <NA>
#>    5:       STD  final                                                  <NA>
#>   ---                                                                       
#> 6296:    TRNMNT  final                 College Basketball Crown - Semifinals
#> 6297:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 6298:    TRNMNT  final            Men's Basketball Championship - Final Four
#> 6299:    TRNMNT  final            College Basketball Crown Championship Game
#> 6300:    TRNMNT  final Men's Basketball Championship - National Championship
#>       attendance home_team_id     home_team_name home_conference_id
#>            <int>        <int>             <char>              <int>
#>    1:       3716          115    IU Indianapolis                 14
#>    2:          0          273  Southern Illinois                 19
#>    3:       7501          171   Middle Tennessee                 13
#>    4:       1001          324       UT Arlington                 30
#>    5:       2055          128 Jacksonville State                 13
#>   ---                                                              
#> 6296:       2972          338          Villanova                  7
#> 6297:      68252           16             Auburn                 24
#> 6298:      68252           72               Duke                  2
#> 6299:       3314          312                UCF                  6
#> 6300:      66602          113            Houston                  6
#>       home_conference_short_name home_seed home_points home_period_points_1
#>                           <char>     <int>       <int>                <int>
#>    1:                    Horizon        NA         100                   48
#>    2:                        MVC        NA          80                   42
#>    3:                       CUSA        NA          97                   51
#>    4:                        WAC        NA          90                   48
#>    5:                       CUSA        NA         112                   55
#>   ---                                                                      
#> 6296:                   Big East        NA          98                   44
#> 6297:                        SEC        NA          73                   46
#> 6298:                        ACC        NA          67                   34
#> 6299:                     Big 12        NA          66                   35
#> 6300:                     Big 12        NA          63                   31
#>       home_period_points_2 home_period_points_3 home_period_points_4
#>                      <int>                <int>                <int>
#>    1:                   52                   NA                   NA
#>    2:                   38                   NA                   NA
#>    3:                   46                   NA                   NA
#>    4:                   42                   NA                   NA
#>    5:                   57                   NA                   NA
#>   ---                                                               
#> 6296:                   44                   10                   NA
#> 6297:                   27                   NA                   NA
#> 6298:                   33                   NA                   NA
#> 6299:                   31                   NA                   NA
#> 6300:                   32                   NA                   NA
#>       home_period_points_5 home_period_points_6 home_period_points_7
#>                      <int>                <int>                <int>
#>    1:                   NA                   NA                   NA
#>    2:                   NA                   NA                   NA
#>    3:                   NA                   NA                   NA
#>    4:                   NA                   NA                   NA
#>    5:                   NA                   NA                   NA
#>   ---                                                               
#> 6296:                   NA                   NA                   NA
#> 6297:                   NA                   NA                   NA
#> 6298:                   NA                   NA                   NA
#> 6299:                   NA                   NA                   NA
#> 6300:                   NA                   NA                   NA
#>       home_winner away_team_id away_team_name away_conference_id
#>            <lgcl>        <int>         <char>              <int>
#>    1:        TRUE          492    IU Columbus                 NA
#>    2:       FALSE           46     Charleston                 12
#>    3:        TRUE          493     Oglethorpe                 NA
#>    4:        TRUE          494     UNT Dallas                 NA
#>    5:        TRUE          495       LaGrange                 NA
#>   ---                                                           
#> 6296:       FALSE          312            UCF                  6
#> 6297:       FALSE           87        Florida                 24
#> 6298:       FALSE          113        Houston                  6
#> 6299:       FALSE          188       Nebraska                 10
#> 6300:       FALSE           87        Florida                 24
#>       away_conference_short_name away_seed away_points away_period_points_1
#>                           <char>     <int>       <int>                <int>
#>    1:                       <NA>        NA          44                   21
#>    2:                        CAA        NA          90                   45
#>    3:                       <NA>        NA          51                   22
#>    4:                       <NA>        NA          55                   25
#>    5:                       <NA>        NA          48                   20
#>   ---                                                                      
#> 6296:                     Big 12        NA         104                   48
#> 6297:                        SEC        NA          79                   38
#> 6298:                     Big 12        NA          70                   28
#> 6299:                    Big Ten        NA          77                   35
#> 6300:                        SEC        NA          65                   28
#>       away_period_points_2 away_period_points_3 away_period_points_4
#>                      <int>                <int>                <int>
#>    1:                   23                   NA                   NA
#>    2:                   45                   NA                   NA
#>    3:                   29                   NA                   NA
#>    4:                   30                   NA                   NA
#>    5:                   28                   NA                   NA
#>   ---                                                               
#> 6296:                   40                   16                   NA
#> 6297:                   41                   NA                   NA
#> 6298:                   42                   NA                   NA
#> 6299:                   42                   NA                   NA
#> 6300:                   37                   NA                   NA
#>       away_period_points_5 away_period_points_6 away_period_points_7
#>                      <int>                <int>                <int>
#>    1:                   NA                   NA                   NA
#>    2:                   NA                   NA                   NA
#>    3:                   NA                   NA                   NA
#>    4:                   NA                   NA                   NA
#>    5:                   NA                   NA                   NA
#>   ---                                                               
#> 6296:                   NA                   NA                   NA
#> 6297:                   NA                   NA                   NA
#> 6298:                   NA                   NA                   NA
#> 6299:                   NA                   NA                   NA
#> 6300:                   NA                   NA                   NA
#>       away_winner excitement venue_id                 venue   venue_city
#>            <lgcl>      <num>    <int>                <char>       <char>
#>    1:       FALSE        0.6       19      Corteva Coliseum Indianapolis
#>    2:        TRUE        7.2       22      Sanford Pentagon  Sioux Falls
#>    3:       FALSE        0.6       20         Murphy Center Murfreesboro
#>    4:       FALSE        1.5       21   College Park Center    Arlington
#>    5:       FALSE        1.0       23 Pete Mathews Coliseum Jacksonville
#>   ---                                                                   
#> 6296:        TRUE       12.3       13        T-Mobile Arena    Las Vegas
#> 6297:        TRUE        8.6      629             Alamodome  San Antonio
#> 6298:        TRUE        6.2      629             Alamodome  San Antonio
#> 6299:        TRUE        7.0       13        T-Mobile Arena    Las Vegas
#> 6300:        TRUE        7.7      629             Alamodome  San Antonio
#>       venue_state
#>            <char>
#>    1:          IN
#>    2:          SD
#>    3:          TN
#>    4:          TX
#>    5:          AL
#>   ---            
#> 6296:          NV
#> 6297:          TX
#> 6298:          TX
#> 6299:          NV
#> 6300:          TX
# }