Load media for cbb games

Pull in broadcast information for all games in a given season

Usage

load_media(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of broadcast information for all games for the specified season(s).

See also

Examples

# \donttest{
load_media()
#>       game_id season season_label season_type               start_date
#>         <int>  <int>       <char>      <char>                   <char>
#>    1:      24   2025     20242025     regular 2024-11-04T11:00:00.000Z
#>    2:      26   2025     20242025     regular 2024-11-04T12:00:00.000Z
#>    3:      25   2025     20242025     regular 2024-11-04T12:00:00.000Z
#>    4:      28   2025     20242025     regular 2024-11-04T12:30:00.000Z
#>    5:      29   2025     20242025     regular 2024-11-04T15:00:00.000Z
#>   ---                                                                 
#> 6698:  209747   2025     20242025  postseason 2025-04-05T20:00:00.000Z
#> 6699:  209731   2025     20242025  postseason 2025-04-05T22:09:00.000Z
#> 6700:  209732   2025     20242025  postseason 2025-04-06T01:28:00.000Z
#> 6701:  209749   2025     20242025  postseason 2025-04-06T21:30:00.000Z
#> 6702:  209751   2025     20242025  postseason 2025-04-08T00:50:00.000Z
#>       start_time_tbd home_team_id          home_team home_conference_short_name
#>               <lgcl>        <int>             <char>                     <char>
#>    1:          FALSE          115    IU Indianapolis                    Horizon
#>    2:          FALSE          324       UT Arlington                        WAC
#>    3:          FALSE          171   Middle Tennessee                       CUSA
#>    4:          FALSE          128 Jacksonville State                       CUSA
#>    5:          FALSE          261        Santa Clara                        WCC
#>   ---                                                                          
#> 6698:          FALSE          338          Villanova                   Big East
#> 6699:          FALSE           16             Auburn                        SEC
#> 6700:          FALSE           72               Duke                        ACC
#> 6701:          FALSE          312                UCF                     Big 12
#> 6702:          FALSE          113            Houston                     Big 12
#>       away_team_id   away_team away_conference_short_name neutral_site
#>              <int>      <char>                     <char>       <lgcl>
#>    1:          492 IU Columbus                       <NA>        FALSE
#>    2:          494  UNT Dallas                       <NA>        FALSE
#>    3:          493  Oglethorpe                       <NA>        FALSE
#>    4:          495    LaGrange                       <NA>        FALSE
#>    5:          252 Saint Louis                       A-10         TRUE
#>   ---                                                                 
#> 6698:          312         UCF                     Big 12         TRUE
#> 6699:           87     Florida                        SEC         TRUE
#> 6700:          113     Houston                     Big 12         TRUE
#> 6701:          188    Nebraska                    Big Ten         TRUE
#> 6702:           87     Florida                        SEC         TRUE
#>       conference_game game_type
#>                <lgcl>    <char>
#>    1:           FALSE       STD
#>    2:           FALSE       STD
#>    3:           FALSE       STD
#>    4:           FALSE       STD
#>    5:           FALSE    TRNMNT
#>   ---                          
#> 6698:           FALSE    TRNMNT
#> 6699:           FALSE    TRNMNT
#> 6700:           FALSE    TRNMNT
#> 6701:           FALSE    TRNMNT
#> 6702:           FALSE    TRNMNT
#>                                                  game_notes broadcast_type
#>                                                      <char>         <char>
#>    1:                                                  <NA>      Streaming
#>    2:                                                  <NA>      Streaming
#>    3:                                                  <NA>      Streaming
#>    4:                                                  <NA>      Streaming
#>    5:                                  Field of 68 Showcase      Streaming
#>   ---                                                                     
#> 6698:                 College Basketball Crown - Semifinals             TV
#> 6699:            Men's Basketball Championship - Final Four             TV
#> 6700:            Men's Basketball Championship - Final Four             TV
#> 6701:            College Basketball Crown Championship Game             TV
#> 6702: Men's Basketball Championship - National Championship             TV
#>       broadcast_name
#>               <char>
#>    1:          ESPN+
#>    2:          ESPN+
#>    3:          ESPN+
#>    4:          ESPN+
#>    5:        YouTube
#>   ---               
#> 6698:            FOX
#> 6699:            CBS
#> 6700:            CBS
#> 6701:            FOX
#> 6702:            CBS

load_media(2024)
#>       game_id season season_label season_type               start_date
#>         <int>  <int>       <char>      <char>                   <char>
#>    1:    5901   2024     20232024     regular 2023-11-06T11:00:00.000Z
#>    2:    5903   2024     20232024     regular 2023-11-06T12:00:00.000Z
#>    3:    5904   2024     20232024     regular 2023-11-06T12:30:00.000Z
#>    4:    5906   2024     20232024     regular 2023-11-06T16:30:00.000Z
#>    5:    5907   2024     20232024     regular 2023-11-06T17:30:00.000Z
#>   ---                                                                 
#> 5523:   12125   2024     20232024  postseason 2024-04-04T19:00:00.000Z
#> 5524:   12125   2024     20232024  postseason 2024-04-04T19:00:00.000Z
#> 5525:   12126   2024     20232024  postseason 2024-04-06T18:09:00.000Z
#> 5526:   12127   2024     20232024  postseason 2024-04-06T20:51:00.000Z
#> 5527:   12128   2024     20232024  postseason 2024-04-08T21:20:00.000Z
#>       start_time_tbd home_team_id         home_team home_conference_short_name
#>               <lgcl>        <int>            <char>                     <char>
#>    1:          FALSE          115   IU Indianapolis                    Horizon
#>    2:          FALSE          302              Troy                   Sun Belt
#>    3:          FALSE          282 Stephen F. Austin                        WAC
#>    4:          FALSE          223            Oregon                     Pac-12
#>    5:          FALSE          224      Oregon State                     Pac-12
#>   ---                                                                         
#> 5523:          FALSE          263        Seton Hall                   Big East
#> 5524:          FALSE          263        Seton Hall                   Big East
#> 5525:          FALSE          236            Purdue                    Big Ten
#> 5526:          FALSE          314             UConn                   Big East
#> 5527:          FALSE          314             UConn                   Big East
#>       away_team_id        away_team away_conference_short_name neutral_site
#>              <int>           <char>                     <char>       <lgcl>
#>    1:          831         Spalding                       <NA>        FALSE
#>    2:          797  Fort Lauderdale                       <NA>        FALSE
#>    3:          832   North American                       <NA>        FALSE
#>    4:           98          Georgia                        SEC         TRUE
#>    5:          833 Linfield College                       <NA>        FALSE
#>   ---                                                                      
#> 5523:          122    Indiana State                        MVC         TRUE
#> 5524:          122    Indiana State                        MVC         TRUE
#> 5525:          185         NC State                        ACC         TRUE
#> 5526:            5          Alabama                        SEC         TRUE
#> 5527:          236           Purdue                    Big Ten         TRUE
#>       conference_game game_type
#>                <lgcl>    <char>
#>    1:           FALSE       STD
#>    2:           FALSE       STD
#>    3:           FALSE       STD
#>    4:           FALSE       STD
#>    5:           FALSE       STD
#>   ---                          
#> 5523:           FALSE    TRNMNT
#> 5524:           FALSE    TRNMNT
#> 5525:           FALSE    TRNMNT
#> 5526:           FALSE    TRNMNT
#> 5527:           FALSE    TRNMNT
#>                                                  game_notes broadcast_type
#>                                                      <char>         <char>
#>    1:                                                  <NA>      Streaming
#>    2:                                                  <NA>      Streaming
#>    3:                                                  <NA>      Streaming
#>    4:                       Hall of Fame Series - Las Vegas             TV
#>    5:                                                  <NA>             TV
#>   ---                                                                     
#> 5523:                                    NIT - Championship      Streaming
#> 5524:                                    NIT - Championship             TV
#> 5525:            Men's Basketball Championship - Final Four             TV
#> 5526:            Men's Basketball Championship - Final Four             TV
#> 5527: Men's Basketball Championship - National Championship             TV
#>       broadcast_name
#>               <char>
#>    1:          ESPN+
#>    2:          ESPN+
#>    3:          ESPN+
#>    4:          truTV
#>    5:           PACN
#>   ---               
#> 5523:          ESPN+
#> 5524:           ESPN
#> 5525:            TBS
#> 5526:            TBS
#> 5527:            TBS

load_media(TRUE)
#>        game_id season season_label season_type               start_date
#>          <int>  <int>       <char>      <char>                   <char>
#>     1:   83467   2011     20102011     regular 2010-11-08T19:00:00.000Z
#>     2:   83470   2011     20102011     regular 2010-11-08T21:00:00.000Z
#>     3:   83472   2011     20102011     regular 2010-11-10T19:00:00.000Z
#>     4:   83474   2011     20102011     regular 2010-11-10T21:00:00.000Z
#>     5:   83482   2011     20102011     regular 2010-11-12T18:30:00.000Z
#>    ---                                                                 
#> 59712:  209747   2025     20242025  postseason 2025-04-05T20:00:00.000Z
#> 59713:  209731   2025     20242025  postseason 2025-04-05T22:09:00.000Z
#> 59714:  209732   2025     20242025  postseason 2025-04-06T01:28:00.000Z
#> 59715:  209749   2025     20242025  postseason 2025-04-06T21:30:00.000Z
#> 59716:  209751   2025     20242025  postseason 2025-04-08T00:50:00.000Z
#>        start_time_tbd home_team_id  home_team home_conference away_team_id
#>                <lgcl>        <int>     <char>          <char>        <int>
#>     1:          FALSE          229 Pittsburgh            <NA>          240
#>     2:          FALSE          295      Texas            <NA>          187
#>     3:          FALSE          160   Maryland            <NA>           46
#>     4:          FALSE          295      Texas            <NA>          149
#>     5:          FALSE          121    Indiana            <NA>           88
#>    ---                                                                    
#> 59712:          FALSE          338  Villanova            <NA>          312
#> 59713:          FALSE           16     Auburn            <NA>           87
#> 59714:          FALSE           72       Duke            <NA>          113
#> 59715:          FALSE          312        UCF            <NA>          188
#> 59716:          FALSE          113    Houston            <NA>           87
#>                 away_team away_conference neutral_site conference_game
#>                    <char>          <char>       <lgcl>          <lgcl>
#>     1:       Rhode Island            <NA>        FALSE           FALSE
#>     2:               Navy            <NA>        FALSE           FALSE
#>     3:         Charleston            <NA>        FALSE           FALSE
#>     4:     Louisiana Tech            <NA>        FALSE           FALSE
#>     5: Florida Gulf Coast            <NA>        FALSE           FALSE
#>    ---                                                                
#> 59712:                UCF            <NA>         TRUE           FALSE
#> 59713:            Florida            <NA>         TRUE           FALSE
#> 59714:            Houston            <NA>         TRUE           FALSE
#> 59715:           Nebraska            <NA>         TRUE           FALSE
#> 59716:            Florida            <NA>         TRUE           FALSE
#>        game_type  notes home_conference_short_name away_conference_short_name
#>           <char> <char>                     <char>                     <char>
#>     1:       STD   <NA>                   Big East                       A-10
#>     2:       STD   <NA>                     Big 12                    Patriot
#>     3:       STD   <NA>                        ACC                      SoCon
#>     4:       STD   <NA>                     Big 12                        WAC
#>     5:       STD   <NA>                    Big Ten                       ASUN
#>    ---                                                                       
#> 59712:    TRNMNT   <NA>                   Big East                     Big 12
#> 59713:    TRNMNT   <NA>                        SEC                        SEC
#> 59714:    TRNMNT   <NA>                        ACC                     Big 12
#> 59715:    TRNMNT   <NA>                     Big 12                    Big Ten
#> 59716:    TRNMNT   <NA>                     Big 12                        SEC
#>                                                   game_notes broadcast_type
#>                                                       <char>         <char>
#>     1:                   2K SPORTS COACHES vs CANCER CLASSIC             TV
#>     2:                   2K SPORTS COACHES vs CANCER CLASSIC             TV
#>     3:                   2K SPORTS COACHES vs CANCER CLASSIC             TV
#>     4:                   2K SPORTS COACHES vs CANCER CLASSIC             TV
#>     5:                                                  <NA>             TV
#>    ---                                                                     
#> 59712:                 College Basketball Crown - Semifinals             TV
#> 59713:            Men's Basketball Championship - Final Four             TV
#> 59714:            Men's Basketball Championship - Final Four             TV
#> 59715:            College Basketball Crown Championship Game             TV
#> 59716: Men's Basketball Championship - National Championship             TV
#>        broadcast_name
#>                <char>
#>     1:          ESPNU
#>     2:          ESPNU
#>     3:          ESPNU
#>     4:          ESPNU
#>     5:            BTN
#>    ---               
#> 59712:            FOX
#> 59713:            CBS
#> 59714:            CBS
#> 59715:            FOX
#> 59716:            CBS
# }