Load cbb player box score stats

Pull in statistics aggregated by player and game for specified season(s)

This data includes both traditional box score stats (such as points, OREB, AST) as well as more advanced stats (such as USG%, ORTG). For more information on each stat, see the data dictionary dictionary_player_box_scores.

Usage

load_player_box_scores(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of player box scores.

See also

Examples

# \donttest{
try({ # avoid cran errors
load_player_box_scores()
})
#>        game_id season season_label season_type tournament
#>          <int>  <int>       <char>      <char>     <char>
#>     1:      24   2025     20242025     regular       <NA>
#>     2:      24   2025     20242025     regular       <NA>
#>     3:      24   2025     20242025     regular       <NA>
#>     4:      24   2025     20242025     regular       <NA>
#>     5:      24   2025     20242025     regular       <NA>
#>    ---                                                   
#> 63705:  209751   2025     20242025  postseason       NCAA
#> 63706:  209751   2025     20242025  postseason       NCAA
#> 63707:  209751   2025     20242025  postseason       NCAA
#> 63708:  209751   2025     20242025  postseason       NCAA
#> 63709:  209751   2025     20242025  postseason       NCAA
#>                      start_date start_time_tbd conference_game neutral_site
#>                          <char>         <lgcl>          <lgcl>       <lgcl>
#>     1: 2024-11-04T11:00:00.000Z          FALSE           FALSE        FALSE
#>     2: 2024-11-04T11:00:00.000Z          FALSE           FALSE        FALSE
#>     3: 2024-11-04T11:00:00.000Z          FALSE           FALSE        FALSE
#>     4: 2024-11-04T11:00:00.000Z          FALSE           FALSE        FALSE
#>     5: 2024-11-04T11:00:00.000Z          FALSE           FALSE        FALSE
#>    ---                                                                     
#> 63705: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#> 63706: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#> 63707: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#> 63708: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#> 63709: 2025-04-08T00:50:00.000Z          FALSE           FALSE         TRUE
#>        is_home game_type                                            game_notes
#>         <lgcl>    <char>                                                <char>
#>     1:   FALSE       STD                                                  <NA>
#>     2:   FALSE       STD                                                  <NA>
#>     3:   FALSE       STD                                                  <NA>
#>     4:   FALSE       STD                                                  <NA>
#>     5:   FALSE       STD                                                  <NA>
#>    ---                                                                        
#> 63705:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#> 63706:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#> 63707:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#> 63708:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#> 63709:   FALSE    TRNMNT Men's Basketball Championship - National Championship
#>        team_id   team_name team_conference_short_name team_seed opponent_id
#>          <int>      <char>                     <char>     <int>       <int>
#>     1:     492 IU Columbus                       <NA>        NA         115
#>     2:     492 IU Columbus                       <NA>        NA         115
#>     3:     492 IU Columbus                       <NA>        NA         115
#>     4:     492 IU Columbus                       <NA>        NA         115
#>     5:     492 IU Columbus                       <NA>        NA         115
#>    ---                                                                     
#> 63705:      87     Florida                        SEC        NA         113
#> 63706:      87     Florida                        SEC        NA         113
#> 63707:      87     Florida                        SEC        NA         113
#> 63708:      87     Florida                        SEC        NA         113
#> 63709:      87     Florida                        SEC        NA         113
#>          opponent_name opponent_conference_short_name opponent_seed
#>                 <char>                         <char>         <int>
#>     1: IU Indianapolis                        Horizon            NA
#>     2: IU Indianapolis                        Horizon            NA
#>     3: IU Indianapolis                        Horizon            NA
#>     4: IU Indianapolis                        Horizon            NA
#>     5: IU Indianapolis                        Horizon            NA
#>    ---                                                             
#> 63705:         Houston                         Big 12            NA
#> 63706:         Houston                         Big 12            NA
#> 63707:         Houston                         Big 12            NA
#> 63708:         Houston                         Big 12            NA
#> 63709:         Houston                         Big 12            NA
#>        game_minutes game_pace athlete_id athlete_source_id       athlete_name
#>               <int>     <num>      <int>            <char>             <char>
#>     1:           40        73        704           5246619      Gavin Dowling
#>     2:           40        73        705           5246612           Owen Law
#>     3:           40        73        706           5246607  Damontae Thompson
#>     4:           40        73        707           5246604     Nathan Pickett
#>     5:           40        73        708           5246603      Deonta Luster
#>    ---                                                                       
#> 63705:           40        67    4255910           4702656      Alijah Martin
#> 63706:           40        67    4255909           4896372 Walter Clayton Jr.
#> 63707:           40        67    4255908           4897262       Will Richard
#> 63708:           40        67    4255907           5174971    Rueben Chinyelu
#> 63709:           40        67    4255906           5174657        Alex Condon
#>        athlete_position starter ejected minutes points turnovers fouls assists
#>                  <char>  <lgcl>  <lgcl>   <int>  <int>     <int> <int>   <int>
#>     1:              ATH    TRUE   FALSE      31      9         2     3       4
#>     2:              ATH    TRUE   FALSE      21     13         1     2       0
#>     3:              ATH    TRUE   FALSE      25     10         3     1       1
#>     4:              ATH    TRUE   FALSE      22      0         2     0       0
#>     5:              ATH    TRUE   FALSE      17      0         0     4       1
#>    ---                                                                        
#> 63705:                G    TRUE   FALSE      34      7         1     0       0
#> 63706:                G    TRUE   FALSE      35     11         3     2       7
#> 63707:                G    TRUE   FALSE      33     18         2     2       2
#> 63708:                C    TRUE   FALSE      19      3         0     2       0
#> 63709:                F    TRUE   FALSE      28     12         4     3       0
#>        steals blocks game_score offensive_rating defensive_rating net_rating
#>         <int>  <int>      <num>            <num>            <num>      <num>
#>     1:      0      1        7.4            101.6            138.6      -37.0
#>     2:      2      0        7.4             92.8            127.9      -35.1
#>     3:      0      0        4.3             89.8            143.8      -54.0
#>     4:      0      0       -4.9              4.6            140.7     -136.1
#>     5:      0      0       -5.8              6.6            144.5     -137.9
#>    ---                                                                      
#> 63705:      0      0        0.5             68.6            108.4      -39.8
#> 63706:      1      1        9.5             97.4             94.2        3.2
#> 63707:      2      1       17.1            132.6             85.7       46.9
#> 63708:      0      1        3.8            123.4             95.5       27.9
#> 63709:      4      0        9.0             82.6             75.9        6.7
#>        usage effective_field_goal_pct true_shooting_pct assists_turnover_ratio
#>        <num>                    <num>             <num>                  <num>
#>     1: 0.160                    0.750             0.580                    2.0
#>     2: 0.314                    0.542             0.542                    0.0
#>     3: 0.212                    0.714             0.672                    0.3
#>     4: 0.184                    0.000             0.000                    0.0
#>     5: 0.209                    0.000             0.000                     NA
#>    ---                                                                        
#> 63705: 0.186                    0.250             0.322                    0.0
#> 63706: 0.224                    0.350             0.468                    2.3
#> 63707: 0.222                    0.700             0.765                    1.0
#> 63708: 0.081                    0.500             0.521                     NA
#> 63709: 0.264                    0.556             0.607                    0.0
#>        free_throw_rate offensive_rebound_pct field_goals_made
#>                  <num>                 <num>            <int>
#>     1:           0.667                 0.400                4
#>     2:           0.000                 0.000                5
#>     3:           0.143                 0.000                5
#>     4:           0.000                 0.333                0
#>     5:           0.000                    NA                0
#>    ---                                                       
#> 63705:           0.200                 1.000                2
#> 63706:           0.400                 0.000                3
#> 63707:           0.400                 0.125                5
#> 63708:           1.000                 0.400                1
#> 63709:           0.222                 0.143                5
#>        field_goals_attempted field_goals_pct two_point_field_goals_made
#>                        <int>           <num>                      <int>
#>     1:                     6           0.667                          3
#>     2:                    12           0.417                          2
#>     3:                     7           0.714                          5
#>     4:                     6           0.000                          0
#>     5:                     7           0.000                          0
#>    ---                                                                 
#> 63705:                    10           0.200                          1
#> 63706:                    10           0.300                          2
#> 63707:                    10           0.500                          1
#> 63708:                     2           0.500                          1
#> 63709:                     9           0.556                          5
#>        two_point_field_goals_attempted two_point_field_goals_pct
#>                                  <int>                     <num>
#>     1:                               5                     0.600
#>     2:                               7                     0.286
#>     3:                               7                     0.714
#>     4:                               2                     0.000
#>     5:                               4                     0.000
#>    ---                                                          
#> 63705:                               4                     0.250
#> 63706:                               3                     0.667
#> 63707:                               3                     0.333
#> 63708:                               2                     0.500
#> 63709:                               8                     0.625
#>        three_point_field_goals_made three_point_field_goals_attempted
#>                               <int>                             <int>
#>     1:                            1                                 1
#>     2:                            3                                 5
#>     3:                            0                                 0
#>     4:                            0                                 4
#>     5:                            0                                 3
#>    ---                                                               
#> 63705:                            1                                 6
#> 63706:                            1                                 7
#> 63707:                            4                                 7
#> 63708:                            0                                 0
#> 63709:                            0                                 1
#>        three_point_field_goals_pct free_throws_made free_throws_attempted
#>                              <num>            <int>                 <int>
#>     1:                       1.000                0                     4
#>     2:                       0.600                0                     0
#>     3:                       0.000                0                     1
#>     4:                       0.000                0                     0
#>     5:                       0.000                0                     0
#>    ---                                                                   
#> 63705:                       0.167                2                     2
#> 63706:                       0.143                4                     4
#> 63707:                       0.571                4                     4
#> 63708:                       0.000                1                     2
#> 63709:                       0.000                2                     2
#>        free_throws_pct rebounds_offensive rebounds_defensive rebounds_total
#>                  <num>              <int>              <int>          <int>
#>     1:             0.0                  2                  3              5
#>     2:             0.0                  0                  2              2
#>     3:             0.0                  0                  1              1
#>     4:             0.0                  1                  2              3
#>     5:             0.0                  0                  0              0
#>    ---                                                                     
#> 63705:             1.0                  1                  0              1
#> 63706:             1.0                  0                  5              5
#> 63707:             1.0                  1                  7              8
#> 63708:             0.5                  2                  3              5
#> 63709:             1.0                  1                  6              7
# }