Load cbb player stats

Pull in statistics aggregated by player for specified season(s)

This data includes both traditional box score stats (such as points, OREB, AST) as well as more advanced stats (such as USG%, ORTG). For more information on each stat, see the data dictionary dictionary_player_stats.

Usage

load_player_stats(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of player statistics for college basketball players

See also

Examples

# \donttest{
load_player_stats()
#>       season season_label team_id            team_name conference_short_name
#>        <int>       <char>   <int>               <char>                <char>
#>    1:   2025     20242025       1    Abilene Christian                   WAC
#>    2:   2025     20242025       1    Abilene Christian                   WAC
#>    3:   2025     20242025       1    Abilene Christian                   WAC
#>    4:   2025     20242025       1    Abilene Christian                   WAC
#>    5:   2025     20242025       1    Abilene Christian                   WAC
#>   ---                                                                       
#> 9813:   2025     20242025     830 Washington Adventist                  <NA>
#> 9814:   2025     20242025     830 Washington Adventist                  <NA>
#> 9815:   2025     20242025     830 Washington Adventist                  <NA>
#> 9816:   2025     20242025     830 Washington Adventist                  <NA>
#> 9817:   2025     20242025     830 Washington Adventist                  <NA>
#>       athlete_id athlete_source_id       athlete_name athlete_position games
#>            <int>            <char>             <char>           <char> <int>
#>    1:       6045           4896505   Leonardo Bettiol                F    32
#>    2:       6046           5177704       Nasir DeGruy                G    32
#>    3:       6047           5106274     Quion Williams                G    30
#>    4:       6048           4897572     Joseph Venzant                G     8
#>    5:       6049           4701572 Hunter Jack Madden                G    32
#>   ---                                                                       
#> 9813:      15168           5264286     Emanuel Hylton              ATH     1
#> 9814:      15169           5264285  Jayden Rowe-Perry              ATH     1
#> 9815:      15170           5264300       Mateo Cataid              ATH     1
#> 9816:      15171           5264299        James Quinn              ATH     1
#> 9817:      15172           5264294       Felix Mcghee              ATH     1
#>       starts minutes points turnovers fouls assists steals blocks
#>        <int>   <int>  <int>     <int> <int>   <int>  <int>  <int>
#>    1:     31     828    419        61   111      37     33     13
#>    2:      9     579    117        49    59      42     42      3
#>    3:     30     888    411        77    69      75     35     22
#>    4:      6     157     13         8    17       4      7      1
#>    5:     29     883    285        65    45      68     41      4
#>   ---                                                            
#> 9813:      1      34     15         4     1       1      1      0
#> 9814:      1      29      8         4     4       0      2      0
#> 9815:      0       3      0         1     1       1      0      0
#> 9816:      0      23      4         3     2       3      0      0
#> 9817:      0      34     10         5     1       0      2      0
#>       offensive_rating defensive_rating net_rating porpag usage
#>                  <num>            <num>      <num>  <num> <num>
#>    1:            107.3            103.0        4.3    2.1 0.260
#>    2:             76.6            100.4      -23.8   -0.5 0.163
#>    3:            102.0            103.1       -1.1    1.7 0.262
#>    4:             70.1            110.9      -40.8   -0.4 0.071
#>    5:             98.1            104.6       -6.5    0.9 0.195
#>   ---                                                          
#> 9813:             59.1            127.3      -68.2   -5.4 0.336
#> 9814:             46.8            121.2      -74.4   -5.5 0.281
#> 9815:               NA            130.9         NA     NA 0.148
#> 9816:             75.2            129.7      -54.5   -0.5 0.113
#> 9817:             78.0            124.2      -46.2   -0.8 0.148
#>       assists_turnover_ratio offensive_rebound_pct free_throw_rate
#>                        <num>                 <num>           <num>
#>    1:                   0.61                 0.381           0.409
#>    2:                   0.86                 0.157           0.213
#>    3:                   0.97                 0.327           0.357
#>    4:                   0.50                 0.125           1.200
#>    5:                   1.05                 0.253           0.277
#>   ---                                                             
#> 9813:                   0.25                 0.200           0.200
#> 9814:                   0.00                 0.167           1.000
#> 9815:                   1.00                    NA              NA
#> 9816:                   1.00                 0.250           1.000
#> 9817:                   0.00                 0.000           0.600
#>       effective_field_goal_pct true_shooting_pct field_goals_made
#>                          <num>             <num>            <int>
#>    1:                    0.526             0.006              164
#>    2:                    0.382             0.004               46
#>    3:                    0.492             0.005              155
#>    4:                    0.400             0.004                4
#>    5:                    0.446             0.005               85
#>   ---                                                            
#> 9813:                    0.325             0.003                5
#> 9814:                    0.150             0.003                1
#> 9815:                       NA                NA                0
#> 9816:                    0.750             0.007                1
#> 9817:                    0.700             0.008                3
#>       field_goals_attempted field_goals_pct two_point_field_goals_made
#>                       <int>           <num>                      <int>
#>    1:                   313           0.524                        163
#>    2:                   127           0.362                         41
#>    3:                   333           0.465                        137
#>    4:                    10           0.400                          4
#>    5:                   249           0.341                         33
#>   ---                                                                 
#> 9813:                    20           0.250                          2
#> 9814:                    10           0.100                          0
#> 9815:                     0           0.000                          0
#> 9816:                     2           0.500                          0
#> 9817:                     5           0.600                          2
#>       two_point_field_goals_attempted two_point_field_goals_pct
#>                                 <int>                     <num>
#>    1:                             304                     0.536
#>    2:                              98                     0.418
#>    3:                             279                     0.491
#>    4:                              10                     0.400
#>    5:                              78                     0.423
#>   ---                                                          
#> 9813:                              11                     0.182
#> 9814:                               6                     0.000
#> 9815:                               0                     0.000
#> 9816:                               1                     0.000
#> 9817:                               2                     1.000
#>       three_point_field_goals_made three_point_field_goals_attempted
#>                              <int>                             <int>
#>    1:                            1                                 9
#>    2:                            5                                29
#>    3:                           18                                54
#>    4:                            0                                 0
#>    5:                           52                               171
#>   ---                                                               
#> 9813:                            3                                 9
#> 9814:                            1                                 4
#> 9815:                            0                                 0
#> 9816:                            1                                 1
#> 9817:                            1                                 3
#>       three_point_field_goals_pct free_throws_made free_throws_attempted
#>                             <num>            <int>                 <int>
#>    1:                       0.111               90                   128
#>    2:                       0.172               20                    27
#>    3:                       0.333               83                   119
#>    4:                       0.000                5                    12
#>    5:                       0.304               63                    69
#>   ---                                                                   
#> 9813:                       0.333                2                     4
#> 9814:                       0.250                5                    10
#> 9815:                       0.000                0                     0
#> 9816:                       1.000                1                     2
#> 9817:                       0.333                3                     3
#>       free_throws_pct rebounds_offensive rebounds_defensive rebounds_total
#>                 <num>              <int>              <int>          <int>
#>    1:           0.703                 48                 78            126
#>    2:           0.741                  8                 43             51
#>    3:           0.697                 50                103            153
#>    4:           0.417                  4                 28             32
#>    5:           0.913                 19                 56             75
#>   ---                                                                     
#> 9813:           0.500                  1                  4              5
#> 9814:           0.500                  1                  5              6
#> 9815:           0.000                  0                  0              0
#> 9816:           0.500                  1                  3              4
#> 9817:           1.000                  0                  4              4
#>       win_shares_offensive win_shares_defensive win_shares_total
#>                      <num>                <num>            <num>
#>    1:                  1.5                  1.2              2.7
#>    2:                 -0.7                  1.0              0.3
#>    3:                  1.1                  1.2              2.3
#>    4:                 -0.1                  0.1              0.0
#>    5:                  0.5                  1.1              1.6
#>   ---                                                           
#> 9813:                 -0.2                  0.0             -0.2
#> 9814:                 -0.2                  0.0             -0.2
#> 9815:                   NA                  0.0               NA
#> 9816:                  0.0                  0.0              0.0
#> 9817:                  0.0                  0.0              0.0
#>       win_shares_total_per40
#>                        <num>
#>    1:                  0.130
#>    2:                  0.021
#>    3:                  0.104
#>    4:                  0.000
#>    5:                  0.072
#>   ---                       
#> 9813:                 -0.235
#> 9814:                 -0.276
#> 9815:                     NA
#> 9816:                  0.000
#> 9817:                  0.000

load_player_stats(2024)
#>       season season_label team_id                    team_name
#>        <int>       <char>   <int>                       <char>
#>    1:   2024     20232024       1            Abilene Christian
#>    2:   2024     20232024       1            Abilene Christian
#>    3:   2024     20232024       1            Abilene Christian
#>    4:   2024     20232024       1            Abilene Christian
#>    5:   2024     20232024       1            Abilene Christian
#>   ---                                                         
#> 9872:   2024     20232024     972 Indiana University Northwest
#> 9873:   2024     20232024     972 Indiana University Northwest
#> 9874:   2024     20232024     972 Indiana University Northwest
#> 9875:   2024     20232024     972 Indiana University Northwest
#> 9876:   2024     20232024     972 Indiana University Northwest
#>       conference_short_name athlete_id athlete_source_id       athlete_name
#>                      <char>      <int>            <char>             <char>
#>    1:                   WAC        747           4896862          Ali Dibba
#>    2:                   WAC       4056           5177703     Kavion McClain
#>    3:                   WAC       6045           4896505   Leonardo Bettiol
#>    4:                   WAC       6046           5177704       Nasir DeGruy
#>    5:                   WAC       6049           4701572 Hunter Jack Madden
#>   ---                                                                      
#> 9872:                  <NA>      19824           5192074     Ladaion Barnes
#> 9873:                  <NA>      19825           5192077      Adam Eastland
#> 9874:                  <NA>      19826           5192073       Kionne Layne
#> 9875:                  <NA>      19827           5192070    Brian Hernandez
#> 9876:                  <NA>      19828           5192068      Nicky Flesher
#>       athlete_position games starts minutes points turnovers fouls assists
#>                 <char> <int>  <int>   <int>  <int>     <int> <int>   <int>
#>    1:                G    31     30     985    482        72    65      36
#>    2:                G    34     30     919    211        68    56     115
#>    3:                F    31     22     667    303        41   111      29
#>    4:                G    34      6     483    131        31    37      36
#>    5:                G    34     34    1108    400        54    67      68
#>   ---                                                                     
#> 9872:                F     1      0      21      3         0     4       2
#> 9873:                G     1      0       2      0         0     0       0
#> 9874:                G     1      0      12      2         0     1       1
#> 9875:                G     1      0      16      3         0     4       2
#> 9876:                G     1      0      19      9         0     1       0
#>       steals blocks offensive_rating defensive_rating net_rating porpag usage
#>        <int>  <int>            <num>            <num>      <num>  <num> <num>
#>    1:     43      7            102.5            110.1       -7.6    1.9 0.263
#>    2:     41      2             97.7            112.3      -14.6    0.6 0.145
#>    3:     12      6            111.9            107.8        4.1    1.9 0.227
#>    4:     17      5             93.2            111.5      -18.3    0.2 0.174
#>    5:     39      1            109.6            113.3       -3.7    2.1 0.188
#>   ---                                                                        
#> 9872:      0      0             95.1            122.6      -27.5    0.2 0.098
#> 9873:      0      0             22.1            134.4     -112.3   -0.9 0.413
#> 9874:      1      1             50.6            116.8      -66.2   -1.8 0.241
#> 9875:      1      0             65.7            122.8      -57.1   -1.2 0.207
#> 9876:      0      0            137.7            134.3        3.4    2.4 0.152
#>       assists_turnover_ratio offensive_rebound_pct free_throw_rate
#>                        <num>                 <num>           <num>
#>    1:                   0.50                 0.220           0.431
#>    2:                   1.69                 0.106           0.259
#>    3:                   0.71                 0.397           0.483
#>    4:                   1.16                 0.220           0.316
#>    5:                   1.26                 0.216           0.417
#>   ---                                                             
#> 9872:                     NA                 0.200           0.000
#> 9873:                     NA                 1.000           0.000
#> 9874:                     NA                 0.500           0.000
#> 9875:                     NA                 0.500           0.000
#> 9876:                     NA                    NA           0.000
#>       effective_field_goal_pct true_shooting_pct field_goals_made
#>                          <num>             <num>            <int>
#>    1:                    0.500             0.006              175
#>    2:                    0.494             0.005               78
#>    3:                    0.559             0.006              118
#>    4:                    0.440             0.005               49
#>    5:                    0.498             0.006              118
#>   ---                                                            
#> 9872:                    0.300             0.003                1
#> 9873:                    0.000             0.000                0
#> 9874:                    0.143             0.001                1
#> 9875:                    0.188             0.002                1
#> 9876:                    0.643             0.006                3
#>       field_goals_attempted field_goals_pct two_point_field_goals_made
#>                       <int>           <num>                      <int>
#>    1:                   367           0.477                        158
#>    2:                   174           0.448                         62
#>    3:                   211           0.559                        118
#>    4:                   117           0.419                         44
#>    5:                   300           0.393                         55
#>   ---                                                                 
#> 9872:                     5           0.200                          0
#> 9873:                     2           0.000                          0
#> 9874:                     7           0.143                          1
#> 9875:                     8           0.125                          0
#> 9876:                     7           0.429                          0
#>       two_point_field_goals_attempted two_point_field_goals_pct
#>                                 <int>                     <num>
#>    1:                             318                     0.497
#>    2:                             121                     0.512
#>    3:                             211                     0.559
#>    4:                              98                     0.449
#>    5:                             138                     0.399
#>   ---                                                          
#> 9872:                               2                     0.000
#> 9873:                               1                     0.000
#> 9874:                               6                     0.167
#> 9875:                               2                     0.000
#> 9876:                               1                     0.000
#>       three_point_field_goals_made three_point_field_goals_attempted
#>                              <int>                             <int>
#>    1:                           17                                49
#>    2:                           16                                53
#>    3:                            0                                 0
#>    4:                            5                                19
#>    5:                           63                               162
#>   ---                                                               
#> 9872:                            1                                 3
#> 9873:                            0                                 1
#> 9874:                            0                                 1
#> 9875:                            1                                 6
#> 9876:                            3                                 6
#>       three_point_field_goals_pct free_throws_made free_throws_attempted
#>                             <num>            <int>                 <int>
#>    1:                       0.347              115                   158
#>    2:                       0.302               39                    45
#>    3:                       0.000               67                   102
#>    4:                       0.263               28                    37
#>    5:                       0.389              101                   125
#>   ---                                                                   
#> 9872:                       0.333                0                     0
#> 9873:                       0.000                0                     0
#> 9874:                       0.000                0                     0
#> 9875:                       0.167                0                     0
#> 9876:                       0.500                0                     0
#>       free_throws_pct rebounds_offensive rebounds_defensive rebounds_total
#>                 <num>              <int>              <int>          <int>
#>    1:           0.728                 31                110            141
#>    2:           0.867                  7                 59             66
#>    3:           0.657                 62                 94            156
#>    4:           0.757                 13                 46             59
#>    5:           0.808                 22                 80            102
#>   ---                                                                     
#> 9872:           0.000                  1                  4              5
#> 9873:           0.000                  1                  0              1
#> 9874:           0.000                  1                  1              2
#> 9875:           0.000                  1                  1              2
#> 9876:           0.000                  0                  0              0
#>       win_shares_offensive win_shares_defensive win_shares_total
#>                      <num>                <num>            <num>
#>    1:                  1.4                  0.7              2.1
#>    2:                  0.4                  0.5              0.9
#>    3:                  1.5                  0.6              2.1
#>    4:                  0.1                  0.3              0.4
#>    5:                  1.9                  0.5              2.4
#>   ---                                                           
#> 9872:                  0.0                  0.0              0.0
#> 9873:                  0.0                  0.0              0.0
#> 9874:                  0.0                  0.0              0.0
#> 9875:                  0.0                  0.0              0.0
#> 9876:                  0.1                  0.0              0.1
#>       win_shares_total_per40
#>                        <num>
#>    1:                  0.085
#>    2:                  0.039
#>    3:                  0.126
#>    4:                  0.033
#>    5:                  0.087
#>   ---                       
#> 9872:                  0.000
#> 9873:                  0.000
#> 9874:                  0.000
#> 9875:                  0.000
#> 9876:                  0.211

load_player_stats(TRUE)
#>         season season_label team_id            team_name conference_short_name
#>          <int>       <char>   <int>               <char>                <char>
#>      1:   2003     20022003     102              Gonzaga                   WCC
#>      2:   2003     20022003     102              Gonzaga                   WCC
#>      3:   2003     20022003     102              Gonzaga                   WCC
#>      4:   2003     20022003     102              Gonzaga                   WCC
#>      5:   2003     20022003     102              Gonzaga                   WCC
#>     ---                                                                       
#> 157937:   2025     20242025     830 Washington Adventist                  <NA>
#> 157938:   2025     20242025     830 Washington Adventist                  <NA>
#> 157939:   2025     20242025     830 Washington Adventist                  <NA>
#> 157940:   2025     20242025     830 Washington Adventist                  <NA>
#> 157941:   2025     20242025     830 Washington Adventist                  <NA>
#>         athlete_id athlete_source_id      athlete_name athlete_position games
#>              <int>            <char>            <char>           <char> <int>
#>      1:      89361              8021 Dustin Villepigue                F     1
#>      2:      90765              6585      Ronny Turiaf                F     1
#>      3:      90766              6580  Brian Michaelson                G     1
#>      4:      91830              8020    Joshua Reisman                G     1
#>      5:      92102              3546     Kyle Bankhead              G-F     1
#>     ---                                                                      
#> 157937:      15168           5264286    Emanuel Hylton              ATH     1
#> 157938:      15169           5264285 Jayden Rowe-Perry              ATH     1
#> 157939:      15170           5264300      Mateo Cataid              ATH     1
#> 157940:      15171           5264299       James Quinn              ATH     1
#> 157941:      15172           5264294      Felix Mcghee              ATH     1
#>         starts minutes points turnovers fouls assists steals blocks
#>          <int>   <int>  <int>     <int> <int>   <int>  <int>  <int>
#>      1:      0       1      0         0     0       0      0      0
#>      2:      0      17      8         0     4       0      2      2
#>      3:      0       1      0         0     0       0      0      0
#>      4:      0       5      2         0     0       0      1      0
#>      5:      0      30      5         0     3       5      0      0
#>     ---                                                            
#> 157937:      1      34     15         4     1       1      1      0
#> 157938:      1      29      8         4     4       0      2      0
#> 157939:      0       3      0         1     1       1      0      0
#> 157940:      0      23      4         3     2       3      0      0
#> 157941:      0      34     10         5     1       0      2      0
#>         offensive_rating defensive_rating net_rating porpag usage
#>                    <num>            <num>      <num>  <num> <num>
#>      1:               NA               NA         NA     NA 0.000
#>      2:             89.3               NA         NA    0.1 0.377
#>      3:               NA               NA         NA     NA 0.000
#>      4:            114.5               NA         NA    0.5 0.214
#>      5:            145.2               NA         NA    2.5 0.089
#>     ---                                                          
#> 157937:             59.1            127.3      -68.2   -5.4 0.336
#> 157938:             46.8            121.2      -74.4   -5.5 0.281
#> 157939:               NA            130.9         NA     NA 0.148
#> 157940:             75.2            129.7      -54.5   -0.5 0.113
#> 157941:             78.0            124.2      -46.2   -0.8 0.148
#>         assists_turnover_ratio offensive_rebound_pct free_throw_rate
#>                          <num>                 <num>           <num>
#>      1:                     NA                 0.000              NA
#>      2:                     NA                 0.600             0.0
#>      3:                     NA                    NA              NA
#>      4:                     NA                 0.000             0.0
#>      5:                     NA                 0.000             0.0
#>     ---                                                             
#> 157937:                   0.25                 0.200             0.2
#> 157938:                   0.00                 0.167             1.0
#> 157939:                   1.00                    NA              NA
#> 157940:                   1.00                 0.250             1.0
#> 157941:                   0.00                 0.000             0.6
#>         effective_field_goal_pct true_shooting_pct field_goals_made
#>                            <num>             <num>            <int>
#>      1:                       NA                NA                0
#>      2:                    0.333             0.003                4
#>      3:                       NA                NA                0
#>      4:                    0.500             0.005                1
#>      5:                    0.500             0.005                2
#>     ---                                                            
#> 157937:                    0.325             0.003                5
#> 157938:                    0.150             0.003                1
#> 157939:                       NA                NA                0
#> 157940:                    0.750             0.007                1
#> 157941:                    0.700             0.008                3
#>         field_goals_attempted field_goals_pct two_point_field_goals_made
#>                         <int>           <num>                      <int>
#>      1:                     0           0.000                          0
#>      2:                    12           0.333                          4
#>      3:                     0           0.000                          0
#>      4:                     2           0.500                          1
#>      5:                     5           0.400                          1
#>     ---                                                                 
#> 157937:                    20           0.250                          2
#> 157938:                    10           0.100                          0
#> 157939:                     0           0.000                          0
#> 157940:                     2           0.500                          0
#> 157941:                     5           0.600                          2
#>         two_point_field_goals_attempted two_point_field_goals_pct
#>                                   <int>                     <num>
#>      1:                               0                     0.000
#>      2:                              12                     0.333
#>      3:                               0                     0.000
#>      4:                               2                     0.500
#>      5:                               2                     0.500
#>     ---                                                          
#> 157937:                              11                     0.182
#> 157938:                               6                     0.000
#> 157939:                               0                     0.000
#> 157940:                               1                     0.000
#> 157941:                               2                     1.000
#>         three_point_field_goals_made three_point_field_goals_attempted
#>                                <int>                             <int>
#>      1:                            0                                 0
#>      2:                            0                                 0
#>      3:                            0                                 0
#>      4:                            0                                 0
#>      5:                            1                                 3
#>     ---                                                               
#> 157937:                            3                                 9
#> 157938:                            1                                 4
#> 157939:                            0                                 0
#> 157940:                            1                                 1
#> 157941:                            1                                 3
#>         three_point_field_goals_pct free_throws_made free_throws_attempted
#>                               <num>            <int>                 <int>
#>      1:                       0.000                0                     0
#>      2:                       0.000                0                     0
#>      3:                       0.000                0                     0
#>      4:                       0.000                0                     0
#>      5:                       0.333                0                     0
#>     ---                                                                   
#> 157937:                       0.333                2                     4
#> 157938:                       0.250                5                    10
#> 157939:                       0.000                0                     0
#> 157940:                       1.000                1                     2
#> 157941:                       0.333                3                     3
#>         free_throws_pct rebounds_offensive rebounds_defensive rebounds_total
#>                   <num>              <int>              <int>          <int>
#>      1:             0.0                  0                  1              1
#>      2:             0.0                  3                  2              5
#>      3:             0.0                  0                  0              0
#>      4:             0.0                  0                  1              1
#>      5:             0.0                  0                  3              3
#>     ---                                                                     
#> 157937:             0.5                  1                  4              5
#> 157938:             0.5                  1                  5              6
#> 157939:             0.0                  0                  0              0
#> 157940:             0.5                  1                  3              4
#> 157941:             1.0                  0                  4              4
#>         win_shares_offensive win_shares_defensive win_shares_total
#>                        <num>                <num>            <num>
#>      1:                   NA                   NA               NA
#>      2:                   NA                   NA               NA
#>      3:                   NA                   NA               NA
#>      4:                   NA                   NA               NA
#>      5:                   NA                   NA               NA
#>     ---                                                           
#> 157937:                 -0.2                    0             -0.2
#> 157938:                 -0.2                    0             -0.2
#> 157939:                   NA                    0               NA
#> 157940:                  0.0                    0              0.0
#> 157941:                  0.0                    0              0.0
#>         win_shares_total_per40
#>                          <num>
#>      1:                     NA
#>      2:                     NA
#>      3:                     NA
#>      4:                     NA
#>      5:                     NA
#>     ---                       
#> 157937:                 -0.235
#> 157938:                 -0.276
#> 157939:                     NA
#> 157940:                  0.000
#> 157941:                  0.000
# }