Load cbb play by play data

Grabs entire seasons of play-by-play data from the collegebasketballdata.com API for specified season(s)

Includes a roughly complete record of every tracked play in for the games in the specified seasons.

For many games with clean player substitution data, the identities of the players on the floor are included. Sometimes, substitution data is unclean, and this results in too many or too fewer players listed as being on the floor at a given time.

A primary player is identified for some plays as participants_*_1. For example, in an unassisted shot, the ID of the player taking the shot is participants_id_1. If two players are involved in a play, the initiating player is listed as participants_id_1 and the receiving player is listed as participants_id_2. If a player passes the ball to another player who makes a shot, the assisting player is listed as participants_id_1 and the shooting player is listed as participants_id_2.

Shot location data is provided where available.

The win probability for the home team at the time of each play is provided as win_probability. These are predictions provided by the collegebasketballdata.com API, not built into the package.

Usage

load_plays(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of play-by-play data from all D1 college basketball games.

See also

Examples

# \donttest{
load_plays(2024)
#>          play_id on_floor_athlete_id_1 on_floor_athlete_id_2
#>            <int>                 <int>                 <int>
#>       1: 1160702                    47                  2938
#>       2: 1160703                    47                  2938
#>       3: 1160704                    47                  2938
#>       4: 1160705                    47                  2938
#>       5: 1160706                    47                  2938
#>      ---                                                    
#> 2004993: 3165694                   210                   369
#> 2004994: 3165695                   210                   369
#> 2004995: 3165696                   210                   369
#> 2004996: 3165697                  8146                 15262
#> 2004997: 3165698                  8146                 15262
#>          on_floor_athlete_id_3 on_floor_athlete_id_4 on_floor_athlete_id_5
#>                          <int>                 <int>                 <int>
#>       1:                 15214                    48                 15207
#>       2:                 15214                    48                 15207
#>       3:                 15214                    48                 15207
#>       4:                 15214                    48                 15207
#>       5:                 15214                    48                 15207
#>      ---                                                                  
#> 2004993:                 15235                 15258                  7792
#> 2004994:                 15235                 15258                  7792
#> 2004995:                 15235                 15258                  7792
#> 2004996:                  7797                  7794                  7802
#> 2004997:                  7797                  7794                  7802
#>          on_floor_athlete_id_6 on_floor_athlete_id_7 on_floor_athlete_id_8
#>                          <int>                 <int>                 <int>
#>       1:                    51                    63                 15206
#>       2:                    51                    63                 15206
#>       3:                    51                    63                 15206
#>       4:                    51                    63                 15206
#>       5:                    51                    63                 15206
#>      ---                                                                  
#> 2004993:                 15259                 15260                 15261
#> 2004994:                 15259                 15260                 15261
#> 2004995:                 15259                 15260                 15261
#> 2004996:                  1486                   373                  6789
#> 2004997:                  1486                   373                  6789
#>          on_floor_athlete_id_9 on_floor_athlete_id_10 on_floor_athlete_id_11
#>                          <int>                  <int>                  <int>
#>       1:                 15205                   2940                     NA
#>       2:                 15205                   2940                     NA
#>       3:                 15205                   2940                     NA
#>       4:                 15205                   2940                     NA
#>       5:                 15205                   2940                     NA
#>      ---                                                                    
#> 2004993:                   368                  15234                     NA
#> 2004994:                   368                  15234                     NA
#> 2004995:                   368                  15234                     NA
#> 2004996:                 15236                    210                     NA
#> 2004997:                 15236                    210                     NA
#>          on_floor_athlete_id_12 on_floor_athlete_id_13 on_floor_athlete_id_14
#>                           <int>                  <int>                  <int>
#>       1:                     NA                     NA                     NA
#>       2:                     NA                     NA                     NA
#>       3:                     NA                     NA                     NA
#>       4:                     NA                     NA                     NA
#>       5:                     NA                     NA                     NA
#>      ---                                                                     
#> 2004993:                     NA                     NA                     NA
#> 2004994:                     NA                     NA                     NA
#> 2004995:                     NA                     NA                     NA
#> 2004996:                     NA                     NA                     NA
#> 2004997:                     NA                     NA                     NA
#>          on_floor_athlete_name_1 on_floor_athlete_name_2
#>                           <char>                  <char>
#>       1:            KJ Adams Jr.           Po'Boigh King
#>       2:            KJ Adams Jr.           Po'Boigh King
#>       3:            KJ Adams Jr.           Po'Boigh King
#>       4:            KJ Adams Jr.           Po'Boigh King
#>       5:            KJ Adams Jr.           Po'Boigh King
#>      ---                                                
#> 2004993:            Mason Gillis          Fletcher Loyer
#> 2004994:            Mason Gillis          Fletcher Loyer
#> 2004995:            Mason Gillis          Fletcher Loyer
#> 2004996:     Apostolos Roumoglou           Andrew Hurley
#> 2004997:     Apostolos Roumoglou           Andrew Hurley
#>          on_floor_athlete_name_3 on_floor_athlete_name_4
#>                           <char>                  <char>
#>       1:      Kevin McCullar Jr.        Hunter Dickinson
#>       2:      Kevin McCullar Jr.        Hunter Dickinson
#>       3:      Kevin McCullar Jr.        Hunter Dickinson
#>       4:      Kevin McCullar Jr.        Hunter Dickinson
#>       5:      Kevin McCullar Jr.        Hunter Dickinson
#>      ---                                                
#> 2004993:             Lance Jones         Donovan Clingan
#> 2004994:             Lance Jones         Donovan Clingan
#> 2004995:             Lance Jones         Donovan Clingan
#> 2004996:          Jaylin Stewart               Solo Ball
#> 2004997:          Jaylin Stewart               Solo Ball
#>          on_floor_athlete_name_5 on_floor_athlete_name_6
#>                           <char>                  <char>
#>       1:      Fred Cleveland Jr.       Dajuan Harris Jr.
#>       2:      Fred Cleveland Jr.       Dajuan Harris Jr.
#>       3:      Fred Cleveland Jr.       Dajuan Harris Jr.
#>       4:      Fred Cleveland Jr.       Dajuan Harris Jr.
#>       5:      Fred Cleveland Jr.       Dajuan Harris Jr.
#>      ---                                                
#> 2004993:            Alex Karaban          Stephon Castle
#> 2004994:            Alex Karaban          Stephon Castle
#> 2004995:            Alex Karaban          Stephon Castle
#> 2004996:             Jayden Ross            Chase Martin
#> 2004997:             Jayden Ross            Chase Martin
#>          on_floor_athlete_name_7 on_floor_athlete_name_8
#>                           <char>                  <char>
#>       1:         Elmarko Jackson          Cobey Harraway
#>       2:         Elmarko Jackson          Cobey Harraway
#>       3:         Elmarko Jackson          Cobey Harraway
#>       4:         Elmarko Jackson          Cobey Harraway
#>       5:         Elmarko Jackson          Cobey Harraway
#>      ---                                                
#> 2004993:          Tristen Newton             Cam Spencer
#> 2004994:          Tristen Newton             Cam Spencer
#> 2004995:          Tristen Newton             Cam Spencer
#> 2004996:             Caleb Furst            Ethan Morton
#> 2004997:             Caleb Furst            Ethan Morton
#>          on_floor_athlete_name_9 on_floor_athlete_name_10
#>                           <char>                   <char>
#>       1:       Emmanuel Izunabor           Keishon Porter
#>       2:       Emmanuel Izunabor           Keishon Porter
#>       3:       Emmanuel Izunabor           Keishon Porter
#>       4:       Emmanuel Izunabor           Keishon Porter
#>       5:       Emmanuel Izunabor           Keishon Porter
#>      ---                                                 
#> 2004993:            Braden Smith                Zach Edey
#> 2004994:            Braden Smith                Zach Edey
#> 2004995:            Braden Smith                Zach Edey
#> 2004996:          Carson Barrett             Mason Gillis
#> 2004997:          Carson Barrett             Mason Gillis
#>          on_floor_athlete_name_11 on_floor_athlete_name_12
#>                            <char>                   <char>
#>       1:                     <NA>                     <NA>
#>       2:                     <NA>                     <NA>
#>       3:                     <NA>                     <NA>
#>       4:                     <NA>                     <NA>
#>       5:                     <NA>                     <NA>
#>      ---                                                  
#> 2004993:                     <NA>                     <NA>
#> 2004994:                     <NA>                     <NA>
#> 2004995:                     <NA>                     <NA>
#> 2004996:                     <NA>                     <NA>
#> 2004997:                     <NA>                     <NA>
#>          on_floor_athlete_name_13 on_floor_athlete_name_14 on_floor_team_name_1
#>                            <char>                   <char>               <char>
#>       1:                     <NA>                     <NA>               Kansas
#>       2:                     <NA>                     <NA>               Kansas
#>       3:                     <NA>                     <NA>               Kansas
#>       4:                     <NA>                     <NA>               Kansas
#>       5:                     <NA>                     <NA>               Kansas
#>      ---                                                                       
#> 2004993:                     <NA>                     <NA>               Purdue
#> 2004994:                     <NA>                     <NA>               Purdue
#> 2004995:                     <NA>                     <NA>               Purdue
#> 2004996:                     <NA>                     <NA>                UConn
#> 2004997:                     <NA>                     <NA>                UConn
#>            on_floor_team_name_2 on_floor_team_name_3 on_floor_team_name_4
#>                          <char>               <char>               <char>
#>       1: North Carolina Central               Kansas               Kansas
#>       2: North Carolina Central               Kansas               Kansas
#>       3: North Carolina Central               Kansas               Kansas
#>       4: North Carolina Central               Kansas               Kansas
#>       5: North Carolina Central               Kansas               Kansas
#>      ---                                                                 
#> 2004993:                 Purdue               Purdue                UConn
#> 2004994:                 Purdue               Purdue                UConn
#> 2004995:                 Purdue               Purdue                UConn
#> 2004996:                  UConn                UConn                UConn
#> 2004997:                  UConn                UConn                UConn
#>            on_floor_team_name_5 on_floor_team_name_6 on_floor_team_name_7
#>                          <char>               <char>               <char>
#>       1: North Carolina Central               Kansas               Kansas
#>       2: North Carolina Central               Kansas               Kansas
#>       3: North Carolina Central               Kansas               Kansas
#>       4: North Carolina Central               Kansas               Kansas
#>       5: North Carolina Central               Kansas               Kansas
#>      ---                                                                 
#> 2004993:                  UConn                UConn                UConn
#> 2004994:                  UConn                UConn                UConn
#> 2004995:                  UConn                UConn                UConn
#> 2004996:                  UConn               Purdue               Purdue
#> 2004997:                  UConn               Purdue               Purdue
#>            on_floor_team_name_8   on_floor_team_name_9  on_floor_team_name_10
#>                          <char>                 <char>                 <char>
#>       1: North Carolina Central North Carolina Central North Carolina Central
#>       2: North Carolina Central North Carolina Central North Carolina Central
#>       3: North Carolina Central North Carolina Central North Carolina Central
#>       4: North Carolina Central North Carolina Central North Carolina Central
#>       5: North Carolina Central North Carolina Central North Carolina Central
#>      ---                                                                     
#> 2004993:                  UConn                 Purdue                 Purdue
#> 2004994:                  UConn                 Purdue                 Purdue
#> 2004995:                  UConn                 Purdue                 Purdue
#> 2004996:                 Purdue                 Purdue                 Purdue
#> 2004997:                 Purdue                 Purdue                 Purdue
#>          on_floor_team_name_11 on_floor_team_name_12 on_floor_team_name_13
#>                         <char>                <char>                <char>
#>       1:                  <NA>                  <NA>                  <NA>
#>       2:                  <NA>                  <NA>                  <NA>
#>       3:                  <NA>                  <NA>                  <NA>
#>       4:                  <NA>                  <NA>                  <NA>
#>       5:                  <NA>                  <NA>                  <NA>
#>      ---                                                                  
#> 2004993:                  <NA>                  <NA>                  <NA>
#> 2004994:                  <NA>                  <NA>                  <NA>
#> 2004995:                  <NA>                  <NA>                  <NA>
#> 2004996:                  <NA>                  <NA>                  <NA>
#> 2004997:                  <NA>                  <NA>                  <NA>
#>          on_floor_team_name_14 shooter_athlete_id shooter_athlete_name
#>                         <char>              <int>               <char>
#>       1:                  <NA>              15214   Kevin McCullar Jr.
#>       2:                  <NA>                 NA                 <NA>
#>       3:                  <NA>                 48     Hunter Dickinson
#>       4:                  <NA>                 NA                 <NA>
#>       5:                  <NA>                 48     Hunter Dickinson
#>      ---                                                              
#> 2004993:                  <NA>              15258      Donovan Clingan
#> 2004994:                  <NA>                368         Braden Smith
#> 2004995:                  <NA>                 NA                 <NA>
#> 2004996:                  <NA>                 NA                 <NA>
#> 2004997:                  <NA>                 NA                 <NA>
#>          shot_made    shot_range shot_assisted assist_athlete_id
#>             <lgcl>        <char>        <lgcl>             <int>
#>       1:     FALSE three_pointer         FALSE                NA
#>       2:        NA          <NA>            NA                NA
#>       3:     FALSE           rim         FALSE                NA
#>       4:        NA          <NA>            NA                NA
#>       5:      TRUE           rim         FALSE                NA
#>      ---                                                        
#> 2004993:      TRUE           rim          TRUE              7792
#> 2004994:     FALSE three_pointer         FALSE                NA
#> 2004995:        NA          <NA>            NA                NA
#> 2004996:        NA          <NA>            NA                NA
#> 2004997:        NA          <NA>            NA                NA
#>          assist_athlete_name shot_x_coordinate shot_y_coordinate
#>                       <char>             <num>             <int>
#>       1:                <NA>             780.2               465
#>       2:                <NA>                NA                NA
#>       3:                <NA>             883.6               250
#>       4:                <NA>                NA                NA
#>       5:                <NA>             883.6               250
#>      ---                                                        
#> 2004993:        Alex Karaban                NA                NA
#> 2004994:                <NA>                NA                NA
#> 2004995:                <NA>                NA                NA
#> 2004996:                <NA>                NA                NA
#> 2004997:                <NA>                NA                NA
#>          primary_participant_athlete_id secondary_participant_athlete_id
#>                                   <int>                            <int>
#>       1:                          15214                               NA
#>       2:                             48                               NA
#>       3:                             48                               NA
#>       4:                             48                               NA
#>       5:                             48                               NA
#>      ---                                                                
#> 2004993:                           7792                            15258
#> 2004994:                            368                               NA
#> 2004995:                             NA                               NA
#> 2004996:                             NA                               NA
#> 2004997:                             NA                               NA
#>          primary_participant_athlete_name secondary_participant_athlete_name
#>                                    <char>                             <char>
#>       1:               Kevin McCullar Jr.                               <NA>
#>       2:                 Hunter Dickinson                               <NA>
#>       3:                 Hunter Dickinson                               <NA>
#>       4:                 Hunter Dickinson                               <NA>
#>       5:                 Hunter Dickinson                               <NA>
#>      ---                                                                    
#> 2004993:                     Alex Karaban                    Donovan Clingan
#> 2004994:                     Braden Smith                               <NA>
#> 2004995:                             <NA>                               <NA>
#> 2004996:                             <NA>                               <NA>
#> 2004997:                             <NA>                               <NA>
#>          game_id game_source_id          game_start_date season season_type
#>            <int>         <char>                   <char>  <int>      <char>
#>       1:    5880      401575451 2023-11-06T20:00:00.000Z   2024     regular
#>       2:    5880      401575451 2023-11-06T20:00:00.000Z   2024     regular
#>       3:    5880      401575451 2023-11-06T20:00:00.000Z   2024     regular
#>       4:    5880      401575451 2023-11-06T20:00:00.000Z   2024     regular
#>       5:    5880      401575451 2023-11-06T20:00:00.000Z   2024     regular
#>      ---                                                                   
#> 2004993:   12128      401638645 2024-04-08T21:20:00.000Z   2024  postseason
#> 2004994:   12128      401638645 2024-04-08T21:20:00.000Z   2024  postseason
#> 2004995:   12128      401638645 2024-04-08T21:20:00.000Z   2024  postseason
#> 2004996:   12128      401638645 2024-04-08T21:20:00.000Z   2024  postseason
#> 2004997:   12128      401638645 2024-04-08T21:20:00.000Z   2024  postseason
#>          game_type tournament     play_source_id          play_type
#>             <char>     <char>             <char>             <char>
#>       1:       STD       <NA> 401575451101805401           JumpShot
#>       2:       STD       <NA> 401575451101806001  Offensive Rebound
#>       3:       STD       <NA> 401575451101806002          LayUpShot
#>       4:       STD       <NA> 401575451101806201  Offensive Rebound
#>       5:       STD       <NA> 401575451101806202          LayUpShot
#>      ---                                                           
#> 2004993:    TRNMNT       NCAA 401638645102995401          LayUpShot
#> 2004994:    TRNMNT       NCAA 401638645102996201           JumpShot
#> 2004995:    TRNMNT       NCAA 401638645102996202  Defensive Rebound
#> 2004996:    TRNMNT       NCAA 401638645102999301 Lost Ball Turnover
#> 2004997:    TRNMNT       NCAA 401638645102999901           End Game
#>          is_home_team team_id team_name conference_short_name team_seed
#>                <lgcl>   <int>    <char>                <char>     <int>
#>       1:         TRUE     131    Kansas                Big 12        NA
#>       2:         TRUE     131    Kansas                Big 12        NA
#>       3:         TRUE     131    Kansas                Big 12        NA
#>       4:         TRUE     131    Kansas                Big 12        NA
#>       5:         TRUE     131    Kansas                Big 12        NA
#>      ---                                                               
#> 2004993:         TRUE     314     UConn              Big East         1
#> 2004994:        FALSE     236    Purdue               Big Ten         1
#> 2004995:         TRUE     314     UConn              Big East         1
#> 2004996:         TRUE     314     UConn              Big East         1
#> 2004997:           NA      NA      <NA>                  <NA>        NA
#>          opponent_team_id          opponent_name opponent_conference_short_name
#>                     <int>                 <char>                         <char>
#>       1:              199 North Carolina Central                           MEAC
#>       2:              199 North Carolina Central                           MEAC
#>       3:              199 North Carolina Central                           MEAC
#>       4:              199 North Carolina Central                           MEAC
#>       5:              199 North Carolina Central                           MEAC
#>      ---                                                                       
#> 2004993:              236                 Purdue                        Big Ten
#> 2004994:              314                  UConn                       Big East
#> 2004995:              236                 Purdue                        Big Ten
#> 2004996:              236                 Purdue                        Big Ten
#> 2004997:               NA                   <NA>                           <NA>
#>          opponent_seed home_score away_score home_win_probability period  clock
#>                  <int>      <int>      <int>                <num>  <int> <char>
#>       1:            NA          0          0                0.641      1  19:45
#>       2:            NA          0          0                0.635      1  19:39
#>       3:            NA          0          0                0.635      1  19:39
#>       4:            NA          0          0                0.635      1  19:37
#>       5:            NA          2          0                0.694      1  19:37
#>      ---                                                                       
#> 2004993:             1         75         60                1.000      2   0:45
#> 2004994:             1         75         60                1.000      2   0:37
#> 2004995:             1         75         60                1.000      2   0:37
#> 2004996:             1         75         60                1.000      2   0:06
#> 2004997:            NA         75         60                1.000      2   0:00
#>          seconds_remaining scoring_play shooting_play score_value
#>                      <int>       <lgcl>        <lgcl>       <int>
#>       1:              1185        FALSE          TRUE           3
#>       2:              1179        FALSE         FALSE           0
#>       3:              1179        FALSE          TRUE           2
#>       4:              1177        FALSE         FALSE           0
#>       5:              1177         TRUE          TRUE           2
#>      ---                                                         
#> 2004993:                45         TRUE          TRUE           2
#> 2004994:                37        FALSE          TRUE           3
#> 2004995:                37        FALSE         FALSE           0
#> 2004996:                 6        FALSE         FALSE           0
#> 2004997:                 0        FALSE         FALSE           0
#>                        wall_clock
#>                            <char>
#>       1: 2023-11-06T20:03:19.000Z
#>       2: 2023-11-06T20:03:19.000Z
#>       3: 2023-11-06T20:03:19.000Z
#>       4: 2023-11-06T20:03:19.000Z
#>       5: 2023-11-06T20:03:19.000Z
#>      ---                         
#> 2004993: 2024-04-08T23:28:07.000Z
#> 2004994: 2024-04-08T23:28:17.000Z
#> 2004995: 2024-04-08T23:28:17.000Z
#> 2004996: 2024-04-08T23:29:58.000Z
#> 2004997: 2024-04-08T23:29:58.000Z
#>                                                      play_text
#>                                                         <char>
#>       1:         Kevin McCullar Jr. missed Three Point Jumper.
#>       2:                   Hunter Dickinson Offensive Rebound.
#>       3:                        Hunter Dickinson missed Layup.
#>       4:                   Hunter Dickinson Offensive Rebound.
#>       5:                          Hunter Dickinson made Layup.
#>      ---                                                      
#> 2004993: Donovan Clingan made Layup. Assisted by Alex Karaban.
#> 2004994:               Braden Smith missed Three Point Jumper.
#> 2004995:                              UConn Defensive Rebound.
#> 2004996:                                      UConn  Turnover.
#> 2004997:                                           End of Game
# }