Load cbb play by play data

Grabs entire seasons of play-by-play data from the collegebasketballdata.com API for specified season(s)

Includes a roughly complete record of every tracked play in for the games in the specified seasons.

For many games with clean player substitution data, the identities of the players on the floor are included. Sometimes, substitution data is unclean, and this results in too many or too fewer players listed as being on the floor at a given time.

A primary player is identified for some plays as participants_*_1. For example, in an unassisted shot, the ID of the player taking the shot is participants_id_1. If two players are involved in a play, the initiating player is listed as participants_id_1 and the receiving player is listed as participants_id_2. If a player passes the ball to another player who makes a shot, the assisting player is listed as participants_id_1 and the shooting player is listed as participants_id_2.

Shot location data is provided where available. build vignette showing how to use shot location data

The win probability for the home team at the time of each play is provided as win_probability. These are predictions provided by the collegebasketballdata.com API, not built into the package.

Usage

load_plays(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of play-by-play data from all D1 college basketball games.

See also

Examples

# \donttest{
load_plays(2024)
#>          play_id on_floor_id_1 on_floor_id_2 on_floor_id_3 on_floor_id_4
#>            <int>         <int>         <int>         <int>         <int>
#>       1:     322         15452          3352         10985         15451
#>       2:     327         15452          3352         10985         15451
#>       3:     328         15452          3352         10985         15451
#>       4:     329         15452          3352         10985         15451
#>       5:     331         15452          3352         10985         15451
#>      ---                                                                
#> 2016357: 1904703            NA            NA            NA            NA
#> 2016358: 1904707            NA            NA            NA            NA
#> 2016359: 1904709            NA            NA            NA            NA
#> 2016360: 1904715            NA            NA            NA            NA
#> 2016361: 1904716            NA            NA            NA            NA
#>          on_floor_id_5 on_floor_id_6 on_floor_id_7 on_floor_id_8 on_floor_id_9
#>                  <int>         <int>         <int>         <int>         <int>
#>       1:         15453         10986           316         11008         10988
#>       2:         15453         10986           316         11008         10988
#>       3:         15453         10986           316         11008         10988
#>       4:         15453         10986           316         11008         10988
#>       5:         15453         10986           316         11008         10988
#>      ---                                                                      
#> 2016357:            NA            NA            NA            NA            NA
#> 2016358:            NA            NA            NA            NA            NA
#> 2016359:            NA            NA            NA            NA            NA
#> 2016360:            NA            NA            NA            NA            NA
#> 2016361:            NA            NA            NA            NA            NA
#>          on_floor_id_10 on_floor_id_11 on_floor_id_12 on_floor_id_13
#>                   <int>          <int>          <int>          <int>
#>       1:          11007             NA             NA             NA
#>       2:          11007             NA             NA             NA
#>       3:          11007             NA             NA             NA
#>       4:          11007             NA             NA             NA
#>       5:          11007             NA             NA             NA
#>      ---                                                            
#> 2016357:             NA             NA             NA             NA
#> 2016358:             NA             NA             NA             NA
#> 2016359:             NA             NA             NA             NA
#> 2016360:             NA             NA             NA             NA
#> 2016361:             NA             NA             NA             NA
#>          on_floor_id_14 on_floor_name_1 on_floor_name_2 on_floor_name_3
#>                   <int>          <char>          <char>          <char>
#>       1:             NA    Tyler Thomas   Jaquan Carlos  Spencer Malloy
#>       2:             NA    Tyler Thomas   Jaquan Carlos  Spencer Malloy
#>       3:             NA    Tyler Thomas   Jaquan Carlos  Spencer Malloy
#>       4:             NA    Tyler Thomas   Jaquan Carlos  Spencer Malloy
#>       5:             NA    Tyler Thomas   Jaquan Carlos  Spencer Malloy
#>      ---                                                               
#> 2016357:             NA            <NA>            <NA>            <NA>
#> 2016358:             NA            <NA>            <NA>            <NA>
#> 2016359:             NA            <NA>            <NA>            <NA>
#> 2016360:             NA            <NA>            <NA>            <NA>
#> 2016361:             NA            <NA>            <NA>            <NA>
#>          on_floor_name_4  on_floor_name_5 on_floor_name_6   on_floor_name_7
#>                   <char>           <char>          <char>            <char>
#>       1:     Jacco Fritz Bryce Washington     Alec Tabada Darlinstone Dubar
#>       2:     Jacco Fritz Bryce Washington     Alec Tabada Darlinstone Dubar
#>       3:     Jacco Fritz Bryce Washington     Alec Tabada Darlinstone Dubar
#>       4:     Jacco Fritz Bryce Washington     Alec Tabada Darlinstone Dubar
#>       5:     Jacco Fritz Bryce Washington     Alec Tabada Darlinstone Dubar
#>      ---                                                                   
#> 2016357:            <NA>             <NA>            <NA>              <NA>
#> 2016358:            <NA>             <NA>            <NA>              <NA>
#> 2016359:            <NA>             <NA>            <NA>              <NA>
#> 2016360:            <NA>             <NA>            <NA>              <NA>
#> 2016361:            <NA>             <NA>            <NA>              <NA>
#>          on_floor_name_8 on_floor_name_9 on_floor_name_10 on_floor_name_11
#>                   <char>          <char>           <char>           <char>
#>       1:      Terrin Roy   Justin Turner     Xavier Allen             <NA>
#>       2:      Terrin Roy   Justin Turner     Xavier Allen             <NA>
#>       3:      Terrin Roy   Justin Turner     Xavier Allen             <NA>
#>       4:      Terrin Roy   Justin Turner     Xavier Allen             <NA>
#>       5:      Terrin Roy   Justin Turner     Xavier Allen             <NA>
#>      ---                                                                  
#> 2016357:            <NA>            <NA>             <NA>             <NA>
#> 2016358:            <NA>            <NA>             <NA>             <NA>
#> 2016359:            <NA>            <NA>             <NA>             <NA>
#> 2016360:            <NA>            <NA>             <NA>             <NA>
#> 2016361:            <NA>            <NA>             <NA>             <NA>
#>          on_floor_name_12 on_floor_name_13 on_floor_name_14 on_floor_team_1
#>                    <char>           <char>           <char>          <char>
#>       1:             <NA>             <NA>             <NA>         Hofstra
#>       2:             <NA>             <NA>             <NA>         Hofstra
#>       3:             <NA>             <NA>             <NA>         Hofstra
#>       4:             <NA>             <NA>             <NA>         Hofstra
#>       5:             <NA>             <NA>             <NA>         Hofstra
#>      ---                                                                   
#> 2016357:             <NA>             <NA>             <NA>            <NA>
#> 2016358:             <NA>             <NA>             <NA>            <NA>
#> 2016359:             <NA>             <NA>             <NA>            <NA>
#> 2016360:             <NA>             <NA>             <NA>            <NA>
#> 2016361:             <NA>             <NA>             <NA>            <NA>
#>          on_floor_team_2            on_floor_team_3 on_floor_team_4
#>                   <char>                     <char>          <char>
#>       1:         Hofstra Saint Joseph's Long Island         Hofstra
#>       2:         Hofstra Saint Joseph's Long Island         Hofstra
#>       3:         Hofstra Saint Joseph's Long Island         Hofstra
#>       4:         Hofstra Saint Joseph's Long Island         Hofstra
#>       5:         Hofstra Saint Joseph's Long Island         Hofstra
#>      ---                                                           
#> 2016357:            <NA>                       <NA>            <NA>
#> 2016358:            <NA>                       <NA>            <NA>
#> 2016359:            <NA>                       <NA>            <NA>
#> 2016360:            <NA>                       <NA>            <NA>
#> 2016361:            <NA>                       <NA>            <NA>
#>          on_floor_team_5            on_floor_team_6 on_floor_team_7
#>                   <char>                     <char>          <char>
#>       1:         Hofstra Saint Joseph's Long Island         Hofstra
#>       2:         Hofstra Saint Joseph's Long Island         Hofstra
#>       3:         Hofstra Saint Joseph's Long Island         Hofstra
#>       4:         Hofstra Saint Joseph's Long Island         Hofstra
#>       5:         Hofstra Saint Joseph's Long Island         Hofstra
#>      ---                                                           
#> 2016357:            <NA>                       <NA>            <NA>
#> 2016358:            <NA>                       <NA>            <NA>
#> 2016359:            <NA>                       <NA>            <NA>
#> 2016360:            <NA>                       <NA>            <NA>
#> 2016361:            <NA>                       <NA>            <NA>
#>                     on_floor_team_8            on_floor_team_9
#>                              <char>                     <char>
#>       1: Saint Joseph's Long Island Saint Joseph's Long Island
#>       2: Saint Joseph's Long Island Saint Joseph's Long Island
#>       3: Saint Joseph's Long Island Saint Joseph's Long Island
#>       4: Saint Joseph's Long Island Saint Joseph's Long Island
#>       5: Saint Joseph's Long Island Saint Joseph's Long Island
#>      ---                                                      
#> 2016357:                       <NA>                       <NA>
#> 2016358:                       <NA>                       <NA>
#> 2016359:                       <NA>                       <NA>
#> 2016360:                       <NA>                       <NA>
#> 2016361:                       <NA>                       <NA>
#>                    on_floor_team_10 on_floor_team_11 on_floor_team_12
#>                              <char>           <char>           <char>
#>       1: Saint Joseph's Long Island             <NA>             <NA>
#>       2: Saint Joseph's Long Island             <NA>             <NA>
#>       3: Saint Joseph's Long Island             <NA>             <NA>
#>       4: Saint Joseph's Long Island             <NA>             <NA>
#>       5: Saint Joseph's Long Island             <NA>             <NA>
#>      ---                                                             
#> 2016357:                       <NA>             <NA>             <NA>
#> 2016358:                       <NA>             <NA>             <NA>
#> 2016359:                       <NA>             <NA>             <NA>
#> 2016360:                       <NA>             <NA>             <NA>
#> 2016361:                       <NA>             <NA>             <NA>
#>          on_floor_team_13 on_floor_team_14 shot_info_shooter_id
#>                    <char>           <char>                <int>
#>       1:             <NA>             <NA>                15452
#>       2:             <NA>             <NA>                 3352
#>       3:             <NA>             <NA>                11008
#>       4:             <NA>             <NA>                15453
#>       5:             <NA>             <NA>                11007
#>      ---                                                       
#> 2016357:             <NA>             <NA>                   NA
#> 2016358:             <NA>             <NA>                   NA
#> 2016359:             <NA>             <NA>                   NA
#> 2016360:             <NA>             <NA>                   NA
#> 2016361:             <NA>             <NA>                   NA
#>          shot_info_shooter_name shot_info_made shot_info_range
#>                          <char>         <lgcl>          <char>
#>       1:           Tyler Thomas          FALSE   three_pointer
#>       2:          Jaquan Carlos           TRUE   three_pointer
#>       3:             Terrin Roy           TRUE             rim
#>       4:       Bryce Washington          FALSE   three_pointer
#>       5:           Xavier Allen          FALSE          jumper
#>      ---                                                      
#> 2016357:                   <NA>             NA            <NA>
#> 2016358:                   <NA>             NA            <NA>
#> 2016359:                   <NA>             NA            <NA>
#> 2016360:                   <NA>             NA            <NA>
#> 2016361:                   <NA>             NA            <NA>
#>          shot_info_assisted shot_info_assisted_by_id shot_info_assisted_by_name
#>                      <lgcl>                    <int>                     <char>
#>       1:              FALSE                       NA                       <NA>
#>       2:               TRUE                    15452               Tyler Thomas
#>       3:               TRUE                    10985             Spencer Malloy
#>       4:              FALSE                       NA                       <NA>
#>       5:              FALSE                       NA                       <NA>
#>      ---                                                                       
#> 2016357:                 NA                       NA                       <NA>
#> 2016358:                 NA                       NA                       <NA>
#> 2016359:                 NA                       NA                       <NA>
#> 2016360:                 NA                       NA                       <NA>
#> 2016361:                 NA                       NA                       <NA>
#>          shot_info_location_x shot_info_location_y participants_id_1
#>                         <num>                <int>             <int>
#>       1:                667.4                  130             15452
#>       2:                639.2                  260             15452
#>       3:                   NA                   NA             11008
#>       4:                902.4                   25             15453
#>       5:                   NA                   NA             11007
#>      ---                                                            
#> 2016357:                   NA                   NA                NA
#> 2016358:                   NA                   NA                NA
#> 2016359:                   NA                   NA                NA
#> 2016360:                   NA                   NA                NA
#> 2016361:                   NA                   NA                NA
#>          participants_id_2 participants_name_1 participants_name_2 game_id
#>                      <int>              <char>              <char>   <int>
#>       1:                NA        Tyler Thomas                <NA>    5902
#>       2:              3352        Tyler Thomas       Jaquan Carlos    5902
#>       3:             10985          Terrin Roy      Spencer Malloy    5902
#>       4:                NA    Bryce Washington                <NA>    5902
#>       5:                NA        Xavier Allen                <NA>    5902
#>      ---                                                                  
#> 2016357:                NA                <NA>                <NA>   11785
#> 2016358:                NA                <NA>                <NA>   11785
#> 2016359:                NA                <NA>                <NA>   11785
#> 2016360:                NA                <NA>                <NA>   11785
#> 2016361:                NA                <NA>                <NA>   11785
#>          game_source_id          game_start_date season season_type game_type
#>                  <char>                   <char>  <int>      <char>    <char>
#>       1:      401610539 2023-11-06T11:30:00.000Z   2024     regular       STD
#>       2:      401610539 2023-11-06T11:30:00.000Z   2024     regular       STD
#>       3:      401610539 2023-11-06T11:30:00.000Z   2024     regular       STD
#>       4:      401610539 2023-11-06T11:30:00.000Z   2024     regular       STD
#>       5:      401610539 2023-11-06T11:30:00.000Z   2024     regular       STD
#>      ---                                                                     
#> 2016357:      401625646 2024-03-10T00:54:00.000Z   2024     regular    TRNMNT
#> 2016358:      401625646 2024-03-10T00:54:00.000Z   2024     regular    TRNMNT
#> 2016359:      401625646 2024-03-10T00:54:00.000Z   2024     regular    TRNMNT
#> 2016360:      401625646 2024-03-10T00:54:00.000Z   2024     regular    TRNMNT
#> 2016361:      401625646 2024-03-10T00:54:00.000Z   2024     regular    TRNMNT
#>          tournament      id          source_id          play_type is_home_team
#>              <char>   <int>             <char>             <char>       <lgcl>
#>       1:       <NA> 1168313 401610539101804601           JumpShot         TRUE
#>       2:       <NA> 1168318 401610539101808501           JumpShot         TRUE
#>       3:       <NA> 1168319 401610539101814801          LayUpShot        FALSE
#>       4:       <NA> 1168320 401610539101816401           JumpShot         TRUE
#>       5:       <NA> 1168322 401610539101818701           JumpShot        FALSE
#>      ---                                                                      
#> 2016357:       <NA> 3054696 401625646102978801 Lost Ball Turnover         TRUE
#> 2016358:       <NA> 3054700 401625646102989001  Offensive Rebound        FALSE
#> 2016359:       <NA> 3054702 401625646102989601  Defensive Rebound         TRUE
#> 2016360:       <NA> 3054708 401625646102999901         End Period           NA
#> 2016361:       <NA> 3054709 401625646102999902           End Game           NA
#>          team_id                       team conference team_seed opponent_id
#>            <int>                     <char>     <char>     <int>       <int>
#>       1:     110                    Hofstra        CAA        NA         677
#>       2:     110                    Hofstra        CAA        NA         677
#>       3:     677 Saint Joseph's Long Island       <NA>        NA         110
#>       4:     110                    Hofstra        CAA        NA         677
#>       5:     677 Saint Joseph's Long Island       <NA>        NA         110
#>      ---                                                                    
#> 2016357:     259              San Francisco        WCC        NA         230
#> 2016358:     230                   Portland        WCC        NA         259
#> 2016359:     259              San Francisco        WCC        NA         230
#> 2016360:      NA                       <NA>       <NA>        NA          NA
#> 2016361:      NA                       <NA>       <NA>        NA          NA
#>                            opponent opponent_conference opponent_seed
#>                              <char>              <char>         <int>
#>       1: Saint Joseph's Long Island                <NA>            NA
#>       2: Saint Joseph's Long Island                <NA>            NA
#>       3:                    Hofstra                 CAA            NA
#>       4: Saint Joseph's Long Island                <NA>            NA
#>       5:                    Hofstra                 CAA            NA
#>      ---                                                             
#> 2016357:                   Portland                 WCC            NA
#> 2016358:              San Francisco                 WCC            NA
#> 2016359:                   Portland                 WCC            NA
#> 2016360:                       <NA>                <NA>            NA
#> 2016361:                       <NA>                <NA>            NA
#>          home_score away_score home_win_probability period  clock
#>               <int>      <int>                <num>  <int> <char>
#>       1:          0          0                0.641      1  19:53
#>       2:          3          0                0.729      1  19:14
#>       3:          3          2                0.669      1  18:51
#>       4:          3          2                0.666      1  18:35
#>       5:          3          2                0.666      1  18:12
#>      ---                                                         
#> 2016357:         70         48                1.000      2   2:11
#> 2016358:         72         51                1.000      2   1:09
#> 2016359:         72         51                1.000      2   1:03
#> 2016360:         72         51                1.000      2   0:00
#> 2016361:         72         51                1.000      2   0:00
#>          seconds_remaining scoring_play shooting_play score_value
#>                      <int>       <lgcl>        <lgcl>       <int>
#>       1:              1193        FALSE          TRUE           3
#>       2:              1154         TRUE          TRUE           3
#>       3:              1131         TRUE          TRUE           2
#>       4:              1115        FALSE          TRUE           3
#>       5:              1092        FALSE          TRUE           2
#>      ---                                                         
#> 2016357:               131        FALSE         FALSE           0
#> 2016358:                69        FALSE         FALSE           0
#> 2016359:                63        FALSE         FALSE           0
#> 2016360:                 0        FALSE         FALSE           0
#> 2016361:                 0        FALSE         FALSE           0
#>                         wallclock
#>                            <char>
#>       1: 2023-11-06T11:33:22.000Z
#>       2: 2023-11-06T11:34:10.000Z
#>       3: 2023-11-06T11:34:33.000Z
#>       4: 2023-11-06T11:34:47.000Z
#>       5: 2023-11-06T11:35:13.000Z
#>      ---                         
#> 2016357: 2024-03-10T03:32:56.000Z
#> 2016358: 2024-03-10T03:34:21.000Z
#> 2016359: 2024-03-10T03:34:40.000Z
#> 2016360: 2024-03-10T03:35:47.000Z
#> 2016361: 2024-03-10T03:35:47.000Z
#>                                                                 play_text
#>                                                                    <char>
#>       1:                          Tyler Thomas missed Three Point Jumper.
#>       2: Jaquan Carlos made Three Point Jumper. Assisted by Tyler Thomas.
#>       3:               Terrin Roy made Layup. Assisted by Spencer Malloy.
#>       4:                      Bryce Washington missed Three Point Jumper.
#>       5:                                      Xavier Allen missed Jumper.
#>      ---                                                                 
#> 2016357:                                         San Francisco  Turnover.
#> 2016358:                                      Portland Offensive Rebound.
#> 2016359:                                 San Francisco Defensive Rebound.
#> 2016360:                                                  End of 2nd half
#> 2016361:                                                      End of Game
# }