Load cbb team rosters

Pull in rosters for college basketball teams for specified season(s)

Usage

load_rosters(seasons = most_recent_season())

Arguments

seasons

an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of rosters for college basketball teams in the specified season(s).

See also

Examples

# \donttest{
try({ # avoid cran errors
load_rosters()
})
#>        team_id team_source_id                team conference_short_name season
#>          <int>         <char>              <char>                <char>  <int>
#>     1:     161            113       Massachusetts                  A-10   2025
#>     2:     161            113       Massachusetts                  A-10   2025
#>     3:     161            113       Massachusetts                  A-10   2025
#>     4:     161            113       Massachusetts                  A-10   2025
#>     5:     161            113       Massachusetts                  A-10   2025
#>    ---                                                                        
#> 15233:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#> 15234:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#> 15235:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#> 15236:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#> 15237:     616           2985 Concordia-Ann Arbor                  <NA>   2025
#>        athlete_id athlete_source_id           athlete_name athlete_first_name
#>             <int>            <char>                 <char>             <char>
#>     1:       2882           5176382       Malek Abdelgowad              Malek
#>     2:       2883           5105982          Daniel Rivera             Daniel
#>     3:       2884           5105551 Daniel Hankins-Sanford             Daniel
#>     4:       2885           5174611           Jaylen Curry             Jaylen
#>     5:       2886           4683733        Rahsool Diggins            Rahsool
#>    ---                                                                       
#> 15233:       9314           4907749     Tarek Abdel-Kireem              Tarek
#> 15234:       9315           4907748           Abeal Ayalew              Abeal
#> 15235:       9316           4907745            Matt Kohler               Matt
#> 15236:       9317           4907744            Devyn Jones              Devyn
#> 15237:       9318           4907742          Jayden Priddy             Jayden
#>        athlete_last_name athlete_jersey_number athlete_position
#>                   <char>                <char>           <char>
#>     1:        Abdelgowad                    25          Forward
#>     2:            Rivera                     5          Forward
#>     3:   Hankins-Sanford                     1          Forward
#>     4:             Curry                     0            Guard
#>     5:           Diggins                     7            Guard
#>    ---                                                         
#> 15233:      Abdel-Kireem                  <NA>          Athlete
#> 15234:            Ayalew                  <NA>          Athlete
#> 15235:            Kohler                  <NA>          Athlete
#> 15236:             Jones                  <NA>          Athlete
#> 15237:            Priddy                  <NA>          Athlete
#>        athlete_height_inches athlete_weight_pounds athlete_hometown_city
#>                        <int>                 <int>                <char>
#>     1:                    82                   220                 Cairo
#>     2:                    78                   215              San Juan
#>     3:                    80                   236             Charlotte
#>     4:                    72                   168             Charlotte
#>     5:                    74                   190          Philadelphia
#>    ---                                                                  
#> 15233:                    NA                    NA                  <NA>
#> 15234:                    NA                    NA                  <NA>
#> 15235:                    NA                    NA                  <NA>
#> 15236:                    NA                    NA                  <NA>
#> 15237:                    NA                    NA                  <NA>
#>        athlete_hometown_state athlete_hometown_country
#>                        <char>                   <lgcl>
#>     1:                  Egypt                       NA
#>     2:            Puerto Rico                       NA
#>     3:                     NC                       NA
#>     4:                     NC                       NA
#>     5:                     PA                       NA
#>    ---                                                
#> 15233:                   <NA>                       NA
#> 15234:                   <NA>                       NA
#> 15235:                   <NA>                       NA
#> 15236:                   <NA>                       NA
#> 15237:                   <NA>                       NA
#>        athlete_hometown_latitude athlete_hometown_longitude
#>                            <num>                      <num>
#>     1:                        NA                         NA
#>     2:                        NA                         NA
#>     3:                  35.22721                  -80.84308
#>     4:                  35.22721                  -80.84308
#>     5:                  39.95272                  -75.16353
#>    ---                                                     
#> 15233:                        NA                         NA
#> 15234:                        NA                         NA
#> 15235:                        NA                         NA
#> 15236:                        NA                         NA
#> 15237:                        NA                         NA
#>        athlete_hometown_county_fips athlete_date_of_birth athlete_start_season
#>                              <char>                <char>                <int>
#>     1:                         <NA>                  <NA>                 2025
#>     2:                         <NA>                  <NA>                 2025
#>     3:                        37119                  <NA>                 2024
#>     4:                        37119                  <NA>                 2024
#>     5:                        42101                  <NA>                 2023
#>    ---                                                                        
#> 15233:                         <NA>                  <NA>                 2023
#> 15234:                         <NA>                  <NA>                 2022
#> 15235:                         <NA>                  <NA>                 2022
#> 15236:                         <NA>                  <NA>                 2022
#> 15237:                         <NA>                  <NA>                 2023
#>        athlete_end_season
#>                     <int>
#>     1:               2025
#>     2:               2025
#>     3:               2025
#>     4:               2025
#>     5:               2025
#>    ---                   
#> 15233:               2025
#> 15234:               2025
#> 15235:               2025
#> 15236:               2025
#> 15237:               2025
# }