Load cbb team rosters

Pull in rosters for college basketball teams for a given season

Usage

load_rosters(seasons = most_recent_season())

Arguments

seasons

an integer of seasons to fetch data for. Defaults to the most recent season. Pass in TRUE to fetch all seasons.

Value

a dataframe of rosters for college basketball teams in the specified season(s).

Examples

# \donttest{
load_rosters()
#>        team_id team_source_id                team conference season players_id
#>          <int>         <char>              <char>     <char>  <int>      <int>
#>     1:     161            113       Massachusetts       A-10   2025       2882
#>     2:     161            113       Massachusetts       A-10   2025       2883
#>     3:     161            113       Massachusetts       A-10   2025       2884
#>     4:     161            113       Massachusetts       A-10   2025       2885
#>     5:     161            113       Massachusetts       A-10   2025       2886
#>    ---                                                                        
#> 15233:     616           2985 Concordia-Ann Arbor       <NA>   2025       9314
#> 15234:     616           2985 Concordia-Ann Arbor       <NA>   2025       9315
#> 15235:     616           2985 Concordia-Ann Arbor       <NA>   2025       9316
#> 15236:     616           2985 Concordia-Ann Arbor       <NA>   2025       9317
#> 15237:     616           2985 Concordia-Ann Arbor       <NA>   2025       9318
#>        players_source_id           players_name players_first_name
#>                   <char>                 <char>             <char>
#>     1:           5176382       Malek Abdelgowad              Malek
#>     2:           5105982          Daniel Rivera             Daniel
#>     3:           5105551 Daniel Hankins-Sanford             Daniel
#>     4:           5174611           Jaylen Curry             Jaylen
#>     5:           4683733        Rahsool Diggins            Rahsool
#>    ---                                                            
#> 15233:           4907749     Tarek Abdel-Kireem              Tarek
#> 15234:           4907748           Abeal Ayalew              Abeal
#> 15235:           4907745            Matt Kohler               Matt
#> 15236:           4907744            Devyn Jones              Devyn
#> 15237:           4907742          Jayden Priddy             Jayden
#>        players_last_name players_jersey players_position players_height
#>                   <char>         <char>           <char>          <int>
#>     1:        Abdelgowad             25          Forward             82
#>     2:            Rivera              5          Forward             78
#>     3:   Hankins-Sanford              1          Forward             80
#>     4:             Curry              0            Guard             72
#>     5:           Diggins              7            Guard             74
#>    ---                                                                 
#> 15233:      Abdel-Kireem           <NA>          Athlete             NA
#> 15234:            Ayalew           <NA>          Athlete             NA
#> 15235:            Kohler           <NA>          Athlete             NA
#> 15236:             Jones           <NA>          Athlete             NA
#> 15237:            Priddy           <NA>          Athlete             NA
#>        players_weight players_hometown_1_city players_hometown_1_state
#>                 <int>                  <char>                   <char>
#>     1:            220                   Cairo                    Egypt
#>     2:            215                San Juan              Puerto Rico
#>     3:            236               Charlotte                       NC
#>     4:            168               Charlotte                       NC
#>     5:            190            Philadelphia                       PA
#>    ---                                                                
#> 15233:             NA                    <NA>                     <NA>
#> 15234:             NA                    <NA>                     <NA>
#> 15235:             NA                    <NA>                     <NA>
#> 15236:             NA                    <NA>                     <NA>
#> 15237:             NA                    <NA>                     <NA>
#>        players_hometown_1_country players_hometown_1_latitude
#>                            <lgcl>                       <num>
#>     1:                         NA                          NA
#>     2:                         NA                          NA
#>     3:                         NA                    35.22721
#>     4:                         NA                    35.22721
#>     5:                         NA                    39.95272
#>    ---                                                       
#> 15233:                         NA                          NA
#> 15234:                         NA                          NA
#> 15235:                         NA                          NA
#> 15236:                         NA                          NA
#> 15237:                         NA                          NA
#>        players_hometown_1_longitude players_hometown_1_county_fips
#>                               <num>                         <char>
#>     1:                           NA                           <NA>
#>     2:                           NA                           <NA>
#>     3:                    -80.84308                          37119
#>     4:                    -80.84308                          37119
#>     5:                    -75.16353                          42101
#>    ---                                                            
#> 15233:                           NA                           <NA>
#> 15234:                           NA                           <NA>
#> 15235:                           NA                           <NA>
#> 15236:                           NA                           <NA>
#> 15237:                           NA                           <NA>
#>        players_date_of_birth players_start_season players_end_season
#>                       <char>                <int>              <int>
#>     1:                  <NA>                 2025               2025
#>     2:                  <NA>                 2025               2025
#>     3:                  <NA>                 2024               2025
#>     4:                  <NA>                 2024               2025
#>     5:                  <NA>                 2023               2025
#>    ---                                                              
#> 15233:                  <NA>                 2023               2025
#> 15234:                  <NA>                 2022               2025
#> 15235:                  <NA>                 2022               2025
#> 15236:                  <NA>                 2022               2025
#> 15237:                  <NA>                 2023               2025

load_rosters(2024)
#>        team_id team_source_id                team conference season players_id
#>          <int>         <char>              <char>     <char>  <int>      <int>
#>     1:     161            113       Massachusetts       A-10   2024       2631
#>     2:     161            113       Massachusetts       A-10   2024       2884
#>     3:     161            113       Massachusetts       A-10   2024       2885
#>     4:     161            113       Massachusetts       A-10   2024       2886
#>     5:     161            113       Massachusetts       A-10   2024       2889
#>    ---                                                                        
#> 14126:     616           2985 Concordia-Ann Arbor       <NA>   2024       9314
#> 14127:     616           2985 Concordia-Ann Arbor       <NA>   2024       9315
#> 14128:     616           2985 Concordia-Ann Arbor       <NA>   2024       9316
#> 14129:     616           2985 Concordia-Ann Arbor       <NA>   2024       9317
#> 14130:     616           2985 Concordia-Ann Arbor       <NA>   2024       9318
#>        players_source_id           players_name players_first_name
#>                   <char>                 <char>             <char>
#>     1:           5174612       Robert Davis Jr.             Robert
#>     2:           5105551 Daniel Hankins-Sanford             Daniel
#>     3:           5174611           Jaylen Curry             Jaylen
#>     4:           4683733        Rahsool Diggins            Rahsool
#>     5:           5174615         Tarique Foster            Tarique
#>    ---                                                            
#> 14126:           4907749     Tarek Abdel-Kireem              Tarek
#> 14127:           4907748           Abeal Ayalew              Abeal
#> 14128:           4907745            Matt Kohler               Matt
#> 14129:           4907744            Devyn Jones              Devyn
#> 14130:           4907742          Jayden Priddy             Jayden
#>        players_last_name players_jersey players_position players_height
#>                   <char>         <char>           <char>          <int>
#>     1:         Davis Jr.              4            Guard             78
#>     2:   Hankins-Sanford              1          Forward             80
#>     3:             Curry              0            Guard             72
#>     4:           Diggins              7            Guard             74
#>     5:            Foster             12          Forward             80
#>    ---                                                                 
#> 14126:      Abdel-Kireem           <NA>          Athlete             NA
#> 14127:            Ayalew           <NA>          Athlete             NA
#> 14128:            Kohler           <NA>          Athlete             NA
#> 14129:             Jones           <NA>          Athlete             NA
#> 14130:            Priddy           <NA>          Athlete             NA
#>        players_weight players_hometown_1_city players_hometown_1_state
#>                 <int>                  <char>                   <char>
#>     1:            185                 Detroit                       MI
#>     2:            236               Charlotte                       NC
#>     3:            168               Charlotte                       NC
#>     4:            190            Philadelphia                       PA
#>     5:            185                   Bronx                       NY
#>    ---                                                                
#> 14126:             NA                    <NA>                     <NA>
#> 14127:             NA                    <NA>                     <NA>
#> 14128:             NA                    <NA>                     <NA>
#> 14129:             NA                    <NA>                     <NA>
#> 14130:             NA                    <NA>                     <NA>
#>        players_hometown_1_country players_hometown_1_latitude
#>                            <lgcl>                       <num>
#>     1:                         NA                    42.33155
#>     2:                         NA                    35.22721
#>     3:                         NA                    35.22721
#>     4:                         NA                    39.95272
#>     5:                         NA                    40.84665
#>    ---                                                       
#> 14126:                         NA                          NA
#> 14127:                         NA                          NA
#> 14128:                         NA                          NA
#> 14129:                         NA                          NA
#> 14130:                         NA                          NA
#>        players_hometown_1_longitude players_hometown_1_county_fips
#>                               <num>                         <char>
#>     1:                    -83.04664                          26163
#>     2:                    -80.84308                          37119
#>     3:                    -80.84308                          37119
#>     4:                    -75.16353                          42101
#>     5:                    -73.87859                          36005
#>    ---                                                            
#> 14126:                           NA                           <NA>
#> 14127:                           NA                           <NA>
#> 14128:                           NA                           <NA>
#> 14129:                           NA                           <NA>
#> 14130:                           NA                           <NA>
#>        players_date_of_birth players_start_season players_end_season
#>                       <char>                <int>              <int>
#>     1:                  <NA>                 2024               2024
#>     2:                  <NA>                 2024               2025
#>     3:                  <NA>                 2024               2025
#>     4:                  <NA>                 2023               2025
#>     5:                  <NA>                 2024               2025
#>    ---                                                              
#> 14126:                  <NA>                 2023               2025
#> 14127:                  <NA>                 2022               2025
#> 14128:                  <NA>                 2022               2025
#> 14129:                  <NA>                 2022               2025
#> 14130:                  <NA>                 2023               2025

load_rosters(TRUE)
#>         team_id team_source_id                team conference season players_id
#>           <int>         <char>              <char>     <char>  <int>      <int>
#>      1:     161            113       Massachusetts       A-10   2003      88819
#>      2:     161            113       Massachusetts       A-10   2003      91926
#>      3:     161            113       Massachusetts       A-10   2003      91929
#>      4:     161            113       Massachusetts       A-10   2003      92301
#>      5:     161            113       Massachusetts       A-10   2003      92303
#>     ---                                                                        
#> 217184:     616           2985 Concordia-Ann Arbor       <NA>   2025       9314
#> 217185:     616           2985 Concordia-Ann Arbor       <NA>   2025       9315
#> 217186:     616           2985 Concordia-Ann Arbor       <NA>   2025       9316
#> 217187:     616           2985 Concordia-Ann Arbor       <NA>   2025       9317
#> 217188:     616           2985 Concordia-Ann Arbor       <NA>   2025       9318
#>         players_source_id       players_name players_first_name
#>                    <char>             <char>             <char>
#>      1:             11811      Jeff Viggiano               Jeff
#>      2:              5264   Anthony Anderson            Anthony
#>      3:              5262     Brennan Martin            Brennan
#>      4:             11809           Gabe Lee               Gabe
#>      5:              5265     Paco Kotaridis               Paco
#>     ---                                                        
#> 217184:           4907749 Tarek Abdel-Kireem              Tarek
#> 217185:           4907748       Abeal Ayalew              Abeal
#> 217186:           4907745        Matt Kohler               Matt
#> 217187:           4907744        Devyn Jones              Devyn
#> 217188:           4907742      Jayden Priddy             Jayden
#>         players_last_name players_jersey players_position players_height
#>                    <char>         <char>           <char>          <int>
#>      1:          Viggiano            22           Forward             78
#>      2:          Anderson            12             Guard             71
#>      3:            Martin           <NA>          Forward             78
#>      4:               Lee            2             Center             81
#>      5:         Kotaridis            14             Guard             73
#>     ---                                                                 
#> 217184:      Abdel-Kireem           <NA>          Athlete             NA
#> 217185:            Ayalew           <NA>          Athlete             NA
#> 217186:            Kohler           <NA>          Athlete             NA
#> 217187:             Jones           <NA>          Athlete             NA
#> 217188:            Priddy           <NA>          Athlete             NA
#>         players_weight players_hometown_1_city players_hometown_1_state
#>                  <int>                  <char>                   <char>
#>      1:            215                Suffield                       CT
#>      2:            170                    Lynn                       MA
#>      3:            190  Rancho Santa Margarita                       CA
#>      4:            210                 Camilla                       GA
#>      5:            190                  Athens                   Greece
#>     ---                                                                
#> 217184:             NA                    <NA>                     <NA>
#> 217185:             NA                    <NA>                     <NA>
#> 217186:             NA                    <NA>                     <NA>
#> 217187:             NA                    <NA>                     <NA>
#> 217188:             NA                    <NA>                     <NA>
#>         players_hometown_1_country players_hometown_1_latitude
#>                             <lgcl>                       <num>
#>      1:                         NA                    41.98169
#>      2:                         NA                    42.46676
#>      3:                         NA                    33.64088
#>      4:                         NA                    31.23129
#>      5:                         NA                          NA
#>     ---                                                       
#> 217184:                         NA                          NA
#> 217185:                         NA                          NA
#> 217186:                         NA                          NA
#> 217187:                         NA                          NA
#> 217188:                         NA                          NA
#>         players_hometown_1_longitude players_hometown_1_county_fips
#>                                <num>                         <char>
#>      1:                    -72.65066                          09003
#>      2:                    -70.94949                          25009
#>      3:                   -117.60293                          06059
#>      4:                    -84.21037                          13205
#>      5:                           NA                           <NA>
#>     ---                                                            
#> 217184:                           NA                           <NA>
#> 217185:                           NA                           <NA>
#> 217186:                           NA                           <NA>
#> 217187:                           NA                           <NA>
#> 217188:                           NA                           <NA>
#>         players_date_of_birth players_start_season players_end_season
#>                        <char>                <int>              <int>
#>      1:            2003-09-17                 2003               2006
#>      2:                  <NA>                 2003               2005
#>      3:                  <NA>                 2003               2005
#>      4:            2003-09-17                 2003               2004
#>      5:                  <NA>                 2003               2004
#>     ---                                                              
#> 217184:                  <NA>                 2023               2025
#> 217185:                  <NA>                 2022               2025
#> 217186:                  <NA>                 2022               2025
#> 217187:                  <NA>                 2022               2025
#> 217188:                  <NA>                 2023               2025
# }